% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.repSparse}
\alias{as}
\alias{as.data.frame.repSparse}
\alias{as.data.table.repSparse}
\alias{as.matrix.repSparse}
\alias{as.repSparse}
\alias{as.repSparse.dsparseMatrix}
\alias{as.repSparse.matrix}
\title{Coerce to and from repeated sparse matrices}
\usage{
as.repSparse(x, ...)

\method{as.repSparse}{dsparseMatrix}(x, ...)

\method{as.repSparse}{matrix}(x, ...)

\method{as.data.frame}{repSparse}(x, ...)

as.data.table.repSparse(x, keep.rownames = FALSE)

as.matrix.repSparse(x, sparse = FALSE, ...)

as("repSparse", "sparseMatrix")

as("repSparse", "CsparseMatrix")

as("repSparse", "dgCMatrix")

as("repSparse", "TsparseMatrix")

as("repSparse", "dgTMatrix")
}
\arguments{
\item{x}{an object}

\item{...}{dots}

\item{keep.rownames}{passed to \code{as.data.table}}

\item{sparse}{return \code{sparseMatrix}?}
}
\description{
Coerce to and from repeated sparse matrices
}

