% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.strucMatrix}
\alias{as.data.frame.strucMatrix}
\alias{as.matrix.strucMatrix}
\alias{as.matrix.strucMatrixChol}
\alias{as.strucMatrix}
\alias{as.strucMatrix.dCHMsimpl}
\alias{as.strucMatrix.dist}
\alias{as.strucMatrix.dsparseMatrix}
\alias{as.strucMatrix.factor}
\alias{as.strucMatrix.matrix}
\alias{as.strucMatrix.strucMatrix}
\alias{strucMatrix2dgCMatrix}
\alias{strucMatrix2dgTMatrix}
\alias{strucMatrix2dtCMatrix}
\alias{strucMatrix2dtTMatrix}
\alias{strucMatrix2gCsparseMatrix}
\alias{strucMatrix2gTsparseMatrix}
\alias{strucMatrix2sparseMatrix}
\title{Coerce to and from structured sparse matrices}
\usage{
as.strucMatrix(x, ...)

\method{as.strucMatrix}{strucMatrix}(x, ...)

\method{as.strucMatrix}{dsparseMatrix}(x, ...)

\method{as.strucMatrix}{matrix}(x, ...)

\method{as.strucMatrix}{factor}(x, ...)

\method{as.strucMatrix}{dist}(x, ...)

\method{as.strucMatrix}{dCHMsimpl}(x, ...)

\method{as.data.frame}{strucMatrix}(x, ...)

\method{as.matrix}{strucMatrix}(x, sparse = FALSE, ...)

\method{as.matrix}{strucMatrixChol}(x, sparse = FALSE, ...)
}
\arguments{
\item{x}{an object}

\item{...}{dots}

\item{sparse}{return \code{sparseMatrix}?}
}
\description{
Coerce to and from structured sparse matrices

as("strucMatrix", "sparseMatrix")

as("strucMatrix", "CsparseMatrix")

as("strucMatrix", "dgCMatrix")

as("strucMatrix", "dtCMatrix")

as("strucMatrix", "TsparseMatrix")

as("strucMatrix", "dgTMatrix")

as("strucMatrix", "dtTMatrix")
}
\examples{
set.seed(1)
m1 <- as.strucMatrix(matrix(rnorm(6), 2, 3))
m2 <- as.strucMatrix(matrix(rbinom(6, 1, 0.5), 2, 3))
m3 <- sparseMatrix(i = 1:10, j = rep(1:5, 2),
                   x = as.integer(rbinom(10, 1, 0.5)),
                   giveCsparse = FALSE)
as.strucMatrix(m1)
as.strucMatrix(m2)
as.strucMatrix(m3)
as(m1, "TsparseMatrix")
as(m2, "dgCMatrix")
as(m2, "sparseMatrix")
}

