% Generated by roxygen2 (4.0.1): do not edit by hand
\name{assignWith}
\alias{assignWith}
\title{Assign to an environment the value of an expression evaluated in
another environment (or list or data frame)}
\usage{
assignWith(expr, name, data, envir, enclos = parent.frame())
}
\arguments{
\item{expr}{expression to evaluate}

\item{name}{name of object in \code{assignEnv}}

\item{data}{environment (or list or data frame) in which to
evaluate \code{expr}}

\item{envir}{environment in which to store the results of
\code{expr} as \code{name}}
}
\description{
Assign to an environment the value of an expression evaluated in
another environment (or list or data frame)
}

