% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bind}
\alias{bind}
\title{Row, column, and block-diagonal binding for repeated sparse
matrices}
\usage{
bind(..., type = c("row", "col", "diag"))
}
\arguments{
\item{...}{list of \code{repSparse} objects}

\item{type}{type of binding}
}
\description{
Row, column, and block-diagonal binding for repeated sparse
matrices
}

