% Generated by roxygen2 (4.0.1): do not edit by hand
\name{point2ind}
\alias{ind2point}
\alias{point2ind}
\title{Changing sparse format}
\usage{
point2ind(point)

ind2point(ind, maxInd, fillNA = TRUE)
}
\arguments{
\item{point}{vector of column pointers}

\item{ind}{vector of column indices}

\item{maxInd}{number of rows or columns (could be larger than
\code{max(ind)})}

\item{fillNA}{fill in \code{NA}'s with repeated column numbers (as
in \code{Matrix} package)?}
}
\description{
Changing sparse format
}

