% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{chol,repSparse-method}
\alias{chol,repSparse-method}
\alias{chol,repSparseCompSymm-method}
\alias{chol,repSparseOneOffDiag-method}
\title{Cholesky decomposition of repeated sparse matrices}
\usage{
\S4method{chol}{repSparse}(x, ...)

\S4method{chol}{repSparseOneOffDiag}(x, ...)

\S4method{chol}{repSparseCompSymm}(x, ...)
}
\arguments{
\item{x}{an object that inherits from class
\code{\link{repSparse}}}

\item{...}{passed to subsequent functions}
}
\description{
Cholesky decomposition of repeated sparse matrices
}
\note{
These are often just bailout methods, but some special
\code{repSparse} matrices have exploitable structure, which can be
used to keep the number of repeated values down.
}
\examples{
x <- repSparseCompSymm(1.2, -0.11, 4)
as.matrix(chol(x))
t(chol(as.matrix(x)))
}

