% Generated by roxygen2 (4.0.1): do not edit by hand
\name{colBind}
\alias{colBind}
\alias{rowBind}
\title{Column and row binding fo repeated sparse matrices}
\usage{
colBind(...)

rowBind(...)
}
\arguments{
\item{...}{list of \code{repSparse} objects}
}
\description{
Column and row binding fo repeated sparse matrices
}

