% Generated by roxygen2 (4.0.1): do not edit by hand
\name{covExpDecay}
\alias{covExpDecay}
\title{Exponentially decaying covariances}
\usage{
covExpDecay(matPars, minCov = 0.001, distCutoff = 2, nPoints = 100)
}
\arguments{
\item{matPars}{parameters of an \code{\link{expDecay}} term
(FIXME: give more info).}

\item{minCov,distCutoff}{see \code{\link{expDecay}}.}

\item{nPoints}{number of points at which to evaluate the curve.}
}
\description{
Compute the exponential decay associated with an
\code{\link{expDecay}} term.
}
\examples{
covExpDecay(c(1, 1))
}

