% Generated by roxygen2 (4.0.1): do not edit by hand
\name{factorCov}
\alias{factorCov}
\alias{print.factorCov}
\title{Factors with covariance matrices over the levels}
\usage{
factorCov(fac, covMat, stan = TRUE)

print.factorCov(x, ...)
}
\arguments{
\item{fac}{a factor}

\item{covMat}{a covariance matrix over the levels (if missing use
identity matrix)}

\item{stan}{standardize covariance matrix to determinant one?}
}
\value{
a \code{factorCov} object
}
\description{
Factors with covariance matrices over the levels
}

