% Generated by roxygen2 (4.0.1): do not edit by hand
\name{factors}
\alias{factors}
\title{Extract factors}
\usage{
factors(x, ...)
}
\arguments{
\item{x}{an object with \code{x$factors}}

\item{...}{not used}
}
\description{
Intended for factor analysis (\code{\link{factAnal}}), and
modelled after the \code{\link{loadings}} function in the
\code{stats} package.
}

