% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{family.strucGlmer}
\alias{family.strucGlmer}
\alias{family.strucParseFormula}
\title{Extract family object}
\usage{
\method{family}{strucGlmer}(object, ...)

\method{family}{strucParseFormula}(object, ...)
}
\arguments{
\item{object}{\code{\link{strucGlmer}} or
\code{\link{strucParseFormula}} object}

\item{...}{not used}
}
\description{
Extract family object
}

