% Generated by roxygen2 (4.0.1): do not edit by hand
\name{familySimFun}
\alias{familySimFun}
\alias{familySimFun.GammaFamily}
\alias{familySimFun.binomialFamily}
\alias{familySimFun.character}
\alias{familySimFun.family}
\alias{familySimFun.gaussianFamily}
\alias{familySimFun.poissonFamily}
\title{Family simulation functions}
\usage{
familySimFun(object, ...)

\method{familySimFun}{family}(object, ...)

\method{familySimFun}{character}(object, ...)

\method{familySimFun}{gaussianFamily}(object, ...)

\method{familySimFun}{binomialFamily}(object, ...)

\method{familySimFun}{poissonFamily}(object, ...)

\method{familySimFun}{GammaFamily}(object, ...)
}
\arguments{
\item{object}{typically a \code{\link{family}} object}

\item{...}{additional objects to pass on (not yet used).}
}
\value{
a simulation function taking three arguments: \code{wts},
\code{nsim}, and \code{ftd}.
}
\description{
Family simulation functions
}

