% Generated by roxygen2 (4.0.1): do not edit by hand
\name{findReTrmClasses}
\alias{findReTrmClasses}
\title{Find classes with a \code{setReTrm} method}
\usage{
findReTrmClasses(formula = NULL)
}
\arguments{
\item{formula}{generalized mixed model formula.  if \code{NULL}
(the default) \code{findReTrmClasses} returns classes available
(on the search path).}
}
\description{
Find classes with a \code{setReTrm} method
}

