% Generated by roxygen2 (4.0.1): do not edit by hand
\name{flattenIntVec}
\alias{flattenIntVec}
\title{Flatten an integer vector}
\usage{
flattenIntVec(x)
}
\arguments{
\item{x}{vector coercible to nonnegative integers}
}
\description{
Map integers in a vector, \code{x}, to the integers from \code{1,
..., n}, where \code{n} is the number of unique integers in
\code{x}.
}
\examples{
flattenIntVec(rep(seq(0, 100, length = 5), 1:5))
}

