% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Gamma_simfun}
\alias{Gamma_simfun}
\title{FIXME: need a gamma.shape.merMod method in order for this to work.
       (see initial shot at gamma.shape.merMod below)}
\usage{
Gamma_simfun(object, nsim, ftd = fitted(object))
}
\description{
FIXME: need a gamma.shape.merMod method in order for this to work.
       (see initial shot at gamma.shape.merMod below)
}

