% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gblmer}
\alias{gblmer}
\title{Generalized bilinear mixed model}
\usage{
gblmer(formula, data, family, latentName = "latent", latentDims = 0L,
  loadingsDim = 1L, loadingPen = 0L, verbose = 0L, ...)
}
\arguments{
\item{linFormula}{a mixed model formula for the linear part of the
model}

\item{bilinFormula}{a mixed model formula for the bilinear part of
the model}

\item{data}{a \code{\link{data.list}} object}

\item{family}{a \code{\link{family}} object}

\item{latentDims}{number of latent dimensions in the bilinear
component of the model}

\item{loadingPen}{penalty for the size of the loadings}

\item{verbose}{passed to optimizer}

\item{...}{arguments to be passed to \code{\link{glFormula}}}
}
\description{
Generalized bilinear mixed model
}

