% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{gblmerMod-class}
\alias{gblmerMod-class}
\title{Class "gblmerMod"}
\description{
Class \code{"gblmerMod"} is san S4 class that extends
\code{"glmerMod"}
}
\section{Slots}{
 in addition to the slots provided by
\code{"glmerMod"}, there is an additional \code{"loadings"} slot
containing the factor loadings
}
\keyword{classes}

