% Generated by roxygen2 (4.0.1): do not edit by hand
\name{generalGlmer}
\alias{generalGlmer}
\title{General glmer}
\usage{
generalGlmer(formula, data, addArgs = list(), ...)
}
\arguments{
\item{formula}{extended mixed model formula}

\item{data}{data frame}

\item{addArgs}{list of additional arguments to pass to
\code{\link{setReTrm}} methods}

\item{...}{further arguments to \code{\link{mkGeneralGlmerDevfun}}}
}
\description{
General glmer
}

