% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getInit}
\alias{getInit}
\alias{getInit.default}
\alias{getInit.function}
\alias{getInit.repSparse}
\alias{setInit}
\alias{setInit.default}
\alias{setInit.function}
\alias{setInit.repSparse}
\title{Get and set initial parameter values for repeated sparse matrices}
\usage{
getInit(x, ...)

\method{getInit}{default}(x, ...)

\method{getInit}{repSparse}(x, ...)

\method{getInit}{function}(x, ...)

setInit(x, init, ...)

\method{setInit}{default}(x, init, ...)

\method{setInit}{repSparse}(x, init, ...)

\method{setInit}{function}(x, init, ...)
}
\arguments{
\item{x}{object}

\item{...}{not yet used}

\item{init}{initial value for the parameter vector}
}
\description{
Get and set initial parameter values for repeated sparse matrices
}
\examples{
set.seed(1)
m1 <- as.repSparse(matrix(rnorm(6), 2, 3))
m2 <- repSparseCompSymm(1.2, -0.2, 5)
getInit(m1)
getInit(m2)
}

