% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getMEc}
\alias{getMEc}
\title{Get components of a \code{glmerc} object}
\usage{
getMEc(object, name = c("X", "Z", "Zt", "y", "TmodMat", "cnms"))
}
\arguments{
\item{object}{\code{\link{glmerc}} object}

\item{name}{object name}
}
\description{
Get components of a \code{glmerc} object
}

