% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structuredSparseMatrices.R
\name{getRepPattern}
\alias{getRepPattern}
\title{Get repetition pattern of a structured sparse matrix}
\usage{
getRepPattern(object)
}
\arguments{
\item{object}{a \code{\link{strucMatrix}} object}
}
\description{
Get repetition pattern of a structured sparse matrix
}

