% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getReTrm}
\alias{getReTrm}
\title{Get random effects structures from a strucGlmer object}
\usage{
getReTrm(object, name, drop = TRUE)
}
\arguments{
\item{object}{\code{\link{strucGlmer}} or
\code{\link{strucParseFormula}} object}

\item{name}{Names of the random effects structures (if missing, a
list of the names are returned).  The naming convention for
\code{reTrmStruct} objects is
\code{grpFacName.reTrmStructClass}. Partial matching of names is
allowed.}

\item{drop}{If only one term is selected, return the
\code{reTrmStruct} objectitself rather than a length-one list with
the object.}
}
\value{
a fitted random effects structure
}
\description{
Get random effects structures from a strucGlmer object
}
\seealso{
\code{\link{setReTrm}}
}

