% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getReTrms}
\alias{getReTrms}
\title{Get random effects terms from a fitted \code{merMod} object}
\usage{
getReTrms(object, ...)
}
\arguments{
\item{object}{\code{\link{merMod}} object}
}
\value{
see \code{\link{mkReTrms}}
}
\description{
Get random effects terms from a fitted \code{merMod} object
}

