% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getReTrmStruct}
\alias{getReTrmStruct}
\title{Get random effects structures from a strucGlmer object}
\usage{
getReTrmStruct(object, name)
}
\arguments{
\item{object}{\code{\link{strucGlmer}} object}

\item{name}{Names of the random effects structures (if missing, a
list of the names are returned).  The naming convention for
\code{reTrmStruct} objects is
\code{grpFacName.reTrmStructClass}. Partial matching of names is
allowed.}
}
\value{
a fitted random effects structure
}
\description{
Get random effects structures from a strucGlmer object
}

