% Generated by roxygen2 (4.0.1): do not edit by hand
\name{glmercFormula}
\alias{glmercFormula}
\title{Parse mixed model formula for levels-covariance model}
\usage{
glmercFormula(formula, data = NULL, family = binomial, covList, ...)
}
\arguments{
\item{formula}{mixed model formula}

\item{data}{data}

\item{family}{family}

\item{covList}{list of covariance matrices for random effects
grouping factors in \code{formula}}

\item{...}{not used yet}
}
\description{
Parse mixed model formula for levels-covariance model
}

