% Generated by roxygen2 (4.0.1): do not edit by hand
\name{glmerf}
\alias{glmerf}
\title{Generalized linear mixed model with factor loadings}
\usage{
glmerf(formula, data, family, latentName = "latent", latentDims = 0L,
  loadingsDim = 1L, loadingPen = 0L, verbose = 0L, ...)
}
\arguments{
\item{formula}{a mixed model formula for the linear part of the
model}

\item{data}{a \code{\link{data.list}} object}

\item{family}{a \code{\link{family}} object}

\item{latentDims}{number of latent dimensions in the bilinear
component of the model}

\item{loadingsDim}{huh?}

\item{loadingPen}{penalty for the size of the loadings}

\item{verbose}{passed to optimizer}

\item{...}{arguments to be passed to \code{\link{glFormula}}}
}
\description{
Generalized linear mixed model with factor loadings
}

