% Generated by roxygen2 (4.0.1): do not edit by hand
\name{growTree}
\alias{growTree}
\title{Grow the history tree}
\usage{
growTree(newNodes, oldTree)
}
\arguments{
\item{newNodes}{list of new component \code{repSparse} objects}

\item{oldTree}{existing tree (can be \code{NULL})}
}
\description{
Grow the history tree
}

