% Generated by roxygen2 (4.0.1): do not edit by hand
\name{joinReTrms}
\alias{joinReTrms}
\title{Join two lists describing two sets of random effects terms}
\usage{
joinReTrms(reTrms1, reTrms2)
}
\arguments{
\item{reTrms1,reTrms2}{results of \code{\link{glFormula}(.)$reTrms}}
}
\description{
Join two lists describing two sets of random effects terms
}

