% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structuredSparseMatrices.R
\docType{methods}
\name{kron}
\alias{kr}
\alias{kron}
\alias{kronecker,strucMatrix,strucMatrix-method}
\title{Kronecker and Khatri-Rao products for structured sparse matrices}
\usage{
kron(X, Y, trans = "*", makedimnames = FALSE, ...)

kr(X, Y, trans = "*")

\S4method{kronecker}{strucMatrix,strucMatrix}(X, Y, FUN = "*",
  make.dimnames = FALSE, ...)
}
\arguments{
\item{X, Y}{structured sparse matrices (\code{\link{strucMatrix-class}})}

\item{trans}{see argument \code{FUN} in \code{\link{outer}}}

\item{makedimnames}{ignored}

\item{...}{ignored}

\item{FUN}{function to apply; may be a quoted string (currently ignored; fo r compatibility)}

\item{make.dimnames}{currently ignored}
}
\description{
Kronecker and Khatri-Rao products for structured sparse matrices
}
\examples{
set.seed(1)
X <- strucMatrix(c(1, 2, 1, 2), c(1, 1, 2, 2), 1:4, rnorm(4))
Y <- strucMatrix(c(1, 2, 1), c(1, 1, 2), 1:3, rnorm(3))
(kronExample <- kron(X, Y))
(krExample <- kr(X, Y))
}
\seealso{
Other matrixCombining: \code{\link{bind}},
  \code{\link{subset.strucMatrix}}

Other matrixCombining: \code{\link{bind}},
  \code{\link{subset.strucMatrix}}
}

