% Generated by roxygen2 (4.0.1): do not edit by hand
\name{laplace}
\alias{laplace}
\alias{laplace.default}
\alias{laplace.glmerMod}
\alias{laplace.strucGlmer}
\title{Laplace approximation}
\usage{
laplace(object, ...)

\method{laplace}{default}(object, ...)

\method{laplace}{glmerMod}(object, ...)

\method{laplace}{strucGlmer}(object, ...)
}
\arguments{
\item{object}{fitted model object}

\item{...}{not used}
}
\description{
Compute the Laplace approximation of the log-likelihood of a
generalized linear mixed model.
}

