% Generated by roxygen2 (4.0.1): do not edit by hand
\name{listOfSubscriptingCallsWithListOfIndices}
\alias{listOfSubscriptingCallsWithListOfIndices}
\title{List of subscripting calls with a list of indices}
\usage{
listOfSubscriptingCallsWithListOfIndices(vecName, indsListName, indTypeNames)
}
\arguments{
\item{vecName}{length-one character vector giving the name of the
vector to be subscripted}

\item{indsListName}{length-one character vector giving the name of
the index list}

\item{indTypeNames}{character vector giving the names of the index types}
}
\description{
Suppose one has an atomic vector and a named list of indices for
the vector. Now suppose one wants a list of calls subscripting the
vector with each element of the index list. This function does
that.
}
\examples{
listOfSubscriptingCallsWithListOfIndices("pars", "parInds", c("covar", "fixef"))
}

