% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loadings}
\alias{loadings}
\title{Refactor \code{loadings} into a generic function}
\usage{
loadings(x, ...)
}
\arguments{
\item{x}{object with loadings to extract}

\item{...}{additional arguments}
}
\description{
Refactor \code{loadings} into a generic function
}

