% Generated by roxygen2 (4.0.0): do not edit by hand
\name{logisticPcaDevfun}
\alias{logisticPcaDevfun}
\title{Construct deviance function for logistic principal components analysis}
\usage{
logisticPcaDevfun(Y, d, Xrow, Xcol, Upenalty, thetaPenalty, Ustart, thetaStart)
}
\arguments{
\item{Y}{response matrix}

\item{d}{number of axes}

\item{Xrow}{row model matrix}

\item{Xcol}{col model matrix}

\item{Upenalty}{penalty for axis scores}

\item{thetaPenalty}{penalty for theta for the axes}

\item{Ustart}{optional initial matrix of column scores}

\item{thetaStart}{optional initial value for theta vector}
}
\description{
Construct deviance function for logistic principal components analysis
}

