% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rowWiseCombination}
\alias{kr}
\alias{kron}
\alias{mmult}
\alias{rowWiseCombination}
\title{Row-wise combination of two repeated sparse matrices}
\usage{
rowWiseCombination(X, Y)

mmult(X, Y, trans = "*")

kron(X, Y, trans = "*", makedimnames = FALSE, saveComponents = TRUE, ...)

kr(X, Y, trans = "*", saveComponents = TRUE)
}
\arguments{
\item{X,Y}{\code{repSparse} objects}

\item{X,Y}{\code{repSparse} objects}

\item{trans}{two argument transformation from \code{X$vals} and
\code{Y$vals} to the repeated values of the resulting matrix}

\item{X,Y}{\code{repSparse} objects}

\item{trans}{see argument \code{FUN} in \code{\link{outer}}}

\item{makedimnames}{ignored}

\item{saveComponents}{should component matrices be saved?}

\item{...}{ignored}
}
\value{
a row-wise combination of repeated sparse matrices
}
\description{
Row-wise combination of two repeated sparse matrices

Standard matrix multiplication for repeated sparse matrices

Kronecker and Khatri-Rao products for repeated sparse matrices
}

