% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mkGblmerDevfun}
\alias{mkGblmerDevfun}
\title{Modular functions for generalized bilinear mixed models}
\usage{
mkGblmerDevfun(fr, X, linReTrms, bilinReTrms, family, nAGQ = 1,
  verbose = 0L, ...)
}
\arguments{
\item{fr}{model frame}

\item{X}{fixed effects model matrix}

\item{linReTrms}{linear random effects terms}

\item{bilinReTrms}{bilinear random effects terms}

\item{family}{family}

\item{nAGQ}{nAGQ}

\item{verbose}{verbose}

\item{...}{...}
}
\description{
Modular functions for generalized bilinear mixed models
}

