% Generated by roxygen2 (4.0.0): do not edit by hand
\name{mkMod}
\alias{mkMod}
\title{Construct deviance function for logistic principal components analysis}
\usage{
mkMod(rho, opt)
}
\arguments{
\item{rho}{environment of the deviance function}

\item{opt}{output of the optimizer}
}
\description{
Construct deviance function for logistic principal components analysis
}

