% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mkNfun}
\alias{mkNfun}
\title{Make function for computing normalized deviance}
\usage{
mkNfun(mod)
}
\arguments{
\item{mod}{}
}
\value{
function for computing the normalized deviance, which
equals zero at the optimum
}
\description{
Make function for computing normalized deviance
}

