% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mkOuterTrans}
\alias{mkOuterTrans}
\title{Construct function for a transformed outer product}
\usage{
mkOuterTrans(A, B, Atrans, Btrans, ABtrans)
}
\arguments{
\item{A,B}{\code{numeric} \code{vector}s}

\item{Atrans,Btrans}{functions for transforming \code{A} and
\code{B}}

\item{ABtrans}{function to pass as \code{FUN} in
\code{\link{outer}}}
}
\description{
Construct function for a transformed outer product
}

