% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mkParInds}
\alias{mkParInds}
\title{Make parameter indices}
\usage{
mkParInds(parList)
}
\arguments{
\item{parList}{named list of parameters with possible names (see
details): (\code{covar}, \code{fixef}, \code{loads}, \code{weigh})}
}
\value{
A list with the same names as \code{parList} with the
indices for each type of parameter.
}
\description{
One faces two somewhat opposing design goals when deciding on how
to represent the parameters of structured generalized linear mixed
models. On one hand, it would be nice to organize the
representation in terms of the four different types of parameters
(see below): (\code{covar}, \code{fixef}, \code{loads},
\code{weigh}). On the other hand, the nonlinear optimizer takes
vector valued parameter sets. The \code{mkParInds} function is
used to get the best of both worlds, by constructing a list of
indices for extracting various types of parameters from a
parameter vector.
}
\details{
The \code{\link{lme4ord}} package keeps track of parameters using
a simple named list of parameter vectors. The names correspond to
different types of parameters, described in the following list:
\describe{

\item{\code{covar}}{Parameters determining the transposed relative
covariance factor, \code{Lambdat}.}

\item{\code{fixef}}{The fixed effects coefficients.}

\item{\code{loads}}{Parameters (called loadings) determining the
random effects model matrix, \code{Zt}.}

\item{\code{weigh}}{Parameters determining the observation weights
(experimental).}

}
\code{mkParInds} constructs the indices required to extract the
different types of parameters in \code{unlist(parList)}.
}
\examples{
set.seed(1)
parList <- list(covar = 1, fixef = c(0, 0), loads = rnorm(10))
parInds <- mkParInds(parList)
parVec <- unlist(parList)
parVec[parInds$fixef]
}

