% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mkPenaltyFun}
\alias{mkPenaltyFun}
\title{Make penalty function}
\usage{
mkPenaltyFun(penCovar, penFixef, penLoads, env = environment())
}
\arguments{
\item{penCovar,penFixef,penLoads}{Penalty functions for
covariance, fixed effects, and loadings parameters.  Can be
\code{NULL}.}

\item{env}{Environment in which to evaluate the resulting
function.}
}
\value{
A function with two arguments:
\item{pars}{Parameter vector}
\item{parInds}{Named list of indices pointing to the elements of
\code{pars} corresponding to one of the three parameter types. The
names of this list must include \code{covar}, \code{fixef}, and/or
\code{loads} if the correponding type has a supplied
(i.e. non-null) penalty function.}
}
\description{
Make penalty function
}
\examples{
penCovar <- function(x) x^2
penFixef <- NULL
penLoads <- function(gg) abs(gg) + 10
mkPenaltyFun(penCovar, penFixef, penLoads)(1:10, list(covar = 1:5, loads = 6:10))
}

