% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mkPfun}
\alias{mkPfun}
\title{Make function for computing profiled deviance}
\usage{
mkPfun(mod, whichPar)
}
\arguments{
\item{mod}{\code{\link{glmerc}} object}

\item{whichPar}{which parameter to profile}
}
\value{
function for computing the profiled deviance over a single
parameter
}
\description{
Make function for computing profiled deviance
}

