% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mkReStructs}
\alias{mkReStructs}
\title{Construct random effects structures}
\usage{
mkReStructs(splitFormula, data)
}
\arguments{
\item{splitFormula}{results of \code{link{splitForm}}}

\item{data}{data}
}
\description{
Construct random effects structures
}

