% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mkGeneralReTrms}
\alias{mkGeneralReTrms}
\alias{mkSparseTrans}
\title{Make general random effects terms}
\usage{
mkGeneralReTrms(reStructList, parsedForm, ...)

mkSparseTrans(object)
}
\arguments{
\item{reStructList}{list of random effects structure functions}

\item{parsedForm}{list parsed random effects terms}

\item{...}{potential extra parameters}

\item{object}{repeated sparse matrix object}
}
\description{
Make general random effects terms

Make transformation function for column-compressed sparse matrix
}

