% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mkReTrmStruct}
\alias{mkReTrmStruct}
\title{Make random effects term structure}
\usage{
mkReTrmStruct(formula, class, data)
}
\arguments{
\item{formula}{random effects term formula (without special function)}

\item{class}{random effects term class}

\item{data}{a data frame}
}
\description{
Make random effects term structure
}

