% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mkReTrmStructs}
\alias{mkReTrmStructs}
\alias{reTrmStruct}
\alias{reTrmStruct-class}
\title{Construct random effects structures}
\usage{
mkReTrmStructs(splitFormula, data)
}
\arguments{
\item{splitFormula}{results of \code{\link{splitForm}} (or a
\code{formula} itself)}

\item{data}{data}
}
\value{
A list of random effects structures, each associated with
the terms in a generalized mixed model formula of roughly the
following form: \code{(linForm1 | grpFac1) + (linForm2 | grpFac2)
+ ... + specialStruc(linForm3 | grpFac3) + ...}.  Each such
structure may be unstructured with class `unstruc` (indicating
standard `lme4` structures, e.g. the first two terms above) or
class of same name as the function in the formula
(e.g. \code{specialStruc}).  Each structure also contains (at a
minimum) the following elements (although \code{\link{setReTrm}}
adds further structure):

\item{modMat}{A raw model matrix for which the structure is
defined.  This matrix is obtained by evaluating
\code{model.matrix(linForm, data)}, where \code{linForm | grpFac}
is the random effects formula defining the term.}

\item{grpFac}{If present, the grouping factor associated with the
structure, \code{NA} otherwise.}

\item{grpName}{If present, the name of the grouping factor,
\code{NA} otherwise.}  \item{addArgs}{If present, any additional
arguments passed to the special function in the formula.}
}
\description{
Construct random effects structures from a formula and data.
}
\seealso{
\code{\link{setReTrm}} for initializing the structures as
is required for model fitting, and \code{\link{splitForm}} for
breaking the formula up into terms.
}

