% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mkSfun}
\alias{mkSfun}
\title{Make function for computing deviance slices}
\usage{
mkSfun(mod, whichPar)
}
\arguments{
\item{mod}{\code{\link{glmerc}} object}

\item{whichPar}{which parameter to slice}
}
\value{
function for computing 1D slices through the deviance at
the optimum
}
\description{
Make function for computing deviance slices
}

