% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mkSparseTrans}
\alias{mkSparseTrans}
\title{Make transformation function to be used with a column-compressed
sparse matrix}
\usage{
mkSparseTrans(object)
}
\arguments{
\item{object}{repeated sparse matrix object}
}
\description{
Make transformation function to be used with a column-compressed
sparse matrix
}
\examples{
set.seed(1)
X <- rRepSparse(3, 7, 2, 12)
(Xsparse <- as.matrix(X, sparse = TRUE))
Xtrans <- mkSparseTrans(X)
slot(Xsparse, "x") <- Xtrans(rnorm(2))
print(Xsparse)
}

