% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mkStrucGlmer}
\alias{mkStrucGlmer}
\title{Make strucGlmer object}
\usage{
mkStrucGlmer(opt, parsedForm, dfun, mc)
}
\arguments{
\item{opt,parsedForm,dfun,mc}{See \code{\link{strucGlmer-class}}}
}
\value{
An object of \code{\link{strucGlmer-class}}
}
\description{
Make strucGlmer object
}

