% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mkTemplateReTrm}
\alias{mkTemplateReTrm}
\title{Make covariance template random effects term}
\usage{
mkTemplateReTrm(modMat, grpFac1, grpFac2, covMat1, covMat2)
}
\arguments{
\item{modMat}{a model matrix}

\item{grpFac1,grpFac2}{grouping factors (typically dimension IDs
of a melted response matrix)}

\item{covMat1,covMat2}{covariance matrices among the levels of
\code{grpFac1} and \code{grpFac2}}
}
\description{
Make covariance template random effects term
}

