% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mkTemplateReTrms}
\alias{mkTemplateReTrms}
\title{Make several covariance template random effects terms}
\usage{
mkTemplateReTrms(modMat, grpFac1, grpFac2, covMat1, covMat2)
}
\arguments{
\item{modMat,grpFac1,grpFac2,covMat1,covMat2}{lists of inputs to
\code{\link{mkTemplateReTrm}}}
}
\description{
Make several covariance template random effects terms
}

