% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mkTemplateTermLambdat}
\alias{mkTemplateTermLambdat}
\title{Make covariance factor for a covariance template term}
\usage{
mkTemplateTermLambdat(covMat, nLevels, diagModel = TRUE)
}
\arguments{
\item{covMat}{template covariance matrix}

\item{nLevels}{number of levels over which the grouping factor varies}

\item{diagModel}{templates on the block diagonal (\code{TRUE}) or
densely tiled over the entire matrix (\code{FALSE})?}
}
\description{
Make covariance factor for a covariance template term
}
\examples{
cm <- crossprod(matrix(rnorm(25), 5, 5))
LamtDiag <- mkTemplateTermLambdat(cm, 4, TRUE)
LamtDense <- mkTemplateTermLambdat(cm, 4, FALSE)
image(LamtDiag)
image(LamtDense)
}

