% Generated by roxygen2 (4.0.1): do not edit by hand
\name{nChoose2Inv}
\alias{nChoose2Inv}
\title{Inverse of the n-choose-2 function}
\usage{
nChoose2Inv(m)
}
\arguments{
\item{m}{a vector coercible to integer}
}
\description{
Inverse of the n-choose-2 function
}
\note{
No checks are made for non-integer input or output.
}
\examples{
nChoose2Inv(choose(1:10, 2)) # = 1:10
}

