% Generated by roxygen2 (4.0.1): do not edit by hand
\name{nobarsWithSpecials}
\alias{nobarsWithSpecials}
\title{Version of the recursive nobars function from lme4}
\usage{
nobarsWithSpecials(term)
}
\arguments{
\item{term}{term}
}
\description{
Version of the recursive nobars function from lme4
}

