% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{Ops.repSparse}
\alias{Ops.repSparse}
\alias{crossprod,repSparse,missing-method}
\alias{crossprod,repSparse,repSparse-method}
\alias{tcrossprod,repSparse,missing-method}
\alias{tcrossprod,repSparse,repSparse-method}
\title{Repeated sparse matrix multiplication (and other binary operations)}
\usage{
\method{Ops}{repSparse}(e1, e2)

\S4method{crossprod}{repSparse,missing}(x, y = NULL, ...)

\S4method{tcrossprod}{repSparse,missing}(x, y = NULL, ...)

\S4method{crossprod}{repSparse,repSparse}(x, y = NULL, ...)

\S4method{tcrossprod}{repSparse,repSparse}(x, y = NULL, ...)
}
\arguments{
\item{e1,e2}{\code{repSparse} objects}

\item{x,y}{\code{\link{repSparse}} matrix objects}

\item{...}{not used (only for consistency with generic)}
}
\description{
Because of the marginal summations involved, the results of a
matrix multiplication of \code{\link{repSparse}} objects is not
profitably stored as a \code{repSparse}. Therefore, the output is
a \code{\link{sparseMatrix}}.
}
\note{
The \code{*} operator is matrix multiplication, not
element-wise multiplication.
}

