% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{Ops.strucMatrix}
\alias{Ops.strucMatrix}
\alias{crossprod,strucMatrix,missing-method}
\alias{crossprod,strucMatrix,strucMatrix-method}
\alias{tcrossprod,strucMatrix,missing-method}
\alias{tcrossprod,strucMatrix,strucMatrix-method}
\title{Structured sparse matrix multiplication (and other binary operations)}
\usage{
\method{Ops}{strucMatrix}(e1, e2)

\S4method{crossprod}{strucMatrix,missing}(x, y = NULL, ...)

\S4method{tcrossprod}{strucMatrix,missing}(x, y = NULL, ...)

\S4method{crossprod}{strucMatrix,strucMatrix}(x, y = NULL, ...)

\S4method{tcrossprod}{strucMatrix,strucMatrix}(x, y = NULL, ...)
}
\arguments{
\item{e1,e2}{\code{strucMatrix} objects}

\item{x,y}{\code{\link{strucMatrix}} matrix objects}

\item{...}{not used (only for consistency with generic)}
}
\description{
Because of the marginal summations involved, the results of a
matrix multiplication of \code{\link{strucMatrix}} objects is not
profitably stored as a \code{strucMatrix}. Therefore, the output is
a \code{\link{sparseMatrix}}.
}
\note{
The \code{*} operator is matrix multiplication, not
element-wise multiplication.
}

