% Generated by roxygen2 (4.0.1): do not edit by hand
\name{orthProcrustesRotMat}
\alias{orthProcrustesRotMat}
\title{Orthogonal procrustean rotation matrix}
\usage{
orthProcrustesRotMat(X, Z)
}
\arguments{
\item{X}{input matrix}

\item{Z}{target matrix}
}
\value{
the rotation matrix
}
\description{
Orthogonal procrustean rotation matrix
}

