% Generated by roxygen2 (4.0.1): do not edit by hand
\name{packReTrm}
\alias{packReTrm}
\title{Package up a random effects term structure}
\usage{
packReTrm(object, Zt, Lambdat, lowerLoads, upperLoads, lowerCovar, upperCovar,
  devfunEnv = NULL)
}
\arguments{
\item{Zt}{transposed model matrix}

\item{Lambdat}{relative covariance factor}

\item{lowerLoads,upperLoads,lowerCovar,upperCovar}{lower and upper
bounds on possible loadings parameters (for \code{Zt}) and
possible covariance parameters (for \code{Lambdat})}
}
\description{
This function is useful as the final line of a method for
\code{\link{setReTrm}}.
}

