% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{parLength}
\alias{parLength}
\alias{parLength.strucGlmer}
\alias{parLength.strucMatrix}
\alias{parLength.strucParseFormula}
\title{Lengths of parameter vectors}
\usage{
parLength(object, ...)

\method{parLength}{strucMatrix}(object, ...)

\method{parLength}{strucParseFormula}(object, ...)

\method{parLength}{strucGlmer}(object, ...)
}
\arguments{
\item{object}{parameterized object}

\item{...}{extra parameters}
}
\description{
Lengths of parameter vectors
}

