% Generated by roxygen2 (4.0.1): do not edit by hand
\name{covar}
\alias{covar}
\alias{loads}
\alias{mkParInds}
\alias{pars}
\title{Get parameters}
\usage{
covar(object, ...)

loads(object, ...)

pars(object, ...)

mkParInds(parList)
}
\arguments{
\item{object}{\code{lme4ord} fitted model object}

\item{parList}{named list of parameters with possible names:
(\code{covar}, \code{fixef}, \code{weigh}, \code{loads})}
}
\description{
Get parameters
}

