% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pglmer}
\alias{pglmer}
\title{Phylogenetic generalized linear mixed model}
\usage{
pglmer(formula, data = NULL, family = binomial, covList = list(),
  optControl = list(iprint = 0L), ...)
}
\arguments{
\item{formula}{formula}

\item{data}{data}

\item{family}{family}

\item{covList}{covList}

\item{optControl}{optControl}

\item{...}{...}
}
\description{
Phylogenetic generalized linear mixed model
}

