% Generated by roxygen2 (4.0.1): do not edit by hand
\name{printReTrm}
\alias{printReTrm}
\alias{printReTrm.default}
\title{Print random effects term}
\usage{
printReTrm(object, forSummary = FALSE, ...)

\method{printReTrm}{default}(object, forSummary = FALSE, ...)
}
\arguments{
\item{object}{\code{\link{repSparse}} object}

\item{forSummary}{print for \code{\link{summary}} instead of
\code{\link{print}}?}

\item{...}{additional arguments}
}
\description{
Print random effects term
}

