% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repSparse-class}
\alias{dim.repSparse}
\alias{image.repSparse}
\alias{print.repSparse}
\alias{repSparse-class}
\alias{sort.repSparse}
\alias{t.repSparse}
\alias{update.repSparse}
\title{\code{repSparse} class}
\usage{
print.repSparse(x, ...)

update.repSparse(object, newPars, ...)

image.repSparse(x, ...)

t.repSparse(x)

\method{sort}{repSparse}(x, decreasing = FALSE, type = c("col", "row",
  "val"), ...)

\method{dim}{repSparse}(x)
}
\arguments{
\item{x}{\code{repSparse} object}

\item{...}{passed to subsequent functions}

\item{object}{\code{repSparse} object}

\item{newPars}{new parameter values}

\item{decreasing}{see \code{\link{sort}}}

\item{type}{sort by column, row, or value indices?}
}
\description{
\code{repSparse} class
}

