% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repSparse}
\alias{repSparse}
\title{Sparse matrix with repeated values}
\usage{
repSparse(rowInds, colInds, valInds, vals, trans, Dim)
}
\arguments{
\item{rowInds}{1-based row indices (converted to 0-based in
output)}

\item{colInds}{1-based column indices (converted to 0-based in
output)}

\item{valInds}{indices for the nonzero values}

\item{vals}{nonzero values}

\item{trans}{function for transforming some parameters into
\code{vals}.  the \code{\link{environment}} of this function
should contain a vector of these parameters (called \code{init}),
which could be taken as the arguments to \code{trans}.  these
initial values can be obtained by \code{\link{getInit}} and set by
\code{\link{setInit}}.  an \code{\link{update.repSparse}} method
will update these parameters.}

\item{Dim}{matrix dimensions}
}
\value{
object of class \code{repSparse} with list elements
\code{rowInds}, \code{colInds}, \code{valInds}, \code{vals}, and
\code{trans}, and a \code{Dim} attibute
}
\description{
Sparse matrix with repeated values
}

