% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repSparseBlank}
\alias{repSparseBlank}
\title{Blank repeated sparse matrix}
\usage{
repSparseBlank(nrow, ncol)
}
\arguments{
\item{nrow,ncol}{number of rows and columns}
}
\description{
Blank repeated sparse matrix
}
\examples{
(xBlank <- repSparseBlank(5, 5))
}
\seealso{
Other repSparseSpecial: \code{\link{repSparseCol}};
  \code{\link{repSparseCompSymm}};
  \code{\link{repSparseConstVarChol}};
  \code{\link{repSparseCorFactor}};
  \code{\link{repSparseCorMatChol}};
  \code{\link{repSparseDiag}};
  \code{\link{repSparseExpChol}};
  \code{\link{repSparseFull}};
  \code{\link{repSparseIdent}}; \code{\link{repSparseInd}};
  \code{\link{repSparseOneOffDiag}};
  \code{\link{repSparseOnes}}; \code{\link{repSparseSymm}};
  \code{\link{repSparseTri}};
  \code{\link{repSparseVarWithCovariate}}
}

