% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repSparseCompSymm}
\alias{repSparseCompSymm}
\title{Repeated sparse matrix with compound symmetry}
\usage{
repSparseCompSymm(diagVal, offDiagVal, matSize)
}
\arguments{
\item{diagVal}{value for the diagonal}

\item{offDiagVal}{value for the off-diagonal}

\item{matSize}{size of the resulting matrix}
}
\description{
Repeated sparse matrix with compound symmetry
}

