% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repSparseConstVarChol}
\alias{repSparseConstVarChol}
\title{Repeated sparse Cholesky factor leading to constant variance}
\usage{
repSparseConstVarChol(sdVal, offDiagVals)
}
\arguments{
\item{sdVal}{standard deviation of crossproduct of the result
factor}

\item{offDiagVals}{off diagonal values of the Cholesky factor}
}
\description{
Repeated sparse Cholesky factor leading to constant variance
}
\examples{
set.seed(1)
(xConstVarChol <- repSparseConstVarChol(3, rnorm(choose(5, 2))))
}
\seealso{
Other repSparseSpecial: \code{\link{repSparseBlank}};
  \code{\link{repSparseCol}};
  \code{\link{repSparseCompSymm}};
  \code{\link{repSparseCorFactor}};
  \code{\link{repSparseCorMatChol}};
  \code{\link{repSparseDiag}};
  \code{\link{repSparseExpChol}};
  \code{\link{repSparseFull}};
  \code{\link{repSparseIdent}}; \code{\link{repSparseInd}};
  \code{\link{repSparseOneOffDiag}};
  \code{\link{repSparseOnes}}; \code{\link{repSparseSymm}};
  \code{\link{repSparseTri}};
  \code{\link{repSparseVarWithCovariate}}
}

