% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repSparseCorFactor}
\alias{repSparseCorFactor}
\title{Construct a repeated sparse upper Cholesky factor from an
\code{nlme}-style \code{corStruct} object}
\usage{
repSparseCorFactor(object, sig = 1)
}
\arguments{
\item{object}{a \code{corStruct} object}

\item{sig}{initial standard deviation}
}
\description{
Construct a repeated sparse upper Cholesky factor from an
\code{nlme}-style \code{corStruct} object
}
\examples{
(xCorFactor <- repSparseCorFactor(Initialize(corAR1(0.5, form = ~ 1 | Subject), data = Orthodont)))
}
\seealso{
Other repSparseSpecial: \code{\link{repSparseBlank}};
  \code{\link{repSparseCol}};
  \code{\link{repSparseCompSymm}};
  \code{\link{repSparseConstVarChol}};
  \code{\link{repSparseCorMatChol}};
  \code{\link{repSparseDiag}};
  \code{\link{repSparseExpChol}};
  \code{\link{repSparseIdent}}; \code{\link{repSparseInd}};
  \code{\link{repSparseOneOffDiag}};
  \code{\link{repSparseOnes}}; \code{\link{repSparseSymm}};
  \code{\link{repSparseTri}};
  \code{\link{repSparseVarWithCovariate}}
}

