% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repSparseCorMatChol}
\alias{repSparseCorMatChol}
\title{Repeated sparse Cholesky factor of a correlation matrix}
\usage{
repSparseCorMatChol(offDiagPars)
}
\arguments{
\item{offDiagPars}{parameters determining the off-diagonal of the
Cholesky factor}
}
\description{
Repeated sparse Cholesky factor of a correlation matrix
}
\examples{
set.seed(1)
(xCorMatChol <- repSparseCorMatChol(rnorm(choose(5, 2))))
}
\seealso{
Other repSparseSpecial: \code{\link{repSparseBlank}};
  \code{\link{repSparseCol}};
  \code{\link{repSparseCompSymm}};
  \code{\link{repSparseConstVarChol}};
  \code{\link{repSparseCorFactor}};
  \code{\link{repSparseDiag}};
  \code{\link{repSparseExpChol}};
  \code{\link{repSparseFull}};
  \code{\link{repSparseIdent}}; \code{\link{repSparseInd}};
  \code{\link{repSparseOneOffDiag}};
  \code{\link{repSparseOnes}}; \code{\link{repSparseSymm}};
  \code{\link{repSparseTri}};
  \code{\link{repSparseVarWithCovariate}}
}

