% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repSparseDiag}
\alias{repSparseDiag}
\title{Diagonal repeated sparse matrix}
\usage{
repSparseDiag(vals, valInds)
}
\arguments{
\item{vals}{vector of values}

\item{valInds}{vector of value indices}
}
\description{
Diagonal repeated sparse matrix
}

