% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repSparseExpChol}
\alias{repSparseExpChol}
\title{Cholesky factor of a repeated sparse covariance matrix obeying
exponential distance decay in covariance}
\usage{
repSparseExpChol(distObj, cutOffDist = Inf)
}
\arguments{
\item{distObj}{distance matrix object}

\item{cutOffDist}{maximum distance with nonzero correlations}
}
\description{
Cholesky factor of a repeated sparse covariance matrix obeying
exponential distance decay in covariance
}
\examples{
(xExpChol <- repSparseExpChol(dist(matrix(rnorm(10), 5, 2))))
}
\seealso{
Other repSparseSpecial: \code{\link{repSparseBlank}};
  \code{\link{repSparseCol}};
  \code{\link{repSparseCompSymm}};
  \code{\link{repSparseConstVarChol}};
  \code{\link{repSparseCorFactor}};
  \code{\link{repSparseCorMatChol}};
  \code{\link{repSparseDiag}};
  \code{\link{repSparseIdent}}; \code{\link{repSparseInd}};
  \code{\link{repSparseOneOffDiag}};
  \code{\link{repSparseOnes}}; \code{\link{repSparseSymm}};
  \code{\link{repSparseTri}};
  \code{\link{repSparseVarWithCovariate}}
}

