% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repSparseFull}
\alias{repSparseFull}
\title{Full repeated sparse matrix}
\usage{
repSparseFull(nrow, ncol, vals)
}
\arguments{
\item{nrow,ncol}{numbers of rows and columns}

\item{vals}{vector of values}
}
\description{
Full repeated sparse matrix
}
\examples{
set.seed(1)
(xFull <- repSparseFull(5, 5, rnorm(25)))
}
\seealso{
Other repSparseSpecial: \code{\link{repSparseBlank}};
  \code{\link{repSparseCol}};
  \code{\link{repSparseCompSymm}};
  \code{\link{repSparseConstVarChol}};
  \code{\link{repSparseCorFactor}};
  \code{\link{repSparseCorMatChol}};
  \code{\link{repSparseDiag}};
  \code{\link{repSparseExpChol}};
  \code{\link{repSparseIdent}}; \code{\link{repSparseInd}};
  \code{\link{repSparseOneOffDiag}};
  \code{\link{repSparseOnes}}; \code{\link{repSparseSymm}};
  \code{\link{repSparseTri}};
  \code{\link{repSparseVarWithCovariate}}
}

