% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repSparseGenFullTri}
\alias{repSparseGenFullTri}
\title{General and full triangular repeated sparse matrix}
\usage{
repSparseGenFullTri(nrow, ncol, vals, diag = TRUE, low = TRUE)
}
\arguments{
\item{nrow,ncol}{number of rows and columns}

\item{vals}{values for the nonzero values}

\item{diag}{include diagonal?}

\item{low}{lower triangular?}
}
\description{
General and full triangular repeated sparse matrix
}
\note{
Beware there appear to be bugs here.
}
\examples{
set.seed(1)
(xGenTri <- repSparseGenFullTri(5, 5, rnorm(choose(6, 2))))
}
\seealso{
Other repSparseSpecial: \code{\link{repSparseBlank}};
  \code{\link{repSparseCol}};
  \code{\link{repSparseCompSymm}};
  \code{\link{repSparseConstVarChol}};
  \code{\link{repSparseCorFactor}};
  \code{\link{repSparseCorMatChol}};
  \code{\link{repSparseDiag}};
  \code{\link{repSparseExpChol}};
  \code{\link{repSparseFull}};
  \code{\link{repSparseIdent}}; \code{\link{repSparseInd}};
  \code{\link{repSparseOneOffDiag}};
  \code{\link{repSparseOnes}}; \code{\link{repSparseSymm}};
  \code{\link{repSparseTri}};
  \code{\link{repSparseVarWithCovariate}}
}

