% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repSparseIdent}
\alias{repSparseIdent}
\title{Identity repeated sparse matrix}
\usage{
repSparseIdent(matSize)
}
\arguments{
\item{matSize}{matrix size}
}
\description{
Identity repeated sparse matrix
}
\examples{
(xIdent <- repSparseIdent(5))
}
\seealso{
Other repSparseSpecial: \code{\link{repSparseBlank}};
  \code{\link{repSparseCol}};
  \code{\link{repSparseCompSymm}};
  \code{\link{repSparseConstVarChol}};
  \code{\link{repSparseCorFactor}};
  \code{\link{repSparseCorMatChol}};
  \code{\link{repSparseDiag}};
  \code{\link{repSparseExpChol}};
  \code{\link{repSparseInd}};
  \code{\link{repSparseOneOffDiag}};
  \code{\link{repSparseOnes}}; \code{\link{repSparseSymm}};
  \code{\link{repSparseTri}};
  \code{\link{repSparseVarWithCovariate}}
}

