% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repSparseInd}
\alias{repSparseInd}
\title{Repeated sparse indicator matrix}
\usage{
repSparseInd(fac)
}
\arguments{
\item{fac}{vector coercible to factor}
}
\description{
Repeated sparse indicator matrix
}
\examples{
(xInd <- repSparseInd(rep(1:3, c(2, 2, 1))))
}
\seealso{
Other repSparseSpecial: \code{\link{repSparseBlank}};
  \code{\link{repSparseCol}};
  \code{\link{repSparseCompSymm}};
  \code{\link{repSparseConstVarChol}};
  \code{\link{repSparseCorMatChol}};
  \code{\link{repSparseDiag}};
  \code{\link{repSparseExpChol}};
  \code{\link{repSparseIdent}};
  \code{\link{repSparseOneOffDiag}};
  \code{\link{repSparseOnes}}; \code{\link{repSparseSymm}};
  \code{\link{repSparseTri}};
  \code{\link{repSparseVarWithCovariate}}
}

