% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repSparseOneOffDiag}
\alias{repSparseOneOffDiag}
\title{Repeated sparse covariance matrix with equal variances and one
off-diagonal covariance}
\usage{
repSparseOneOffDiag(diagVal, offDiagVal, offDiagInds, matSize)
}
\arguments{
\item{diagVal}{value for the diagonal}

\item{offDiagVal}{value for the off-diagonal}

\item{offDiagInds}{indices for the two correlated objects}

\item{matSize}{size of the resulting matrix}
}
\description{
Repeated sparse covariance matrix with equal variances and one
off-diagonal covariance
}

