% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repSparseOnes}
\alias{repSparseOnes}
\title{Repeated sparse matrix of ones}
\usage{
repSparseOnes(nrow, ncol)
}
\arguments{
\item{nrow,ncol}{numbers of rows and columns}
}
\description{
Repeated sparse matrix of ones
}
\examples{
(xOnes <- repSparseOnes(5, 5))
}
\seealso{
Other repSparseSpecial: \code{\link{repSparseBlank}};
  \code{\link{repSparseCol}};
  \code{\link{repSparseCompSymm}};
  \code{\link{repSparseConstVarChol}};
  \code{\link{repSparseCorFactor}};
  \code{\link{repSparseCorMatChol}};
  \code{\link{repSparseDiag}};
  \code{\link{repSparseExpChol}};
  \code{\link{repSparseIdent}}; \code{\link{repSparseInd}};
  \code{\link{repSparseOneOffDiag}};
  \code{\link{repSparseSymm}}; \code{\link{repSparseTri}};
  \code{\link{repSparseVarWithCovariate}}
}

