% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repSparseBlank}
\alias{repSparseBlank}
\alias{repSparseCol}
\alias{repSparseCompSymm}
\alias{repSparseConstVarChol}
\alias{repSparseCorMatChol}
\alias{repSparseDiag}
\alias{repSparseExpChol}
\alias{repSparseIdent}
\alias{repSparseInd}
\alias{repSparseOneOffDiag}
\alias{repSparseOnes}
\alias{repSparseSpecial}
\alias{repSparseSymm}
\alias{repSparseTri}
\alias{repSparseVarWithCovariate}
\title{Special sparse repeated matrices}
\usage{
repSparseBlank(nrow, ncol)

repSparseIdent(matSize)

repSparseDiag(vals, valInds)

repSparseCol(vals, valInds)

repSparseInd(fac)

repSparseTri(diagVals, offDiagVals, low = TRUE)

repSparseOnes(nrow, ncol)

repSparseSymm(diagVals, offDiagVals)

repSparseCompSymm(diagVal, offDiagVal, matSize)

repSparseOneOffDiag(diagVal, offDiagVal, offDiagInds, matSize)

repSparseConstVarChol(sdVal, offDiagVals)

repSparseCorMatChol(offDiagPars)

repSparseVarWithCovariate(varPars, covariate, grpFac,
  mkTransFunc = mkVarExpTrans)

repSparseExpChol(distObj, cutOffDist = Inf)
}
\arguments{
\item{nrow,ncol}{number of rows and columns}

\item{vals}{vector of values}

\item{valInds}{vector of value indices}

\item{fac}{vector coercible to factor}

\item{diagVals}{values for the diagonal}

\item{offDiagVals}{values for the off-diagonal}

\item{low}{lower triangular?}

\item{diagVal}{value for the diagonal}

\item{offDiagVal}{value for the off-diagonal}

\item{matSize}{size of the resulting matrix}

\item{offDiagInds}{indices for the two correlated objects}

\item{sdVal}{standard deviation of crossproduct of the result}

\item{offDiagVals}{values for the off-diagonal of the Cholesky
factor}

\item{offDiagPars}{parameters determining the off-diagonal of the
Cholesky factor}

\item{varPars}{vector of variance parameters (one per level of
\code{grpFac})}

\item{covariate}{covariate}

\item{grpFac}{a grouping factor (or anything coercible to
\code{numeric} really) with the number of levels equal to the
length of \code{varPars}}

\item{mkTransFunc}{a function taking arguments \code{varPars},
\code{covariate}, and \code{grpFac} for constructing a
\code{trans} function (see \code{\link{mkVarExpTrans}} for an
example).}

\item{distObj}{distance matrix object}
}
\description{
Special sparse repeated matrices
}
\examples{
(xBlank <- repSparseBlank(5, 5))
(xIdent <- repSparseIdent(5))
set.seed(1)
(xDiag <- repSparseDiag(rnorm(5)))
set.seed(1)
(xCol <- repSparseCol(rnorm(5)))
(xInd <- repSparseInd(rep(1:3, c(2, 2, 1))))
set.seed(1)
(xTri <- repSparseTri(rnorm(5), rnorm(choose(5, 2))))
(xOnes <- repSparseOnes(5, 5))
set.seed(1)
(xSymm <- repSparseSymm(rnorm(5), rnorm(choose(5, 2))))
(xCompSymm <- repSparseCompSymm(1, -0.2, 5))
(xOneOffDiag <- repSparseOneOffDiag(1, -0.2, c(5, 2), 5))
set.seed(1)
(xConstVarChol <- repSparseConstVarChol(3, rnorm(choose(5, 2))))
set.seed(1)
(xCorMatChol <- repSparseCorMatChol(rnorm(choose(5, 2))))
(xVarWithCovariate <- repSparseVarWithCovariate(rep(1, 5), 0.05*(4:0)))
(xExpChol <- repSparseExpChol(dist(matrix(rnorm(10), 5, 2))))
}
\seealso{
Other repeated.sparse.matrix.topics: \code{\link{.bind}},
  \code{\link{bind}}, \code{\link{rep.repSparse}};
  \code{\link{as.data.frame.repSparse}},
  \code{\link{as.data.table.repSparse}},
  \code{\link{as.matrix.repSparse}},
  \code{\link{as.repSparse}},
  \code{\link{as.repSparse.dCHMsimpl}},
  \code{\link{as.repSparse.dist}},
  \code{\link{as.repSparse.dsparseMatrix}},
  \code{\link{as.repSparse.factor}},
  \code{\link{as.repSparse.matrix}},
  \code{\link{as.repSparse.repSparse}},
  \code{\link{repSparse2CsparseMatrix}},
  \code{\link{repSparse2TsparseMatrix}},
  \code{\link{repSparse2dgCMatrix}},
  \code{\link{repSparse2dgTMatrix}},
  \code{\link{repSparse2sparseMatrix}};
  \code{\link{chol.repSparse}},
  \code{\link{chol.repSparseCompSymm}},
  \code{\link{chol.repSparseOneOffDiag}};
  \code{\link{diag.repSparse}},
  \code{\link{dim.repSparse}},
  \code{\link{image.repSparse}},
  \code{\link{plot.repSparse}},
  \code{\link{print.repSparse}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}}, \code{\link{t.repSparse}},
  \code{\link{update.repSparse}}; \code{\link{getInit}},
  \code{\link{getInit.default}},
  \code{\link{getInit.function}},
  \code{\link{getInit.repSparse}}, \code{\link{setInit}},
  \code{\link{setInit.default}},
  \code{\link{setInit.function}},
  \code{\link{setInit.repSparse}}; \code{\link{ind2point}},
  \code{\link{point2ind}}; \code{\link{kr}},
  \code{\link{kron}}; \code{\link{mkCholCompSymmTrans}},
  \code{\link{mkCholOneOffDiagTrans}},
  \code{\link{mkConstVarCholTrans}},
  \code{\link{mkCorMatCholTrans}},
  \code{\link{mkGenCholInPlaceTrans}},
  \code{\link{mkIdentityTrans}}, \code{\link{mkListTrans}},
  \code{\link{mkOuterTrans}}, \code{\link{mkTrans}},
  \code{\link{mkTransExpChol}},
  \code{\link{mkVarExpTrans}},
  \code{\link{mkVarFixedTrans}},
  \code{\link{mkVarIdentTrans}},
  \code{\link{mkVarPowerTrans}}; \code{\link{repSparse}}
}

