% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repSparseSymm}
\alias{repSparseSymm}
\title{Symmetric repeated sparse matrix}
\usage{
repSparseSymm(diagVals, offDiagVals)
}
\arguments{
\item{diagVals}{values for the diagonal}

\item{offDiagVals}{values for the off-diagonal}
}
\description{
Symmetric repeated sparse matrix
}
\examples{
set.seed(1)
(xSymm <- repSparseSymm(rnorm(5), rnorm(choose(5, 2))))
}
\seealso{
Other repSparseSpecial: \code{\link{repSparseBlank}};
  \code{\link{repSparseCol}};
  \code{\link{repSparseCompSymm}};
  \code{\link{repSparseConstVarChol}};
  \code{\link{repSparseCorMatChol}};
  \code{\link{repSparseDiag}};
  \code{\link{repSparseExpChol}};
  \code{\link{repSparseIdent}}; \code{\link{repSparseInd}};
  \code{\link{repSparseOneOffDiag}};
  \code{\link{repSparseOnes}}; \code{\link{repSparseTri}};
  \code{\link{repSparseVarWithCovariate}}
}

