% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repSparseTri}
\alias{repSparseTri}
\title{Triangular repeated sparse matrix}
\usage{
repSparseTri(diagVals, offDiagVals, low = TRUE)
}
\arguments{
\item{diagVals}{values for the diagonal}

\item{offDiagVals}{values for the off-diagonal}

\item{low}{lower triangular?}
}
\description{
Triangular repeated sparse matrix
}
\examples{
set.seed(1)
(xTri <- repSparseTri(rnorm(5), rnorm(choose(5, 2))))
}
\seealso{
Other repSparseSpecial: \code{\link{repSparseBlank}};
  \code{\link{repSparseCol}};
  \code{\link{repSparseCompSymm}};
  \code{\link{repSparseConstVarChol}};
  \code{\link{repSparseCorFactor}};
  \code{\link{repSparseCorMatChol}};
  \code{\link{repSparseDiag}};
  \code{\link{repSparseExpChol}};
  \code{\link{repSparseFull}};
  \code{\link{repSparseGenFullTri}};
  \code{\link{repSparseIdent}}; \code{\link{repSparseInd}};
  \code{\link{repSparseOneOffDiag}};
  \code{\link{repSparseOnes}}; \code{\link{repSparseSymm}};
  \code{\link{repSparseVarWithCovariate}}
}

