% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repSparseVarWithCovariate}
\alias{repSparseVarWithCovariate}
\title{Repeated sparse diagonal covariance matrix with a covariate
determining the diagonal}
\usage{
repSparseVarWithCovariate(varPars, covariate, grpFac,
  mkTransFunc = mkVarExpTrans)
}
\arguments{
\item{varPars}{vector of variance parameters (one per level of
\code{grpFac})}

\item{covariate}{covariate}

\item{grpFac}{a grouping factor (or anything coercible to
\code{numeric} really) with the number of levels equal to the
length of \code{varPars}}

\item{mkTransFunc}{a function taking arguments \code{varPars},
\code{covariate}, and \code{grpFac} for constructing a
\code{trans} function (see \code{\link{mkVarExpTrans}} for an
example).}
}
\description{
Repeated sparse diagonal covariance matrix with a covariate
determining the diagonal
}
\examples{
(xVarWithCovariate <- repSparseVarWithCovariate(rep(1, 5), 0.05*(4:0)))
}
\seealso{
Other repSparseSpecial: \code{\link{repSparseBlank}};
  \code{\link{repSparseCol}};
  \code{\link{repSparseCompSymm}};
  \code{\link{repSparseConstVarChol}};
  \code{\link{repSparseCorFactor}};
  \code{\link{repSparseCorMatChol}};
  \code{\link{repSparseDiag}};
  \code{\link{repSparseExpChol}};
  \code{\link{repSparseIdent}}; \code{\link{repSparseInd}};
  \code{\link{repSparseOneOffDiag}};
  \code{\link{repSparseOnes}}; \code{\link{repSparseSymm}};
  \code{\link{repSparseTri}}
}

