% Generated by roxygen2 (4.0.1): do not edit by hand
\name{resetTransConst}
\alias{resetTransConst}
\title{Reset the transformation function of a repeated sparse matrix}
\usage{
resetTransConst(object)
}
\arguments{
\item{object}{repeated sparse matrix}
}
\description{
Reset the transformation function of a repeated sparse matrix
}

