% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structuredSparseMatrices.R
\name{resetTransConst}
\alias{resetTransConst}
\title{Constant structured sparse matrix}
\usage{
resetTransConst(object)
}
\arguments{
\item{object}{structured sparse matrix}
}
\description{
Convert a parameterized structured sparse matrix into a constant
structured sparse matrix (i.e. \code{getInit(object)} has length
zero).
}
\seealso{
Other modifications: \code{\link{scalarMult}},
  \code{\link{simplifyStrucMatrix}}
}

