% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rowWiseCombination}
\alias{rowWiseCombination}
\title{Row-wise combination of two repeated sparse matrices}
\usage{
rowWiseCombination(X, Y)
}
\arguments{
\item{X,Y}{\code{repSparse} objects}
}
\value{
a row-wise combination of repeated sparse matrices
}
\description{
Row-wise combination of two repeated sparse matrices
}

