% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rRepSparse}
\alias{rRepSparse}
\title{Random repeated sparse matrices}
\usage{
rRepSparse(nrows, ncols, nvals, nnonzeros, rfunc = rnorm, ...)
}
\arguments{
\item{nrows,ncols}{numbers of rows and columns}

\item{nvals}{number of values}

\item{nnonzeros}{number of nonzero elements}

\item{rfunc}{random number function}

\item{...}{dots}
}
\description{
Random repeated sparse matrices
}

