% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structuredSparseMatrices.R
\name{rStrucMatrix}
\alias{rStrucMatrix}
\title{Random structured sparse matrices}
\usage{
rStrucMatrix(nrows, ncols, nvals, nnonzeros, rfunc = rnorm, ...)
}
\arguments{
\item{nrows, ncols}{numbers of rows and columns}

\item{nvals}{number of values}

\item{nnonzeros}{number of nonzero elements}

\item{rfunc}{random number function}

\item{...}{dots}
}
\description{
Random structured sparse matrices
}

