% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setLowerDefault}
\alias{setLowerDefault}
\alias{setUpperDefault}
\title{Get defaults choices for lower and/or upper bound of a model
parameter}
\usage{
setLowerDefault(init, ...)

setUpperDefault(init, ...)
}
\arguments{
\item{init}{initial parameter value}

\item{...}{additional arguments not currently used}
}
\description{
Get defaults choices for lower and/or upper bound of a model
parameter
}
\examples{
init <- rep(c(-1, 0, 1), each = 3)
setLowerDefault(init)
setUpperDefault(init)
}

