% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setReTrm.cooccur}
\alias{cooccur}
\alias{setReTrm.cooccur}
\title{Random effects term for cooccurence model (experimental)}
\usage{
\method{setReTrm}{cooccur}(object, addArgsList, devfunEnv = NULL)
}
\arguments{
\item{object}{a \code{reTrmStruct} object}

\item{addArgsList}{a list of named quantities within which
\code{addArgsExpr} is evaluated}

\item{devfunEnv}{optional environment of the deviance function}
}
\description{
Random effects term for cooccurence model (experimental)
}
\section{Mixed model formula usage}{

\code{cooccur(linForm | grpFac)}
\describe{
\item{linForm}{linear model formula decribing effects} \item{grpFac}{grouping factor}
}
}
\seealso{
Other setReTrm: \code{\link{edge}},
  \code{\link{setReTrm.edge}}; \code{\link{expDecay}},
  \code{\link{setReTrm.expDecay}}; \code{\link{factAnal}},
  \code{\link{setReTrm.factAnal}};
  \code{\link{findReTrmClasses}}, \code{\link{getAddArgs}},
  \code{\link{packReTrm}}, \code{\link{setReTrm}},
  \code{\link{setReTrm.default}}; \code{\link{flexvar}},
  \code{\link{setReTrm.flexvar}}; \code{\link{identity}},
  \code{\link{setReTrm.identity}}; \code{\link{lme4}},
  \code{\link{setReTrm.lme4}}; \code{\link{nlmeCorStruct}},
  \code{\link{setReTrm.nlmeCorStruct}};
  \code{\link{obslev}}, \code{\link{setReTrm.obslev}};
  \code{\link{sem}}, \code{\link{setReTrm.sem}};
  \code{\link{setReTrm.varExp}}, \code{\link{varExp}};
  \code{\link{setReTrm.varIdent}}, \code{\link{varIdent}}
}

