% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomEffectsStructures.R
\name{setReTrm.expDecay}
\alias{expDecay}
\alias{setReTrm.expDecay}
\title{Random effects term with exponential decay in distance-based covariance}
\usage{
\method{setReTrm}{expDecay}(object, addArgsList, auxEnv = NULL,
  devfunEnv = NULL)
}
\arguments{
\item{object}{a \code{\link{reTrmStruct}} object}

\item{addArgsList}{a list of named quantities within which
\code{addArgsExpr} is evaluated}

\item{auxEnv}{an optional auxilliary environment containing
objects that are possibly required for the setting of a random
effects term structure (currently this is almost always the
environment of a call to \code{\link{strucParseFormula}}, unless
\code{setReTrm} is called directly)}

\item{devfunEnv}{optional environment of the deviance function}
}
\description{
Random effects term with exponential decay in distance-based covariance
}
\section{Mixed model formula usage}{

\code{expDecay(1 | grpFac, distCutoff, minCov, distMat)}
\describe{
\item{grpFac:}{grouping factor (e.g. with levels given by geographical sites)} \item{distCutoff:}{maximum distance with covariance greater than minCov} \item{minCov:}{minimum covariance} \item{distMat:}{distance matrix object over the levels of grpFac}
}
}

\section{Parameters}{

\describe{
\item{Covariance:}{length-two vector with the exponential decay parameter and the standard deviation at zero distance}
\item{Loadings:}{none}
}
}
\seealso{
Other setReTrm: \code{\link{setReTrm.cooccur}},
  \code{\link{setReTrm.factAnal}},
  \code{\link{setReTrm.flexvar}},
  \code{\link{setReTrm.identity}},
  \code{\link{setReTrm.lme4}},
  \code{\link{setReTrm.nlmeCorStruct}},
  \code{\link{setReTrm.obslev}},
  \code{\link{setReTrm.phyloEdge}},
  \code{\link{setReTrm.sem}},
  \code{\link{setReTrm.varExp}},
  \code{\link{setReTrm.varIdent}}, \code{\link{setReTrm}}
}

