% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomEffectsStructures.R
\name{setReTrm.factAnal}
\alias{biplot.factAnal}
\alias{factAnal}
\alias{scores.factAnal}
\alias{setReTrm.factAnal}
\title{Random effects term for factor analysis}
\usage{
\method{setReTrm}{factAnal}(object, addArgsList, auxEnv = NULL,
  devfunEnv = NULL)

\method{scores}{factAnal}(x, ...)

\method{biplot}{factAnal}(x, ...)
}
\arguments{
\item{object}{a \code{\link{reTrmStruct}} object}

\item{addArgsList}{a list of named quantities within which
\code{addArgsExpr} is evaluated}

\item{auxEnv}{an optional auxilliary environment containing
objects that are possibly required for the setting of a random
effects term structure (currently this is almost always the
environment of a call to \code{\link{strucParseFormula}}, unless
\code{setReTrm} is called directly)}

\item{devfunEnv}{optional environment of the deviance function}

\item{x}{\code{\link{setReTrm.factAnal}} object}

\item{...}{not used (for consistency with generic in \code{vegan}
package)}
}
\description{
Random effects term for factor analysis
}
\section{Mixed model formula usage}{

\code{factAnal(0 + obsFac | varFac, nAxes, seed)}
\describe{
\item{obsFac:}{grouping factor for multivariate observations} \item{varFac:}{grouping factor for variables} \item{nAxes:}{number of axes} \item{seed:}{random seed for initial axis values}
}
}

\section{Parameters}{

\describe{
\item{Covariance:}{none}
\item{Loadings:}{elements of the lower triangle of the variables by axes loadings matrix}
}
}
\seealso{
Other setReTrm: \code{\link{setReTrm.cooccur}},
  \code{\link{setReTrm.expDecay}},
  \code{\link{setReTrm.flexvar}},
  \code{\link{setReTrm.identity}},
  \code{\link{setReTrm.lme4}},
  \code{\link{setReTrm.nlmeCorStruct}},
  \code{\link{setReTrm.obslev}},
  \code{\link{setReTrm.phyloEdge}},
  \code{\link{setReTrm.sem}},
  \code{\link{setReTrm.varExp}},
  \code{\link{setReTrm.varIdent}}, \code{\link{setReTrm}}
}

