% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setReTrm.flexvar}
\alias{flexvar}
\alias{setReTrm.flexvar}
\title{Random effects term for a flexible variance function}
\usage{
\method{setReTrm}{flexvar}(object, addArgsList, devfunEnv = NULL)
}
\arguments{
\item{object}{a \code{reTrmStruct} object}

\item{addArgsList}{a list of named quantities within which
\code{addArgsExpr} is evaluated}

\item{devfunEnv}{optional environment of the deviance function}
}
\description{
Random effects term for a flexible variance function
}
\section{Mixed model formula usage}{

\code{flexvar(linForm, init, nBasis)}
\describe{
\item{linForm}{linear model formula decribing effects} \item{init}{initial values for coefficients for the linear predictor of the variance function} \item{nBasis}{number of coefficients for the linear predictor}
}
}
\seealso{
Other setReTrm: \code{\link{cooccur}},
  \code{\link{setReTrm.cooccur}}; \code{\link{edge}},
  \code{\link{setReTrm.edge}}; \code{\link{expDecay}},
  \code{\link{setReTrm.expDecay}}; \code{\link{factAnal}},
  \code{\link{setReTrm.factAnal}};
  \code{\link{findReTrmClasses}}, \code{\link{getAddArgs}},
  \code{\link{packReTrm}}, \code{\link{setReTrm}},
  \code{\link{setReTrm.default}}; \code{\link{identity}},
  \code{\link{setReTrm.identity}}; \code{\link{obslev}},
  \code{\link{setReTrm.nlmeCorStruct}};
  \code{\link{obslev}}, \code{\link{setReTrm.obslev}};
  \code{\link{setReTrm.varExp}}, \code{\link{varExp}};
  \code{\link{setReTrm.varIdent}}, \code{\link{varIdent}}
}

