% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomEffectsStructures.R
\name{setReTrm.identity}
\alias{identity}
\alias{setReTrm.identity}
\title{Random effects term for covariance proportional to identity matrix}
\usage{
\method{setReTrm}{identity}(object, addArgsList, auxEnv = NULL,
  devfunEnv = NULL)
}
\arguments{
\item{object}{a \code{\link{reTrmStruct}} object}

\item{addArgsList}{a list of named quantities within which
\code{addArgsExpr} is evaluated}

\item{auxEnv}{an optional auxilliary environment containing
objects that are possibly required for the setting of a random
effects term structure (currently this is almost always the
environment of a call to \code{\link{strucParseFormula}}, unless
\code{setReTrm} is called directly)}

\item{devfunEnv}{optional environment of the deviance function}
}
\description{
Random effects term for covariance proportional to identity matrix
}
\section{Mixed model formula usage}{

\code{identity(linForm | grpFac)}
\describe{
\item{linForm:}{linear model formula decribing effects} \item{grpFac:}{grouping factor}
}
}

\section{Parameters}{

\describe{
\item{Covariance:}{standard deviation of the random effects}
\item{Loadings:}{none}
}
}
\seealso{
Other setReTrm: \code{\link{setReTrm.cooccur}},
  \code{\link{setReTrm.expDecay}},
  \code{\link{setReTrm.factAnal}},
  \code{\link{setReTrm.flexvar}},
  \code{\link{setReTrm.lme4}},
  \code{\link{setReTrm.nlmeCorStruct}},
  \code{\link{setReTrm.obslev}},
  \code{\link{setReTrm.phyloEdge}},
  \code{\link{setReTrm.sem}},
  \code{\link{setReTrm.varExp}},
  \code{\link{setReTrm.varIdent}}, \code{\link{setReTrm}}
}

