% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setReTrm.multiVarExp}
\alias{multiVarExp}
\alias{setReTrm.multiVarExp}
\title{Random effects term with exponential variance structure}
\usage{
\method{setReTrm}{multiVarExp}(object, addArgsList, devfunEnv = NULL)
}
\arguments{
\item{object}{a \code{reTrmStruct} object}

\item{addArgsList}{a list of named quantities within which
\code{addArgsExpr} is evaluated}

\item{devfunEnv}{optional environment of the deviance function}
}
\description{
Random effects term with exponential variance structure
}
\section{Mixed model formula usage}{

\code{multiVarExp(linForm | grpFac)} or \code{multiVarExp(linForm)}
\describe{
\item{linForm}{linear model formula decribing effects} \item{grpFac}{grouping factor}
}
}
\seealso{
Other setReTrm: \code{\link{cooccur}},
  \code{\link{setReTrm.cooccur}}; \code{\link{edge}},
  \code{\link{setReTrm.edge}}; \code{\link{expDecay}},
  \code{\link{setReTrm.expDecay}}; \code{\link{factAnal}},
  \code{\link{setReTrm.factAnal}};
  \code{\link{findReTrmClasses}}, \code{\link{getAddArgs}},
  \code{\link{packReTrm}}, \code{\link{setReTrm}},
  \code{\link{setReTrm.default}}; \code{\link{flexvar}},
  \code{\link{setReTrm.flexvar}}; \code{\link{identity}},
  \code{\link{setReTrm.identity}}; \code{\link{obslev}},
  \code{\link{setReTrm.obslev}};
  \code{\link{setReTrm.varIdent}}, \code{\link{varIdent}}
}

