% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomEffectsStructures.R
\name{setReTrm.nlmeCorStruct}
\alias{getCorStruct}
\alias{nlmeCorStruct}
\alias{setReTrm.nlmeCorStruct}
\title{Random effects term from an \code{nlme}-style \code{corStruct}
object}
\usage{
\method{setReTrm}{nlmeCorStruct}(object, addArgsList, auxEnv = NULL,
  devfunEnv = NULL)

getCorStruct(object, ...)
}
\arguments{
\item{object}{\code{nlmeCorStruct} object}

\item{addArgsList}{a list of named quantities within which
\code{addArgsExpr} is evaluated}

\item{auxEnv}{an optional auxilliary environment containing
objects that are possibly required for the setting of a random
effects term structure (currently this is almost always the
environment of a call to \code{\link{strucParseFormula}}, unless
\code{setReTrm} is called directly)}

\item{devfunEnv}{optional environment of the deviance function}

\item{...}{not used (for consistency with generic)}
}
\description{
Random effects term from an \code{nlme}-style \code{corStruct}
object
}
\section{Mixed model formula usage}{

\code{nlmeCorStruct(1, corObj)} or \code{nlmeCorStruct(1 | grpFac, corObj)}
\describe{
\item{corObj:}{corStruct object} \item{sig:}{initial standard deviation}
}
}

\section{Parameters}{

\describe{
\item{Covariance:}{parameters for the correlation structure}
\item{Loadings:}{none}
}
}
\seealso{
Other setReTrm: \code{\link{setReTrm.cooccur}},
  \code{\link{setReTrm.expDecay}},
  \code{\link{setReTrm.factAnal}},
  \code{\link{setReTrm.flexvar}},
  \code{\link{setReTrm.identity}},
  \code{\link{setReTrm.lme4}},
  \code{\link{setReTrm.obslev}},
  \code{\link{setReTrm.phyloEdge}},
  \code{\link{setReTrm.sem}},
  \code{\link{setReTrm.varExp}},
  \code{\link{setReTrm.varIdent}}, \code{\link{setReTrm}}
}

