% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomEffectsStructures.R
\name{setReTrm}
\alias{findReTrmClasses}
\alias{getAddArgs}
\alias{ranef.reTrmStruct}
\alias{setReTrm}
\alias{setReTrm.default}
\title{Set model matrix slice and relative covariance factor block for a
random effects term}
\usage{
setReTrm(object, addArgsList, auxEnv = NULL, devfunEnv = NULL)

\method{setReTrm}{default}(object, addArgsList, auxEnv = NULL,
  devfunEnv = NULL)

findReTrmClasses(formula = NULL)

getAddArgs(addArgsExpr, addArgsList)

\method{ranef}{reTrmStruct}(object, type = c("u", "Lu", "ZLu"), ...)
}
\arguments{
\item{object}{a \code{\link{reTrmStruct}} object}

\item{addArgsList}{a list of named quantities within which
\code{addArgsExpr} is evaluated}

\item{auxEnv}{an optional auxilliary environment containing
objects that are possibly required for the setting of a random
effects term structure (currently this is almost always the
environment of a call to \code{\link{strucParseFormula}}, unless
\code{setReTrm} is called directly)}

\item{devfunEnv}{optional environment of the deviance function}

\item{formula}{generalized mixed model formula.  if \code{NULL}
(the default) \code{findReTrmClasses} returns classes available
(on the search path).}

\item{addArgsExpr}{a list of expressions for evaluating within
\code{addArgsList}}

\item{type}{see \code{\link{ranef.strucGlmer}}}

\item{...}{not used (for consistency with generic)}
}
\value{
\code{object} with the following additional elements:

\item{Zt}{A \code{\link{strucMatrix}} object describing the slice of
the model matrix associated with the random effects term.}

\item{Lambdat}{A \code{\link{strucMatrix}} object describing the
block of the relative covariance factor associated with the random
effects term.}

\item{lowerLoads,upperLoads,lowerCovar,upperCovar}{Lower/upper
bounds on loading parameters (possible parameters for \code{Zt})
and/or covariance parameters (possible parameters for \code{Lambdat})}
}
\description{
Set model matrix slice and relative covariance factor block for a
random effects term
}
\note{
Almost never called directly, but instead called within
\code{\link{strucParseFormula}}.
}
\section{Mixed model formula usage}{

The default method is identical to a standard \code{lme4} random effects term.
\code{(linForm | grpFac)}
\describe{
\item{linForm}{linear model formula}
\item{grpFac}{grouping factor}
}
}
\seealso{
\code{\link{mkReTrmStructs}} for construction of these objects

getReTrm

Other setReTrm: \code{\link{setReTrm.cooccur}},
  \code{\link{setReTrm.expDecay}},
  \code{\link{setReTrm.factAnal}},
  \code{\link{setReTrm.flexvar}},
  \code{\link{setReTrm.identity}},
  \code{\link{setReTrm.lme4}},
  \code{\link{setReTrm.nlmeCorStruct}},
  \code{\link{setReTrm.obslev}},
  \code{\link{setReTrm.phyloEdge}},
  \code{\link{setReTrm.sem}},
  \code{\link{setReTrm.varExp}},
  \code{\link{setReTrm.varIdent}}
}

