% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setReTrm.sem}
\alias{sem}
\alias{setReTrm.sem}
\title{Random effects term for structural equation model}
\usage{
\method{setReTrm}{sem}(object, addArgsList, auxEnv = NULL, devfunEnv = NULL)
}
\arguments{
\item{object}{a \code{reTrmStruct} object}

\item{addArgsList}{a list of named quantities within which
\code{addArgsExpr} is evaluated}

\item{auxEnv}{an optional auxilliary environment containing
objects that are possibly required for the setting of a random
effects term structure (currently this is almost always the
environment of a call to \code{\link{strucParseFormula}}, unless
\code{setReTrm} is called directly)}

\item{devfunEnv}{optional environment of the deviance function}
}
\description{
Random effects term for structural equation model
}
\section{Mixed model formula usage}{

\code{sem(1 | grpFac, loadMat)}
\describe{
\item{grpFac}{grouping factor for loadings} \item{nAxes}{number of axes} \item{seed}{random seed for initial axis values}
}
}
\seealso{
Other setReTrm: \code{\link{biplot.factAnal}},
  \code{\link{factAnal}}, \code{\link{scores.factAnal}},
  \code{\link{setReTrm.factAnal}}; \code{\link{cooccur}},
  \code{\link{setReTrm.cooccur}}; \code{\link{edge}},
  \code{\link{setReTrm.edge}}; \code{\link{expDecay}},
  \code{\link{setReTrm.expDecay}};
  \code{\link{findReTrmClasses}}, \code{\link{getAddArgs}},
  \code{\link{packReTrm}}, \code{\link{ranef.reTrmStruct}},
  \code{\link{setReTrm}}, \code{\link{setReTrm.default}};
  \code{\link{flexvar}}, \code{\link{setReTrm.flexvar}};
  \code{\link{identity}}, \code{\link{setReTrm.identity}};
  \code{\link{lme4}}, \code{\link{setReTrm.lme4}};
  \code{\link{nlmeCorStruct}},
  \code{\link{setReTrm.nlmeCorStruct}};
  \code{\link{obslev}}, \code{\link{setReTrm.obslev}};
  \code{\link{setReTrm.varExp}}, \code{\link{varExp}};
  \code{\link{setReTrm.varIdent}}, \code{\link{varIdent}}
}

