% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomEffectsStructures.R
\name{simAddArgs}
\alias{simAddArgs}
\alias{simAddArgs.default}
\alias{simAddArgs.phyloEdge}
\alias{simAddArgsList}
\title{Simulate additional arguments (experimental)}
\usage{
simAddArgs(object, ...)

simAddArgsList(object, ...)

\method{simAddArgs}{default}(object, ...)

\method{simAddArgs}{phyloEdge}(object, rtreeArgs = list(),
  compute.brlenArgs = list(), ...)
}
\arguments{
\item{object}{a \code{\link{reTrmStruct}} object}

\item{...}{dots}

\item{rtreeArgs}{arguments for \code{\link{rtree}}}

\item{compute.brlenArgs}{arguments for \code{\link{compute.brlen}}}
}
\description{
Simulate additional arguments (experimental)
}

