% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{simplifyFacList}
\alias{simplifyFacList}
\title{Simplify factor list over random effects terms (not currently
used)}
\usage{
simplifyFacList(facList)
}
\arguments{
\item{facList}{list of grouping factors over random effects terms}
}
\value{
collapse repeated factors and add an \code{assign}
attribute for indicating which factors relate to which terms
}
\description{
Simplify factor list over random effects terms (not currently
used)
}

