% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simReTrm}
\alias{simReTrm}
\title{Simulate from a random effects term}
\usage{
simReTrm(object)
}
\arguments{
\item{object}{\code{\link{reTrmStruct}} object}
}
\value{
a numeric vector of length equal to the number of rows of
the model matrix for the term.
}
\description{
The model matrix times the relative covariance factor times a
simulated vector of random effects coefficients.
}
\examples{
example("strucGlmer")
set.seed(1)
simReTrm(getReTrm(gm, "period.factAnal"))
}

