% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputAndInference.R
\name{simStrucParsedForm}
\alias{simStrucParsedForm}
\title{Simulate from a parsed formula}
\usage{
simStrucParsedForm(parsedForm, family = binomial, weights, nsim)
}
\arguments{
\item{parsedForm}{result of \code{strucParseFormula}}

\item{family}{family object}

\item{weights}{optional weights}

\item{nsim}{number of simulations (FIXME: inconsistent between
binomial and poisson)}
}
\description{
Simulate from a parsed formula
}

