% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simTestPhyloDat}
\alias{simTestPhyloDat}
\title{Generate phylogenetic test data}
\usage{
simTestPhyloDat(seed = 1, n = 10, m = 30, form = y ~ 1 + (1 | species),
  power = 0.1, covarSim = 1, fixefSim = 1)
}
\arguments{
\item{seed}{random seed}

\item{n}{number of sites}

\item{m}{number of species}

\item{form}{formula with \code{y} (comm dat), \code{x} (env),
\code{z} (trait), \code{species}, or \code{sites}}

\item{power}{power for \code{Grafen} method}

\item{covarSim}{covariance parameters}

\item{fixefSim}{fixed effect parameters}
}
\description{
Generate phylogenetic test data
}

