% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sort.repSparse}
\alias{sort.repSparse}
\alias{standardSort}
\title{Sort the indices of a repeated sparse matrix}
\usage{
\method{sort}{repSparse}(x, decreasing = FALSE, type = c("col", "row",
  "val"), ...)

standardSort(x)
}
\arguments{
\item{decreasing}{see \code{\link{sort}}}

\item{type}{sort by column, row, or value indices?}
}
\description{
Sort the indices of a repeated sparse matrix
}

