% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{stanCov}
\alias{stanCov}
\title{Standardize covariance matrix to determinant one}
\usage{
stanCov(covMat)
}
\arguments{
\item{covMat}{covariance matrix}
}
\description{
Standardize covariance matrix to determinant one
}

