% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strucGlmer-class}
\alias{formula.strucGlmer}
\alias{print.strucGlmer}
\alias{strucGlmer-class}
\alias{summary.strucGlmer}
\title{Structured GLMM class}
\usage{
print.strucGlmer(x, ...)

summary.strucGlmer(object, compSE = TRUE, ...)

\method{formula}{strucGlmer}(x, ...)
}
\arguments{
\item{x,object}{\code{strucGlmer} objects}

\item{...}{additional arguments to methods}
}
\description{
An \code{S3} class for generalized linear mixed models with
structured (co)variance terms, which are lists with the following
elements:
\describe{
  \item{opt}{Result of the optimizer (currently \code{\link{minqa}}).}
  \item{parsedForm}{Results of \code{\link{strucParseFormula}}.}
  \item{dfun}{A function for computing the model deviance.
              The environment of \code{dfun} contains objects for
              representing the model.}}
}

