% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strucGlmer}
\alias{covar.strucGlmer}
\alias{covarPerTerm}
\alias{fixef.strucGlmer}
\alias{getStrucGlmerPar}
\alias{loads.strucGlmer}
\alias{loadsPerTerm}
\alias{pars.strucGlmer}
\alias{print.strucGlmer}
\alias{strucGlmer}
\title{Structured glmer}
\usage{
strucGlmer(formula, data, family, addArgs = list(), optVerb = 0L,
  weights = NULL, offset = NULL, etastart = NULL, ...)

print.strucGlmer(x, ...)

getStrucGlmerPar(object, type, ...)

\method{fixef}{strucGlmer}(object, ...)

\method{covar}{strucGlmer}(object, ...)

loads.strucGlmer(object, ...)

\method{pars}{strucGlmer}(object, ...)

covarPerTerm(object)

loadsPerTerm(object)
}
\arguments{
\item{formula}{extended mixed model formula}

\item{data}{data frame}

\item{addArgs}{list of additional arguments to pass to}

\item{optVerb}{verbose
\code{\link{setReTrm}} methods}

\item{...}{further arguments to \code{\link{mkGeneralGlmerDevfun}}}

\item{type}{character string giving the type of parameter
(e.g. \code{"fixef", "covar"})}
}
\description{
Parses a mixed model formula with special structures (with
\code{\link{strucParseFormula}}), constructs a generalized linear
mixed model deviance function (with
\code{\link{mkGeneralGlmerDevfun}}), optimizes this deviance
function (with \code{\link{bobyqa}}), and returns the results as a
\code{strucGlmer} object.
}

