% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{strucMatrix-class}
\alias{diag,strucMatrix-method}
\alias{diag<-,strucMatrix-method}
\alias{dim.strucMatrix}
\alias{image.strucMatrix}
\alias{plot.strucMatrix}
\alias{print.strucMatrix}
\alias{strucMatrix-class}
\alias{t.strucMatrix}
\alias{update.strucMatrix}
\title{A repeated sparse matrix class}
\usage{
\method{print}{strucMatrix}(x, n = 6L, ...)

\method{update}{strucMatrix}(object, newPars, ...)

\method{image}{strucMatrix}(x, plain = FALSE, ...)

\method{plot}{strucMatrix}(x, y, plain = FALSE, ...)

\method{t}{strucMatrix}(x)

\S4method{diag}{strucMatrix}(x)

\S4method{diag}{strucMatrix}(x) <- value

\method{dim}{strucMatrix}(x)
}
\arguments{
\item{x}{\code{strucMatrix} object}

\item{n}{how many indices and repeated values to print?}

\item{...}{passed to subsequent functions}

\item{object}{\code{strucMatrix} object}

\item{newPars}{new parameter values}

\item{plain}{should a completely plain plot be used? (try and see)}

\item{y}{not used}

\item{value}{replacement value for diagonal}
}
\description{
An \code{S3} class for repeated sparse matrices, which are lists
with a \code{Dim} attribute and with the following elements:
\describe{
  \item{\code{rowInds}}{0-based row indices}
  \item{\code{colInds}}{0-based column indices}
  \item{\code{valInds}}{1-based indices for associating each \code{rowInds} and
                        \code{colInds} pair with the elements in \code{vals}}
  \item{\code{trans}}{A function for transforming a parameter vector
                      into \code{vals}.  This function is used by
                      \code{\link{update.strucMatrix}}}
}
Objects in this class can be constructed with
\code{\link{strucMatrix}} (and, for example,
\code{\link{strucMatrixDiag}}), and has several methods including
the following.
}
\details{
The \code{...} arguments for the \code{update.strucMatrix} method
can be used to specify special parameter arguments, if
\code{object$mkNewPars} exits (which is often the case for special
matrices such as \code{\link{strucMatrixTri}}.  This is a more
explicit approach and can therefore be more convenient than having
to figure out what order the parameters should appear in
\code{newPars}.  For example, \code{update(., diagVals = c(1, 1),
offDiagVals = -0.2)} is more explicit than \code{update(., c(1,
-0.2, 1))}.
}
\examples{
(X <- strucMatrix(rowInds = 1:6,
                colInds = rep(1:2, 3),
                valInds = rep(1:2, each = 3),
                vals    = c(-pi, pi)))
(Y <- update(X, c(0.2, 0.8)))
as.matrix(X, sparse = TRUE)
as.matrix(Y, sparse = TRUE)
image(kron(t(X), Y))
image(bind(X, Y, type = "diag"))
}

