% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structuredSparseMatrices.R
\name{strucMatrixCol}
\alias{strucMatrixCol}
\title{Column vector as structured sparse matrix}
\usage{
strucMatrixCol(vals, valInds)
}
\arguments{
\item{vals}{vector of values}

\item{valInds}{vector of value indices}
}
\description{
Column vector as structured sparse matrix
}
\examples{
set.seed(1)
(xCol <- strucMatrixCol(rnorm(5)))
}
\seealso{
Other strucMatrixSpecial: \code{\link{strucMatrixBlank}},
  \code{\link{strucMatrixCompSymm}},
  \code{\link{strucMatrixConstVarChol}},
  \code{\link{strucMatrixCorFactor}},
  \code{\link{strucMatrixCorMatChol}},
  \code{\link{strucMatrixDiag}},
  \code{\link{strucMatrixExpChol}},
  \code{\link{strucMatrixFull}},
  \code{\link{strucMatrixGenFullTri}},
  \code{\link{strucMatrixIdent}},
  \code{\link{strucMatrixInd}},
  \code{\link{strucMatrixOneOffDiag}},
  \code{\link{strucMatrixOnes}},
  \code{\link{strucMatrixSymm}},
  \code{\link{strucMatrixTri}},
  \code{\link{strucMatrixVarWithCovariate}}
}

