% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strucMatrixCompSymm}
\alias{strucMatrixCompSymm}
\title{Repeated sparse matrix with compound symmetry}
\usage{
strucMatrixCompSymm(diagVal, offDiagVal, matSize)
}
\arguments{
\item{diagVal}{value for the diagonal}

\item{offDiagVal}{value for the off-diagonal}

\item{matSize}{size of the resulting matrix}
}
\description{
Repeated sparse matrix with compound symmetry
}
\examples{
(xCompSymm <- strucMatrixCompSymm(1, -0.2, 5))
}
\seealso{
Other strucMatrixSpecial: \code{\link{strucMatrixBlank}};
  \code{\link{strucMatrixCol}};
  \code{\link{strucMatrixConstVarChol}};
  \code{\link{strucMatrixCorFactor}};
  \code{\link{strucMatrixCorMatChol}};
  \code{\link{strucMatrixDiag}};
  \code{\link{strucMatrixExpChol}};
  \code{\link{strucMatrixFull}};
  \code{\link{strucMatrixGenFullTri}};
  \code{\link{strucMatrixIdent}};
  \code{\link{strucMatrixInd}};
  \code{\link{strucMatrixOneOffDiag}};
  \code{\link{strucMatrixOnes}};
  \code{\link{strucMatrixSymm}};
  \code{\link{strucMatrixTri}};
  \code{\link{strucMatrixVarWithCovariate}}
}

