% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strucMatrixCorFactor}
\alias{strucMatrixCorFactor}
\title{Construct a repeated sparse upper Cholesky factor from an
\code{nlme}-style \code{corStruct} object}
\usage{
strucMatrixCorFactor(object, sig = 1)
}
\arguments{
\item{object}{a \code{corStruct} object}

\item{sig}{initial standard deviation}
}
\description{
Construct a repeated sparse upper Cholesky factor from an
\code{nlme}-style \code{corStruct} object
}
\examples{
if(require("nlme")) {
    corObj <- Initialize(corAR1(0.5, form = ~ 1 | Subject),
                         data = Orthodont)
}
(xCorFactor <- strucMatrixCorFactor(corObj))
}
\seealso{
Other strucMatrixSpecial: \code{\link{strucMatrixBlank}};
  \code{\link{strucMatrixCol}};
  \code{\link{strucMatrixCompSymm}};
  \code{\link{strucMatrixConstVarChol}};
  \code{\link{strucMatrixCorMatChol}};
  \code{\link{strucMatrixDiag}};
  \code{\link{strucMatrixExpChol}};
  \code{\link{strucMatrixFull}};
  \code{\link{strucMatrixGenFullTri}};
  \code{\link{strucMatrixIdent}};
  \code{\link{strucMatrixInd}};
  \code{\link{strucMatrixOneOffDiag}};
  \code{\link{strucMatrixOnes}};
  \code{\link{strucMatrixSymm}};
  \code{\link{strucMatrixTri}};
  \code{\link{strucMatrixVarWithCovariate}}
}

