% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strucMatrixCorMatChol}
\alias{strucMatrixCorMatChol}
\title{Repeated sparse Cholesky factor of a correlation matrix}
\usage{
strucMatrixCorMatChol(offDiagPars)
}
\arguments{
\item{offDiagPars}{parameters determining the off-diagonal of the
Cholesky factor}
}
\description{
Repeated sparse Cholesky factor of a correlation matrix
}
\examples{
set.seed(1)
(xCorMatChol <- strucMatrixCorMatChol(rnorm(choose(5, 2))))
}
\seealso{
Other strucMatrixSpecial: \code{\link{strucMatrixBlank}};
  \code{\link{strucMatrixCol}};
  \code{\link{strucMatrixCompSymm}};
  \code{\link{strucMatrixConstVarChol}};
  \code{\link{strucMatrixCorFactor}};
  \code{\link{strucMatrixDiag}};
  \code{\link{strucMatrixExpChol}};
  \code{\link{strucMatrixFull}};
  \code{\link{strucMatrixGenFullTri}};
  \code{\link{strucMatrixIdent}};
  \code{\link{strucMatrixInd}};
  \code{\link{strucMatrixOneOffDiag}};
  \code{\link{strucMatrixOnes}};
  \code{\link{strucMatrixSymm}};
  \code{\link{strucMatrixTri}};
  \code{\link{strucMatrixVarWithCovariate}}
}

