% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strucMatrixDiag}
\alias{strucMatrixDiag}
\title{Diagonal structured sparse matrix}
\usage{
strucMatrixDiag(vals, valInds)
}
\arguments{
\item{vals}{vector of values}

\item{valInds}{vector of value indices}
}
\description{
Diagonal structured sparse matrix
}
\examples{
set.seed(1)
(xDiag <- strucMatrixDiag(rnorm(5)))
}
\seealso{
Other strucMatrixSpecial: \code{\link{strucMatrixBlank}};
  \code{\link{strucMatrixCol}};
  \code{\link{strucMatrixCompSymm}};
  \code{\link{strucMatrixConstVarChol}};
  \code{\link{strucMatrixCorFactor}};
  \code{\link{strucMatrixCorMatChol}};
  \code{\link{strucMatrixExpChol}};
  \code{\link{strucMatrixFull}};
  \code{\link{strucMatrixGenFullTri}};
  \code{\link{strucMatrixIdent}};
  \code{\link{strucMatrixInd}};
  \code{\link{strucMatrixOneOffDiag}};
  \code{\link{strucMatrixOnes}};
  \code{\link{strucMatrixSymm}};
  \code{\link{strucMatrixTri}};
  \code{\link{strucMatrixVarWithCovariate}}
}

