% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strucMatrixExpChol}
\alias{strucMatrixExpChol}
\title{Cholesky factor of a repeated sparse covariance matrix obeying
exponential distance decay in covariance}
\usage{
strucMatrixExpChol(distObj, cutOffDist = Inf)
}
\arguments{
\item{distObj}{distance matrix object}

\item{cutOffDist}{maximum distance with nonzero correlations}
}
\description{
Cholesky factor of a repeated sparse covariance matrix obeying
exponential distance decay in covariance
}
\examples{
(xExpChol <- strucMatrixExpChol(dist(matrix(rnorm(10), 5, 2))))
}
\seealso{
Other strucMatrixSpecial: \code{\link{strucMatrixBlank}};
  \code{\link{strucMatrixCol}};
  \code{\link{strucMatrixCompSymm}};
  \code{\link{strucMatrixConstVarChol}};
  \code{\link{strucMatrixCorFactor}};
  \code{\link{strucMatrixCorMatChol}};
  \code{\link{strucMatrixDiag}};
  \code{\link{strucMatrixFull}};
  \code{\link{strucMatrixGenFullTri}};
  \code{\link{strucMatrixIdent}};
  \code{\link{strucMatrixInd}};
  \code{\link{strucMatrixOneOffDiag}};
  \code{\link{strucMatrixOnes}};
  \code{\link{strucMatrixSymm}};
  \code{\link{strucMatrixTri}};
  \code{\link{strucMatrixVarWithCovariate}}
}

