% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strucMatrixFull}
\alias{strucMatrixFull}
\title{Full structured sparse matrix}
\usage{
strucMatrixFull(nrow, ncol, vals)
}
\arguments{
\item{nrow,ncol}{numbers of rows and columns}

\item{vals}{vector of values}
}
\description{
Full structured sparse matrix
}
\examples{
set.seed(1)
(xFull <- strucMatrixFull(5, 5, rnorm(25)))
}
\seealso{
Other strucMatrixSpecial: \code{\link{strucMatrixBlank}};
  \code{\link{strucMatrixCol}};
  \code{\link{strucMatrixCompSymm}};
  \code{\link{strucMatrixConstVarChol}};
  \code{\link{strucMatrixCorFactor}};
  \code{\link{strucMatrixCorMatChol}};
  \code{\link{strucMatrixDiag}};
  \code{\link{strucMatrixExpChol}};
  \code{\link{strucMatrixGenFullTri}};
  \code{\link{strucMatrixIdent}};
  \code{\link{strucMatrixInd}};
  \code{\link{strucMatrixOneOffDiag}};
  \code{\link{strucMatrixOnes}};
  \code{\link{strucMatrixSymm}};
  \code{\link{strucMatrixTri}};
  \code{\link{strucMatrixVarWithCovariate}}
}

