% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strucMatrixGenFullTri}
\alias{strucMatrixGenFullTri}
\title{General and full triangular repeated sparse matrix}
\usage{
strucMatrixGenFullTri(nrow, ncol, vals, diag = TRUE, low = TRUE)
}
\arguments{
\item{nrow,ncol}{number of rows and columns}

\item{vals}{values for the nonzero values}

\item{diag}{include diagonal?}

\item{low}{lower triangular?}
}
\description{
General and full triangular repeated sparse matrix
}
\note{
Beware there appear to be bugs here.
}
\examples{
set.seed(1)
(xGenTri <- strucMatrixGenFullTri(5, 5, rnorm(choose(6, 2))))
}
\seealso{
Other strucMatrixSpecial: \code{\link{strucMatrixBlank}};
  \code{\link{strucMatrixCol}};
  \code{\link{strucMatrixCompSymm}};
  \code{\link{strucMatrixConstVarChol}};
  \code{\link{strucMatrixCorFactor}};
  \code{\link{strucMatrixCorMatChol}};
  \code{\link{strucMatrixDiag}};
  \code{\link{strucMatrixExpChol}};
  \code{\link{strucMatrixFull}};
  \code{\link{strucMatrixIdent}};
  \code{\link{strucMatrixInd}};
  \code{\link{strucMatrixOneOffDiag}};
  \code{\link{strucMatrixOnes}};
  \code{\link{strucMatrixSymm}};
  \code{\link{strucMatrixTri}};
  \code{\link{strucMatrixVarWithCovariate}}
}

