% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strucMatrixInd}
\alias{strucMatrixInd}
\title{Structured sparse indicator matrix}
\usage{
strucMatrixInd(fac)
}
\arguments{
\item{fac}{vector coercible to factor}
}
\description{
Structured sparse indicator matrix
}
\examples{
(xInd <- strucMatrixInd(rep(1:3, c(2, 2, 1))))
}
\seealso{
Other strucMatrixSpecial: \code{\link{strucMatrixBlank}};
  \code{\link{strucMatrixCol}};
  \code{\link{strucMatrixCompSymm}};
  \code{\link{strucMatrixConstVarChol}};
  \code{\link{strucMatrixCorFactor}};
  \code{\link{strucMatrixCorMatChol}};
  \code{\link{strucMatrixDiag}};
  \code{\link{strucMatrixExpChol}};
  \code{\link{strucMatrixFull}};
  \code{\link{strucMatrixGenFullTri}};
  \code{\link{strucMatrixIdent}};
  \code{\link{strucMatrixOneOffDiag}};
  \code{\link{strucMatrixOnes}};
  \code{\link{strucMatrixSymm}};
  \code{\link{strucMatrixTri}};
  \code{\link{strucMatrixVarWithCovariate}}
}

