% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strucMatrixOneOffDiag}
\alias{strucMatrixOneOffDiag}
\title{Structured sparse matrix with only one non-zero value off the
diagonal}
\usage{
strucMatrixOneOffDiag(diagVal, offDiagVal, offDiagInds, matSize)
}
\arguments{
\item{diagVal}{unique value for the diagonal}

\item{offDiagVal}{unique value for the off-diagonal}

\item{offDiagInds}{indices for the two correlated objects}

\item{matSize}{size of the matrix}
}
\description{
Structured sparse matrix with only one non-zero value off the
diagonal
}
\examples{
(xOneOffDiag <- strucMatrixOneOffDiag(1, -0.2, c(5, 2), 5))
}
\seealso{
Other strucMatrixSpecial: \code{\link{strucMatrixBlank}};
  \code{\link{strucMatrixCol}};
  \code{\link{strucMatrixCompSymm}};
  \code{\link{strucMatrixConstVarChol}};
  \code{\link{strucMatrixCorFactor}};
  \code{\link{strucMatrixCorMatChol}};
  \code{\link{strucMatrixDiag}};
  \code{\link{strucMatrixExpChol}};
  \code{\link{strucMatrixFull}};
  \code{\link{strucMatrixGenFullTri}};
  \code{\link{strucMatrixIdent}};
  \code{\link{strucMatrixInd}};
  \code{\link{strucMatrixOnes}};
  \code{\link{strucMatrixSymm}};
  \code{\link{strucMatrixTri}};
  \code{\link{strucMatrixVarWithCovariate}}
}

