% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strucMatrixOnes}
\alias{strucMatrixOnes}
\title{Repeated sparse matrix of ones}
\usage{
strucMatrixOnes(nrow, ncol)
}
\arguments{
\item{nrow,ncol}{numbers of rows and columns}
}
\description{
Repeated sparse matrix of ones
}
\examples{
(xOnes <- strucMatrixOnes(5, 5))
}
\seealso{
Other strucMatrixSpecial: \code{\link{strucMatrixBlank}};
  \code{\link{strucMatrixCol}};
  \code{\link{strucMatrixCompSymm}};
  \code{\link{strucMatrixConstVarChol}};
  \code{\link{strucMatrixCorFactor}};
  \code{\link{strucMatrixCorMatChol}};
  \code{\link{strucMatrixDiag}};
  \code{\link{strucMatrixExpChol}};
  \code{\link{strucMatrixFull}};
  \code{\link{strucMatrixGenFullTri}};
  \code{\link{strucMatrixIdent}};
  \code{\link{strucMatrixInd}};
  \code{\link{strucMatrixOneOffDiag}};
  \code{\link{strucMatrixSymm}};
  \code{\link{strucMatrixTri}};
  \code{\link{strucMatrixVarWithCovariate}}
}

