% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strucMatrixTri}
\alias{strucMatrixTri}
\title{Triangular repeated sparse matrix}
\usage{
strucMatrixTri(diagVals, offDiagVals, low = TRUE)
}
\arguments{
\item{diagVals}{values for the diagonal}

\item{offDiagVals}{values for the off-diagonal}

\item{low}{lower triangular?}
}
\description{
Triangular repeated sparse matrix
}
\examples{
set.seed(1)
(xTri <- strucMatrixTri(rnorm(5), rnorm(choose(5, 2))))
}
\seealso{
Other strucMatrixSpecial: \code{\link{strucMatrixBlank}};
  \code{\link{strucMatrixCol}};
  \code{\link{strucMatrixCompSymm}};
  \code{\link{strucMatrixConstVarChol}};
  \code{\link{strucMatrixCorFactor}};
  \code{\link{strucMatrixCorMatChol}};
  \code{\link{strucMatrixDiag}};
  \code{\link{strucMatrixExpChol}};
  \code{\link{strucMatrixFull}};
  \code{\link{strucMatrixGenFullTri}};
  \code{\link{strucMatrixIdent}};
  \code{\link{strucMatrixInd}};
  \code{\link{strucMatrixOneOffDiag}};
  \code{\link{strucMatrixOnes}};
  \code{\link{strucMatrixSymm}};
  \code{\link{strucMatrixVarWithCovariate}}
}

