% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strucMatrixVarWithCovariate}
\alias{strucMatrixVarWithCovariate}
\title{Repeated sparse diagonal covariance matrix with a covariate
determining the diagonal}
\usage{
strucMatrixVarWithCovariate(varPars, covariate, grpFac,
  mkTransFunc = mkVarExpTrans)
}
\arguments{
\item{varPars}{vector of variance parameters (one per level of
\code{grpFac})}

\item{covariate}{covariate}

\item{grpFac}{a grouping factor (or anything coercible to
\code{numeric} really) with the number of levels equal to the
length of \code{varPars}}

\item{mkTransFunc}{a function taking arguments \code{varPars},
\code{covariate}, and \code{grpFac} for constructing a
\code{trans} function (see \code{\link{mkVarExpTrans}} for an
example).}
}
\description{
Repeated sparse diagonal covariance matrix with a covariate
determining the diagonal
}
\examples{
(xVarWithCovariate <- strucMatrixVarWithCovariate(rep(1, 5), 0.05*(4:0)))
}
\seealso{
Other strucMatrixSpecial: \code{\link{strucMatrixBlank}};
  \code{\link{strucMatrixCol}};
  \code{\link{strucMatrixCompSymm}};
  \code{\link{strucMatrixConstVarChol}};
  \code{\link{strucMatrixCorFactor}};
  \code{\link{strucMatrixCorMatChol}};
  \code{\link{strucMatrixDiag}};
  \code{\link{strucMatrixExpChol}};
  \code{\link{strucMatrixFull}};
  \code{\link{strucMatrixGenFullTri}};
  \code{\link{strucMatrixIdent}};
  \code{\link{strucMatrixInd}};
  \code{\link{strucMatrixOneOffDiag}};
  \code{\link{strucMatrixOnes}};
  \code{\link{strucMatrixSymm}};
  \code{\link{strucMatrixTri}}
}

