% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devianceFunctionConstruction.R
\name{strucMkDevfun}
\alias{strucMkDevfun}
\title{Make a (Laplace) glmer deviance function from a parsed formula}
\usage{
strucMkDevfun(parsedForm, family, weights = NULL, offset = NULL,
  etastart = NULL, ...)
}
\arguments{
\item{parsedForm}{results of \code{\link{strucParseFormula}}}

\item{family, weights, offset, etastart}{see \code{\link{mkGeneralGlmerDevfun}}}

\item{...}{not used}
}
\description{
Make a (Laplace) glmer deviance function from a parsed formula
}

