% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strucOpt}
\alias{strucOpt}
\title{Optimize a structured GLMM deviance function}
\usage{
strucOpt(optimizer = minqa::bobyqa, initPars, dfun, lower, upper, optVerb,
  optMaxit, ...)
}
\arguments{
\item{optimizer}{nonlinear optimization function (currently only
minqa::bobyqa is allowed)}

\item{initPars}{initial parameter values}

\item{dfun}{structured GLMM deviance function (results of
\code{\link{strucMkDevfun}} or \code{\link{mkGeneralGlmerDevfun}})}

\item{lower,upper}{bounds on the parameter vector}

\item{optVerb,optMaxit}{controls for the optimizer (see
\code{\link{strucGlmer}})}

\item{...}{not used}
}
\description{
Optimize a structured GLMM deviance function
}

