% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strucParseFormula}
\alias{model.matrix.strucParseFormula}
\alias{print.strucParseFormula}
\alias{simulate.strucParseFormula}
\alias{strucParseFormula}
\title{Parse a mixed model formula}
\usage{
strucParseFormula(formula, data, addArgs = list(), reTrmsList = NULL, ...)

\method{print}{strucParseFormula}(x, ...)

\method{model.matrix}{strucParseFormula}(object, ...)

\method{simulate}{strucParseFormula}(object, nsim = 1, seed = NULL,
  weights = NULL, family = binomial(), ...)
}
\arguments{
\item{formula}{mixed model formula}

\item{data}{an object coercible to data frame}

\item{addArgs}{list of additional arguments to
\code{\link{setReTrm}} methods}

\item{reTrmsList}{if \code{NULL} \code{\link{mkReTrmStructs}} is used}

\item{...}{additional parameters to \code{\link{as.data.frame}}}

\item{x}{\code{strucParseFormula} objects}

\item{object}{\code{strucParseFormula} objects}

\item{nsim}{number of simulations}

\item{seed}{random seed (see \code{\link{set.seed}})}

\item{weights}{\code{\link{weights}} for each observation}

\item{family}{\code{\link{family}} object}
}
\value{
A list with components:
\item{response}{The response vector}
\item{fixed}{The fixed effects model matrix}
\item{random}{List of random effects terms returned by
\code{\link{mkReTrmStructs}} and updated with
\code{\link{setReTrm}}.}
\item{Zt}{The sparse transposed random effects model matrix, of
class \code{dgCMatrix}.}
\item{Lambdat}{The sparse transposed relative covariance factor,
of class \code{dgCMatrix}}
\item{mapToModMat}{function taking the \code{loads} parameters and
returning the values of the non-zero elements of \code{Zt}
(i.e. the \code{x} slot of \code{Zt})}
\item{mapToCovFact}{function taking the \code{covar} parameters
and returning the values of the non-zero elements of
\code{Lambdat} (i.e. the \code{x} slot of \code{Lambdat})}
\item{initPars}{The initial parameter vector}
\item{parInds}{List of indices to the different types of
parameters.}
\item{lower, upper}{Vectors of lower and upper bounds on
\code{initPars}.}
\item{devfunEnv}{Environment of the deviance function.}
\item{formula}{Model formula.}
}
\description{
Parse a mixed model formula
}

