% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structuredSparseMatrices.R
\name{subset.strucMatrix}
\alias{strucMatrixRowSubset}
\alias{subset.strucMatrix}
\title{Subsetting structured sparse matices}
\usage{
\method{subset}{strucMatrix}(x, rowInds = NULL, colInds = NULL, ...)

strucMatrixRowSubset(x, rowInds)
}
\arguments{
\item{x}{a \code{\link{strucMatrix-class}} object}

\item{rowInds, colInds}{1-based integer indices for rows and
columns}

\item{...}{unused}
}
\description{
Subsetting structured sparse matices
}
\examples{
set.seed(1)
X <- strucMatrix(c(1, 2, 1, 2), c(1, 1, 2, 2), 1:4, rnorm(4))
(subsetExample <- subset(X, c(1, 2, 2, 2, 1, 1)))
}
\seealso{
Other matrixCombining: \code{\link{bind}},
  \code{\link{kron}}
}

