% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeModel.R
\name{tipNodeRelationship}
\alias{tipNodeRelationship}
\title{Tip-node bigraph adjacency matrix}
\usage{
tipNodeRelationship(phy)
}
\arguments{
\item{phy}{phylo object}
}
\value{
adjacency matrix
}
\description{
Tip-node bigraph adjacency matrix
}

