% Generated by roxygen2 (4.0.1): do not edit by hand
\name{triInds}
\alias{triInds}
\title{Convert row and column indices to vector indices, for a triangular
matrix}
\usage{
triInds(rowInds, colInds, maxInd, type = c("distClass"))
}
\arguments{
\item{rowInds,colInds}{vectors of 1-based row and column indices}

\item{maxInd}{maximum index (could be larger than
\code{max(rowInds, colInds)})}

\item{type}{type of parameterization (currently only
\code{"distClass"} available)}
}
\description{
Convert row and column indices to vector indices, for a triangular
matrix
}
\examples{
n <- 5
set.seed(1)
X <- dist(matrix(rnorm((n + 1) * 2), n + 1, 2))
rowInds <- rep(1:n, 1:n)
colInds <- sequence(1:n)
X[triInds(rowInds, colInds, n)]
X
}

