% Generated by roxygen2 (4.0.1): do not edit by hand
\name{update.reTrmStruct}
\alias{update.reTrmStruct}
\title{Update a random effects term structure with new parameters}
\usage{
update.reTrmStruct(object, newCovar, newLoads)
}
\arguments{
\item{object}{a \code{reTrmStruct} object}

\item{newCovar}{new covariance parameters}

\item{newLoads}{new loadings parameters}
}
\description{
Update a random effects term structure with new parameters
}

