% Generated by roxygen2 (4.0.1): do not edit by hand
\name{vcov.gblmerMod}
\alias{vcov.gblmerMod}
\title{Variance-covariance matrix of the (co)variance parameters and
loadings (and sometimes of fixed effects)}
\usage{
\method{vcov}{gblmerMod}(object, correlation = TRUE, ...)
}
\arguments{
\item{object}{a \code{\link{gblmerMod}} object}

\item{correlation}{should the correlation matrix be computed too?}

\item{...}{not used}
}
\description{
Variance-covariance matrix of the (co)variance parameters and
loadings (and sometimes of fixed effects)
}

