% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.repSparse}
\alias{as}
\alias{as.data.frame.repSparse}
\alias{as.data.table.repSparse}
\alias{as.matrix.repSparse}
\alias{as.repSparse}
\alias{as.repSparse.dsparseMatrix}
\alias{as.repSparse.matrix}
\alias{as.repSparse.repSparse}
\title{Coerce to and from repeated sparse matrices}
\usage{
as.repSparse(x, ...)

\method{as.repSparse}{repSparse}(x, ...)

\method{as.repSparse}{dsparseMatrix}(x, ...)

\method{as.repSparse}{matrix}(x, ...)

\method{as.data.frame}{repSparse}(x, ...)

as.data.table.repSparse(x, keep.rownames = FALSE)

as.matrix.repSparse(x, sparse = FALSE, ...)
}
\arguments{
\item{x}{an object}

\item{...}{dots}

\item{keep.rownames}{passed to \code{as.data.table}}

\item{sparse}{return \code{sparseMatrix}?}
}
\description{
Coerce to and from repeated sparse matrices

as("repSparse", "sparseMatrix")

as("repSparse", "CsparseMatrix")

as("repSparse", "dgCMatrix")

as("repSparse", "TsparseMatrix")

as("repSparse", "dgTMatrix")
}
\examples{
set.seed(1)
m1 <- as.repSparse(matrix(rnorm(6), 2, 3))
m2 <- as.repSparse(matrix(rbinom(6, 1, 0.5), 2, 3))
m3 <- sparseMatrix(i = 1:10, j = rep(1:5, 2),
                   x = as.integer(rbinom(10, 1, 0.5)),
                   giveCsparse = FALSE)
as.repSparse(m1)
as.repSparse(m2)
as.repSparse(m3)
as(m1, "TsparseMatrix")
as(m2, "dgCMatrix")
as(m2, "sparseMatrix")
}

