% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.repSparse}
\alias{as.data.frame.repSparse}
\alias{as.matrix.repSparse}
\alias{as.matrix.repSparseChol}
\alias{as.repSparse}
\alias{as.repSparse.dCHMsimpl}
\alias{as.repSparse.dist}
\alias{as.repSparse.dsparseMatrix}
\alias{as.repSparse.factor}
\alias{as.repSparse.matrix}
\alias{as.repSparse.repSparse}
\alias{repSparse2dgCMatrix}
\alias{repSparse2dgTMatrix}
\alias{repSparse2dtCMatrix}
\alias{repSparse2dtTMatrix}
\alias{repSparse2gCsparseMatrix}
\alias{repSparse2gTsparseMatrix}
\alias{repSparse2sparseMatrix}
\title{Coerce to and from repeated sparse matrices}
\usage{
as.repSparse(x, ...)

\method{as.repSparse}{repSparse}(x, ...)

\method{as.repSparse}{dsparseMatrix}(x, ...)

\method{as.repSparse}{matrix}(x, ...)

\method{as.repSparse}{factor}(x, ...)

\method{as.repSparse}{dist}(x, ...)

\method{as.repSparse}{dCHMsimpl}(x, ...)

\method{as.data.frame}{repSparse}(x, ...)

\method{as.matrix}{repSparse}(x, sparse = FALSE, ...)

\method{as.matrix}{repSparseChol}(x, sparse = FALSE, ...)
}
\arguments{
\item{x}{an object}

\item{...}{dots}

\item{sparse}{return \code{sparseMatrix}?}
}
\description{
Coerce to and from repeated sparse matrices

as("repSparse", "sparseMatrix")

as("repSparse", "CsparseMatrix")

as("repSparse", "dgCMatrix")

as("repSparse", "dtCMatrix")

as("repSparse", "TsparseMatrix")

as("repSparse", "dgTMatrix")

as("repSparse", "dtTMatrix")
}
\examples{
set.seed(1)
m1 <- as.repSparse(matrix(rnorm(6), 2, 3))
m2 <- as.repSparse(matrix(rbinom(6, 1, 0.5), 2, 3))
m3 <- sparseMatrix(i = 1:10, j = rep(1:5, 2),
                   x = as.integer(rbinom(10, 1, 0.5)),
                   giveCsparse = FALSE)
as.repSparse(m1)
as.repSparse(m2)
as.repSparse(m3)
as(m1, "TsparseMatrix")
as(m2, "dgCMatrix")
as(m2, "sparseMatrix")
}

