% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.repSparse}
\alias{as.data.frame.repSparse}
\alias{as.data.table.repSparse}
\alias{as.matrix.repSparse}
\alias{as.repSparse}
\alias{as.repSparse.dsparseMatrix}
\alias{as.repSparse.factor}
\alias{as.repSparse.matrix}
\alias{as.repSparse.repSparse}
\alias{repSparse2CsparseMatrix}
\alias{repSparse2TsparseMatrix}
\alias{repSparse2dgCMatrix}
\alias{repSparse2dgTMatrix}
\alias{repSparse2sparseMatrix}
\title{Coerce to and from repeated sparse matrices}
\usage{
as.repSparse(x, ...)

\method{as.repSparse}{repSparse}(x, ...)

\method{as.repSparse}{dsparseMatrix}(x, ...)

\method{as.repSparse}{matrix}(x, ...)

\method{as.repSparse}{factor}(x, ...)

\method{as.data.frame}{repSparse}(x, ...)

as.data.table.repSparse(x, keep.rownames = FALSE)

\method{as.matrix}{repSparse}(x, sparse = FALSE, ...)
}
\arguments{
\item{x}{an object}

\item{...}{dots}

\item{keep.rownames}{passed to \code{as.data.table}}

\item{sparse}{return \code{sparseMatrix}?}
}
\description{
Coerce to and from repeated sparse matrices

as("repSparse", "sparseMatrix")

as("repSparse", "CsparseMatrix")

as("repSparse", "dgCMatrix")

as("repSparse", "TsparseMatrix")

as("repSparse", "dgTMatrix")
}
\examples{
set.seed(1)
m1 <- as.repSparse(matrix(rnorm(6), 2, 3))
m2 <- as.repSparse(matrix(rbinom(6, 1, 0.5), 2, 3))
m3 <- sparseMatrix(i = 1:10, j = rep(1:5, 2),
                   x = as.integer(rbinom(10, 1, 0.5)),
                   giveCsparse = FALSE)
as.repSparse(m1)
as.repSparse(m2)
as.repSparse(m3)
as(m1, "TsparseMatrix")
as(m2, "dgCMatrix")
as(m2, "sparseMatrix")
}
\seealso{
Other repeated.sparse.matrix.topics: \code{\link{.bind}},
  \code{\link{bind}}, \code{\link{rep.repSparse}};
  \code{\link{chol.repSparse}},
  \code{\link{chol.repSparseCompSymm}},
  \code{\link{chol.repSparseOneOffDiag}};
  \code{\link{dim.repSparse}},
  \code{\link{image.repSparse}},
  \code{\link{plot.repSparse}},
  \code{\link{print.repSparse}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}}, \code{\link{t.repSparse}},
  \code{\link{update.repSparse}}; \code{\link{getInit}},
  \code{\link{getInit.default}},
  \code{\link{getInit.function}},
  \code{\link{getInit.repSparse}}, \code{\link{setInit}},
  \code{\link{setInit.default}},
  \code{\link{setInit.function}},
  \code{\link{setInit.repSparse}}; \code{\link{ind2point}},
  \code{\link{point2ind}}; \code{\link{kr}},
  \code{\link{kron}}; \code{\link{mkCholOneOffDiagTrans}},
  \code{\link{mkConstVarCholTrans}},
  \code{\link{mkCorMatCholTrans}},
  \code{\link{mkGenCholInPlaceTrans}},
  \code{\link{mkIdentityTrans}}, \code{\link{mkListTrans}},
  \code{\link{mkOuterTrans}}, \code{\link{mkVarExpTrans}};
  \code{\link{rRepSparse}}, \code{\link{repSparseBlank}},
  \code{\link{repSparseCompSymm}},
  \code{\link{repSparseConstVarChol}},
  \code{\link{repSparseCorMatChol}},
  \code{\link{repSparseDiag}},
  \code{\link{repSparseIdent}},
  \code{\link{repSparseOneOffDiag}},
  \code{\link{repSparseSymm}}, \code{\link{repSparseTri}},
  \code{\link{repSparseVarExp}}; \code{\link{repSparse}}
}

