% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bind}
\alias{.bind}
\alias{bind}
\alias{rep.repSparse}
\title{Row, column, and block-diagonal binding for repeated sparse
matrices}
\usage{
bind(..., type = c("row", "col", "diag"))

.bind(mats, type = c("row", "col", "diag"))

\method{rep}{repSparse}(x, times, type = c("row", "col", "diag"),
  repVals = FALSE, ...)
}
\arguments{
\item{...}{list of \code{repSparse} objects (but not used for
\code{rep.repSparse})}

\item{type}{type of binding}

\item{mats}{list of \code{repSparse} matrix objects}

\item{x}{\code{repSparse} object}

\item{times}{like \code{rep}}

\item{repVals}{should the unique values (\code{x$vals}) be
repeated too?  in other words, do you want to have each replicate
to have guaranteed identical values (\code{FALSE}, the default),
or do you want each replicate to be settable to it's own values
(\code{TRUE})?}
}
\description{
Row, column, and block-diagonal binding for repeated sparse
matrices

Repeat repeated sparse matrix
}
\examples{
set.seed(1)
X <- repSparse(c(1, 2, 1, 2), c(1, 1, 2, 2), 1:4, rnorm(4))
Y <- repSparse(c(1, 2, 1), c(1, 1, 2), 1:3, rnorm(3))
Z <- repSparse(c(1, 2), c(1, 2), 1:2, rnorm(2))
(bindDiag <- bind(X, Y, Z, type = "diag"))
(bindRow  <- bind(X, Y, Z, type = "row"))
(bindCol  <- bind(X, Y, Z, type = "col"))
set.seed(1)
X <- repSparse(c(1, 2, 1, 2), c(1, 1, 2, 2), 1:4, rnorm(4))
Y <- repSparse(c(1, 2, 1), c(1, 1, 2), 1:3, rnorm(3))
Z <- repSparse(c(1, 2), c(1, 2), 1:2, rnorm(2))
(repDiag <- rep(X, 3, type = "diag"))
(repRow  <- rep(X, 3, type = "row"))
(repCol  <- rep(X, 3, type = "col"))
}
\seealso{
Other matrixCombining: \code{\link{Ops.repSparse}},
  \code{\link{crossprod,repSparse,missing-method}},
  \code{\link{crossprod,repSparse,repSparse-method}},
  \code{\link{kr}}, \code{\link{kron}},
  \code{\link{kronecker,repSparse,repSparse-method}},
  \code{\link{tcrossprod,repSparse,missing-method}},
  \code{\link{tcrossprod,repSparse,repSparse-method}};
  \code{\link{repSparseRowSubset}},
  \code{\link{subset.repSparse}}

Other matrixCombining: \code{\link{Ops.repSparse}},
  \code{\link{crossprod,repSparse,missing-method}},
  \code{\link{crossprod,repSparse,repSparse-method}},
  \code{\link{kr}}, \code{\link{kron}},
  \code{\link{kronecker,repSparse,repSparse-method}},
  \code{\link{tcrossprod,repSparse,missing-method}},
  \code{\link{tcrossprod,repSparse,repSparse-method}};
  \code{\link{repSparseRowSubset}},
  \code{\link{subset.repSparse}}
}

