% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bind}
\alias{.bind}
\alias{bind}
\alias{rep.repSparse}
\title{Row, column, and block-diagonal binding for repeated sparse
matrices}
\usage{
bind(..., type = c("row", "col", "diag"))

.bind(mats, type = c("row", "col", "diag"))

\method{rep}{repSparse}(x, times, type = c("row", "col", "diag"),
  repVals = FALSE, ...)
}
\arguments{
\item{...}{list of \code{repSparse} objects (but not used for
\code{rep.repSparse})}

\item{type}{type of binding}

\item{mats}{list of \code{repSparse} matrix objects}

\item{x}{\code{repSparse} object}

\item{times}{like \code{rep}}

\item{repVals}{should the unique values (\code{x$vals}) be
repeated too?  in other words, do you want to have each replicate
to have guaranteed identical values (\code{FALSE}, the default),
or do you want each replicate to be settable to it's own values
(\code{TRUE})?}
}
\description{
Row, column, and block-diagonal binding for repeated sparse
matrices

Repeat repeated sparse matrix
}
\examples{
set.seed(1)
X <- repSparse(c(1, 2, 1, 2), c(1, 1, 2, 2), 1:4, rnorm(4))
Y <- repSparse(c(1, 2, 1), c(1, 1, 2), 1:3, rnorm(3))
Z <- repSparse(c(1, 2), c(1, 2), 1:2, rnorm(2))
(bindDiag <- bind(X, Y, Z, type = "diag"))
(bindRow  <- bind(X, Y, Z, type = "row"))
(bindCol  <- bind(X, Y, Z, type = "col"))
set.seed(1)
X <- repSparse(c(1, 2, 1, 2), c(1, 1, 2, 2), 1:4, rnorm(4))
Y <- repSparse(c(1, 2, 1), c(1, 1, 2), 1:3, rnorm(3))
Z <- repSparse(c(1, 2), c(1, 2), 1:2, rnorm(2))
(repDiag <- rep(X, 3, type = "diag"))
(repRow  <- rep(X, 3, type = "row"))
(repCol  <- rep(X, 3, type = "col"))
}
\seealso{
Other repeated.sparse.matrix.topics: \code{\link{as.data.frame.repSparse}},
  \code{\link{as.data.table.repSparse}},
  \code{\link{as.matrix.repSparse}},
  \code{\link{as.repSparse}},
  \code{\link{as.repSparse.dCHMsimpl}},
  \code{\link{as.repSparse.dist}},
  \code{\link{as.repSparse.dsparseMatrix}},
  \code{\link{as.repSparse.factor}},
  \code{\link{as.repSparse.matrix}},
  \code{\link{as.repSparse.repSparse}},
  \code{\link{repSparse2CsparseMatrix}},
  \code{\link{repSparse2TsparseMatrix}},
  \code{\link{repSparse2dgCMatrix}},
  \code{\link{repSparse2dgTMatrix}},
  \code{\link{repSparse2sparseMatrix}};
  \code{\link{chol.repSparse}},
  \code{\link{chol.repSparseCompSymm}},
  \code{\link{chol.repSparseOneOffDiag}};
  \code{\link{diag.repSparse}},
  \code{\link{dim.repSparse}},
  \code{\link{image.repSparse}},
  \code{\link{plot.repSparse}},
  \code{\link{print.repSparse}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}}, \code{\link{t.repSparse}},
  \code{\link{update.repSparse}}; \code{\link{getInit}},
  \code{\link{getInit.default}},
  \code{\link{getInit.function}},
  \code{\link{getInit.reTrmStruct}},
  \code{\link{getInit.repSparse}}, \code{\link{parLength}},
  \code{\link{setInit}}, \code{\link{setInit.default}},
  \code{\link{setInit.function}},
  \code{\link{setInit.reTrmStruct}},
  \code{\link{setInit.repSparse}}; \code{\link{ind2point}},
  \code{\link{point2ind}}; \code{\link{kr}},
  \code{\link{kron}}; \code{\link{mkCholCompSymmTrans}},
  \code{\link{mkCholOneOffDiagTrans}},
  \code{\link{mkConstVarCholTrans}},
  \code{\link{mkCorMatCholTrans}},
  \code{\link{mkExpCholTrans}},
  \code{\link{mkFlexDiagTrans}},
  \code{\link{mkGenCholInPlaceTrans}},
  \code{\link{mkIdentityTrans}}, \code{\link{mkListTrans}},
  \code{\link{mkOuterTrans}}, \code{\link{mkTrans}},
  \code{\link{mkVarExpTrans}},
  \code{\link{mkVarFixedTrans}},
  \code{\link{mkVarIdentTrans}},
  \code{\link{mkVarPowerTrans}};
  \code{\link{repSparseBlank}}, \code{\link{repSparseCol}},
  \code{\link{repSparseCompSymm}},
  \code{\link{repSparseConstVarChol}},
  \code{\link{repSparseCorMatChol}},
  \code{\link{repSparseDiag}},
  \code{\link{repSparseExpChol}},
  \code{\link{repSparseIdent}}, \code{\link{repSparseInd}},
  \code{\link{repSparseOneOffDiag}},
  \code{\link{repSparseOnes}},
  \code{\link{repSparseSpecial}},
  \code{\link{repSparseSymm}}, \code{\link{repSparseTri}},
  \code{\link{repSparseVarWithCovariate}};
  \code{\link{repSparse}}
}

