% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bind}
\alias{.bind}
\alias{bind}
\alias{rep.repSparse}
\title{Row, column, and block-diagonal binding for repeated sparse
matrices}
\usage{
bind(..., type = c("row", "col", "diag"))

.bind(mats, type = c("row", "col", "diag"))

\method{rep}{repSparse}(x, times, type = c("row", "col", "diag"))
}
\arguments{
\item{...}{list of \code{repSparse} objects}

\item{type}{type of binding}

\item{mats}{list of \code{repSparse} matrix objects}

\item{x}{\code{repSparse} object}

\item{times}{like \code{rep}}
}
\description{
Row, column, and block-diagonal binding for repeated sparse
matrices

Repeat repeated sparse matrix
}

