% Generated by roxygen2 (4.0.1): do not edit by hand
\name{chol.repSparse}
\alias{chol.repSparse}
\alias{chol.repSparseOneOffDiag}
\title{Cholesky decomposition of repeated sparse matrices}
\usage{
chol.repSparse(x, ...)

chol.repSparseOneOffDiag(x, ...)
}
\arguments{
\item{x}{an object that inherits from class
\code{\link{repSparse}}}

\item{...}{passed to subsequent functions}
}
\description{
Cholesky decomposition of repeated sparse matrices
}
\note{
These are often just bailout methods, but some special
\code{repSparse} matrices have exploitable structure, which can be
used to keep the number of repeated values down.
}

