% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{chol,strucMatrix-method}
\alias{chol,strucMatrix-method}
\alias{chol,strucMatrixCompSymm-method}
\alias{chol,strucMatrixOneOffDiag-method}
\title{Cholesky decomposition of repeated sparse matrices}
\usage{
\S4method{chol}{strucMatrix}(x, ...)

\S4method{chol}{strucMatrixOneOffDiag}(x, ...)

\S4method{chol}{strucMatrixCompSymm}(x, ...)
}
\arguments{
\item{x}{an object that inherits from class
\code{\link{strucMatrix}}}

\item{...}{passed to subsequent functions}
}
\description{
Cholesky decomposition of repeated sparse matrices
}
\note{
These are often just bailout methods, but some special
\code{strucMatrix} matrices have exploitable structure, which can be
used to keep the number of repeated values down.
}
\examples{
x <- strucMatrixCompSymm(1.2, -0.11, 4)
as.matrix(chol(x))
t(chol(as.matrix(x)))
}

