% Generated by roxygen2 (4.0.1): do not edit by hand
\name{chol.repSparse}
\alias{chol.repSparse}
\alias{chol.repSparseCompSymm}
\alias{chol.repSparseOneOffDiag}
\title{Cholesky decomposition of repeated sparse matrices}
\usage{
chol.repSparse(x, ...)

chol.repSparseOneOffDiag(x, ...)

chol.repSparseCompSymm(x, ...)
}
\arguments{
\item{x}{an object that inherits from class
\code{\link{repSparse}}}

\item{...}{passed to subsequent functions}
}
\description{
Cholesky decomposition of repeated sparse matrices
}
\note{
These are often just bailout methods, but some special
\code{repSparse} matrices have exploitable structure, which can be
used to keep the number of repeated values down.
}
\examples{
x <- repSparseCompSymm(1.2, -0.11, 4)
as.matrix(chol(x))
t(chol(as.matrix(x)))
}
\seealso{
Other repeated.sparse.matrix.topics: \code{\link{.bind}},
  \code{\link{bind}}, \code{\link{rep.repSparse}};
  \code{\link{as.data.frame.repSparse}},
  \code{\link{as.data.table.repSparse}},
  \code{\link{as.matrix.repSparse}},
  \code{\link{as.repSparse}},
  \code{\link{as.repSparse.dCHMsimpl}},
  \code{\link{as.repSparse.dist}},
  \code{\link{as.repSparse.dsparseMatrix}},
  \code{\link{as.repSparse.factor}},
  \code{\link{as.repSparse.matrix}},
  \code{\link{as.repSparse.repSparse}},
  \code{\link{repSparse2CsparseMatrix}},
  \code{\link{repSparse2TsparseMatrix}},
  \code{\link{repSparse2dgCMatrix}},
  \code{\link{repSparse2dgTMatrix}},
  \code{\link{repSparse2sparseMatrix}};
  \code{\link{diag.repSparse}},
  \code{\link{dim.repSparse}},
  \code{\link{image.repSparse}},
  \code{\link{plot.repSparse}},
  \code{\link{print.repSparse}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}}, \code{\link{t.repSparse}},
  \code{\link{update.repSparse}}; \code{\link{getInit}},
  \code{\link{getInit.default}},
  \code{\link{getInit.function}},
  \code{\link{getInit.repSparse}}, \code{\link{setInit}},
  \code{\link{setInit.default}},
  \code{\link{setInit.function}},
  \code{\link{setInit.repSparse}}; \code{\link{ind2point}},
  \code{\link{point2ind}}; \code{\link{kr}},
  \code{\link{kron}}; \code{\link{mkCholCompSymmTrans}},
  \code{\link{mkCholOneOffDiagTrans}},
  \code{\link{mkConstVarCholTrans}},
  \code{\link{mkCorMatCholTrans}},
  \code{\link{mkGenCholInPlaceTrans}},
  \code{\link{mkIdentityTrans}}, \code{\link{mkListTrans}},
  \code{\link{mkOuterTrans}}, \code{\link{mkTrans}},
  \code{\link{mkTransExpChol}},
  \code{\link{mkVarExpTrans}},
  \code{\link{mkVarFixedTrans}},
  \code{\link{mkVarIdentTrans}},
  \code{\link{mkVarPowerTrans}};
  \code{\link{repSparseBlank}}, \code{\link{repSparseCol}},
  \code{\link{repSparseCompSymm}},
  \code{\link{repSparseConstVarChol}},
  \code{\link{repSparseCorMatChol}},
  \code{\link{repSparseDiag}},
  \code{\link{repSparseExpChol}},
  \code{\link{repSparseIdent}}, \code{\link{repSparseInd}},
  \code{\link{repSparseOneOffDiag}},
  \code{\link{repSparseOnes}},
  \code{\link{repSparseSpecial}},
  \code{\link{repSparseSymm}}, \code{\link{repSparseTri}},
  \code{\link{repSparseVarWithCovariate}};
  \code{\link{repSparse}}
}

