% Generated by roxygen2 (4.0.1): do not edit by hand
\name{countUnique}
\alias{countInRange}
\alias{countUnique}
\title{Count unique values}
\usage{
countUnique(x)

countInRange(x)
}
\arguments{
\item{x}{numeric (for \code{countUnique}) or integer (for
\code{countInRange}) vector}
}
\value{
\code{\link{data.frame}} with \code{uniqueVals} and
\code{counts} columns
}
\description{
Count unique values
}
\examples{
x <- rep(1:5, 5:1)
countUnique(x)
countInRange(x)
}

