% Generated by roxygen2 (4.0.1): do not edit by hand
\name{denseDiagInds}
\alias{denseDiagInds}
\alias{denseLowerInds}
\alias{denseUpperInds}
\title{Indices for square dense matrices}
\usage{
denseDiagInds(n)

denseLowerInds(n)

denseUpperInds(n)
}
\arguments{
\item{n}{matrix size}
}
\description{
Indices for square dense matrices
}

