% Generated by roxygen2 (4.0.1): do not edit by hand
\name{findPathFromNode}
\alias{edgeTipIndicator}
\alias{edgeTipIndicator.default}
\alias{edgeTipIndicator.matrix}
\alias{edgeTipIndicator.phylo}
\alias{findEdgesFromPath}
\alias{findPathFromNode}
\title{Relationships between tips and edges}
\usage{
findPathFromNode(node, edge)

findEdgesFromPath(path, edge)

edgeTipIndicator(object, ...)

\method{edgeTipIndicator}{default}(object, ...)

\method{edgeTipIndicator}{matrix}(object, ntip, ...)

\method{edgeTipIndicator}{phylo}(object, ...)
}
\arguments{
\item{node}{vector of node indices}

\item{edge}{phylogenetic edge matrix}

\item{path}{vector giving the path from a node back in time
through its ancestor nodes (output of \code{findFromNode})}

\item{object}{either a \code{phylo} or \code{matrix} object}

\item{ntip}{number of tips}
}
\description{
\code{findFromNode} is a recurrsive function for computing the
path from a node back in time through its ancestor nodes.
\code{findEdgesFromPath} returns a \code{logical} vector
indicating the edges associated with a particular path.
\code{edgeTipIndicator} returns a \code{logical} matrix giving the
relationship between tips and the edges associated with their
history.
}
\examples{
set.seed(1)
phy <- rtree(5)
plot(phy)
edgelabels()
edgeTipIndicator(phy)
}

