% Generated by roxygen2 (4.0.1): do not edit by hand
\name{generalGlmer}
\alias{covar.generalGlmer}
\alias{fixef.generalGlmer}
\alias{generalGlmer}
\alias{getGeneralGlmerPar}
\alias{loads.generalGlmer}
\alias{print.generalGlmer}
\title{General glmer}
\usage{
generalGlmer(formula, data, addArgs = list(), optVerb = 0L, weights, offset,
  etastart, ...)

print.generalGlmer(x, ...)

getGeneralGlmerPar(object, type, ...)

\method{fixef}{generalGlmer}(object, ...)

\method{covar}{generalGlmer}(object, ...)

loads.generalGlmer(object, ...)
}
\arguments{
\item{formula}{extended mixed model formula}

\item{data}{data frame}

\item{addArgs}{list of additional arguments to pass to}

\item{optVerb}{verbose
\code{\link{setReTrm}} methods}

\item{...}{further arguments to \code{\link{mkGeneralGlmerDevfun}}}

\item{type}{character string giving the type of parameter
(e.g. \code{"fixef", "covar"})}
}
\description{
General glmer
}

