% Generated by roxygen2 (4.0.1): do not edit by hand
\name{generalParseFormula}
\alias{generalParseFormula}
\alias{simGeneralParsedForm}
\title{Parse a mixed model formula}
\usage{
generalParseFormula(formula, data, addArgs = list(), reTrmsList = NULL, ...)

simGeneralParsedForm(parsedForm, family = binomial, weights, ...)
}
\arguments{
\item{formula}{mixed model formula}

\item{data}{an object coercible to data frame}

\item{addArgs}{list of additional arguments to
\code{\link{setReTrm}} methods}

\item{reTrmsList}{if \code{NULL} \code{\link{mkReStructs}} is used}

\item{...}{additional parameters to \code{\link{as.data.frame}}}

\item{parsedForm}{result of \code{generalParseFormula}}

\item{family}{family object}

\item{weights}{optional weights}

\item{...}{not used}
}
\description{
Parse a mixed model formula
}

