% Generated by roxygen2 (4.0.1): do not edit by hand
\name{generalParseFormula}
\alias{generalParseFormula}
\title{Parse a mixed model formula}
\usage{
generalParseFormula(formula, data, ...)
}
\arguments{
\item{formula}{mixed model formula}

\item{data}{an object coercible to data frame}

\item{...}{additional parameters to \code{\link{as.data.frame}}}
}
\description{
Parse a mixed model formula
}

