% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getInit}
\alias{getInit}
\alias{getInit.default}
\alias{getInit.function}
\alias{getInit.reTrmStruct}
\alias{getInit.strucMatrix}
\alias{getInit.strucParseFormula}
\alias{setInit}
\alias{setInit.default}
\alias{setInit.function}
\alias{setInit.reTrmStruct}
\alias{setInit.strucMatrix}
\alias{setInit.strucParseFormula}
\title{Get and set initial parameter values for repeated sparse matrices}
\usage{
getInit(x, ...)

\method{getInit}{default}(x, ...)

\method{getInit}{strucMatrix}(x, ...)

\method{getInit}{function}(x, ...)

\method{getInit}{reTrmStruct}(x, ...)

\method{getInit}{strucParseFormula}(x, ...)

setInit(x, ...)

\method{setInit}{default}(x, init, ...)

\method{setInit}{strucMatrix}(x, init, ...)

\method{setInit}{function}(x, init, ...)

\method{setInit}{reTrmStruct}(x, initCovar, initLoads, ...)

\method{setInit}{strucParseFormula}(x, init, parType, ...)
}
\arguments{
\item{x}{object}

\item{...}{not yet used}

\item{init}{initial value for the parameter vector}

\item{initCovar}{initial value for the covariance parameter vector}

\item{initLoads}{initial value for the loadings parameter vector}

\item{parType}{character string giving the type of parameter to set
(if \code{parType} is missing, then guidance is provided by printing
available \code{parType}s). types can be \code{"fixef"} if a fixed
effects model matrix is present, \code{"weigh"} if weights
parameters are present, \code{"covar.trmName"} if covariance
parameters are present where \code{"trmName"} is the name of a
random effects term, and \code{"loads.trmName"} if loadings are
present.}
}
\description{
The initial values for a repeated sparse matrix are stored in the
environment of its transformation function, where they have the
name \code{init}.  Random effects term structures
(\code{\link{reTrmStruct}}) contain two repeated sparse matrices:
the transposed relative covariance factor, \code{Lambdat}, and the
transposed random effects model matrix, \code{Zt}. There are
\code{getReTrm.reTrmStruct} and \code{setReTrm.reTrmStruct}
methods for getting and setting the initial values of these
matrices directly.  Setting initial values does not change the
values themselves, only the initial values. Use the
\code{\link{update.strucMatrix}} methods to actually update the
values to the initial values.
}
\examples{
set.seed(1)
m1 <- as.strucMatrix(matrix(rnorm(6), 2, 3))
m2 <- strucMatrixCompSymm(1.2, -0.2, 5)
getInit(m1)
getInit(m2)
setInit(m2, c(10, 0))
}

