% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getInit}
\alias{getInit}
\alias{getInit.default}
\alias{getInit.function}
\alias{getInit.reTrmStruct}
\alias{getInit.repSparse}
\alias{parLength}
\alias{setInit}
\alias{setInit.default}
\alias{setInit.function}
\alias{setInit.reTrmStruct}
\alias{setInit.repSparse}
\title{Get and set initial parameter values for repeated sparse matrices}
\usage{
getInit(x, ...)

\method{getInit}{default}(x, ...)

\method{getInit}{repSparse}(x, ...)

\method{getInit}{function}(x, ...)

\method{getInit}{reTrmStruct}(x, ...)

setInit(x, ...)

\method{setInit}{default}(x, init, ...)

\method{setInit}{repSparse}(x, init, ...)

\method{setInit}{function}(x, init, ...)

\method{setInit}{reTrmStruct}(x, initCovar, initLoads, ...)

parLength(ll)
}
\arguments{
\item{x}{object}

\item{...}{not yet used}

\item{init}{initial value for the parameter vector}

\item{initCovar}{initial value for the covariance parameter vector}

\item{initLoads}{initial value for the loadings parameter vector}

\item{ll}{length of parameter vector}
}
\description{
Get and set initial parameter values for repeated sparse matrices
}
\examples{
set.seed(1)
m1 <- as.repSparse(matrix(rnorm(6), 2, 3))
m2 <- repSparseCompSymm(1.2, -0.2, 5)
getInit(m1)
getInit(m2)
}
\seealso{
Other repeated sparse matrix topics: \code{\link{as.data.frame.repSparse}},
  \code{\link{as.matrix.repSparse}},
  \code{\link{as.repSparse}},
  \code{\link{as.repSparse.dCHMsimpl}},
  \code{\link{as.repSparse.dist}},
  \code{\link{as.repSparse.dsparseMatrix}},
  \code{\link{as.repSparse.factor}},
  \code{\link{as.repSparse.matrix}},
  \code{\link{as.repSparse.repSparse}},
  \code{\link{repSparse2CsparseMatrix}},
  \code{\link{repSparse2TsparseMatrix}},
  \code{\link{repSparse2dgCMatrix}},
  \code{\link{repSparse2dgTMatrix}},
  \code{\link{repSparse2sparseMatrix}};
  \code{\link{chol.repSparse}},
  \code{\link{chol.repSparseCompSymm}},
  \code{\link{chol.repSparseOneOffDiag}};
  \code{\link{diag.repSparse}},
  \code{\link{dim.repSparse}},
  \code{\link{image.repSparse}},
  \code{\link{plot.repSparse}},
  \code{\link{print.repSparse}},
  \code{\link{repSparse-class}}, \code{\link{t.repSparse}},
  \code{\link{update.repSparse}}; \code{\link{ind2point}},
  \code{\link{point2ind}};
  \code{\link{mkCholCompSymmTrans}},
  \code{\link{mkCholOneOffDiagTrans}},
  \code{\link{mkConstVarCholTrans}},
  \code{\link{mkCorMatCholTrans}},
  \code{\link{mkExpCholTrans}},
  \code{\link{mkFlexDiagTrans}},
  \code{\link{mkGenCholInPlaceTrans}},
  \code{\link{mkIdentityTrans}}, \code{\link{mkListTrans}},
  \code{\link{mkMultiVarExpTrans}},
  \code{\link{mkOuterTrans}}, \code{\link{mkTrans}},
  \code{\link{mkVarExpTrans}},
  \code{\link{mkVarFixedTrans}},
  \code{\link{mkVarIdentTrans}},
  \code{\link{mkVarPowerTrans}}; \code{\link{repSparse}}
}

