% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getModMatAndGrpFac}
\alias{getModMatAndGrpFac}
\title{Get model matrix and grouping factor}
\usage{
getModMatAndGrpFac(bar, fr)
}
\arguments{
\item{bar}{random effect language object (e.g. \code{x | g})}

\item{fr}{model frame}
}
\value{
list with model matrix and grouping factor
}
\description{
This is kind of like \code{\link{model.matrix}} but for random
effects terms.
}

