% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getParms}
\alias{getParms}
\title{Get parameters from deviance function}
\usage{
getParms(dfun, parms = c("theta", "phi"), .unlist = TRUE)
}
\arguments{
\item{dfun}{deviance function created from \code{\link{logisticPcaDevfun}}}

\item{parms}{character vectors of parameter names}

\item{.unlist}{should the results be \code{\link{unlist}}ed?}
}
\value{
model parameters at their current value
}
\description{
Get parameters from deviance function
}

