% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getRepPattern}
\alias{getRepPattern}
\title{Get repetition pattern of a repeated sparse matrix}
\usage{
getRepPattern(object)
}
\arguments{
\item{object}{a \code{\link{repSparse}} object}
}
\description{
Get repetition pattern of a repeated sparse matrix
}

