% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getReTrm}
\alias{getReTrm}
\alias{getReTrm.identity}
\title{Get random effects term}
\usage{
getReTrm(modMat, grpFac, grpName)

\method{getReTrm}{identity}(modMat, grpFac, grpName)
}
\arguments{
\item{modMat}{raw random effects model matrix}

\item{grpFac}{grouping factor}

\item{grpName}{grouping factor name}
}
\description{
Get random effects term
}

