% Generated by roxygen2 (4.0.1): do not edit by hand
\name{glmerc}
\alias{glmerc}
\alias{print.glmerc}
\title{Generalized linear mixed model with custom covariance over grouping factor levels}
\usage{
glmerc(formula, data = NULL, family = binomial, covList = list(),
  optControl = list(iprint = 0L), ...)

print.glmerc(x, ...)
}
\arguments{
\item{formula}{formula}

\item{data}{data}

\item{family}{family}

\item{covList}{covList}

\item{optControl}{optControl}

\item{...}{...}
}
\description{
When fitting generalized linear mixed model with
\code{\link{glmer}}, the covariance structure over the grouping
factor levels is assumed to be an identity matrix.  That is,
random effects are sampled indepedently over the grouping factor
levels.  With \code{glmerc}, one may specify the covariance over
the levels, up to a fitted parameter.
}

