% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{kron}
\alias{kr}
\alias{kron}
\alias{kronecker,strucMatrix,strucMatrix-method}
\title{Kronecker and Khatri-Rao products for repeated sparse matrices}
\usage{
kron(X, Y, trans = "*", makedimnames = FALSE, ...)

kr(X, Y, trans = "*")

\S4method{kronecker}{strucMatrix,strucMatrix}(X, Y, FUN = "*",
  make.dimnames = FALSE, ...)
}
\arguments{
\item{X,Y}{repeated sparse matrices (\code{\link{strucMatrix-class}})}

\item{trans}{see argument \code{FUN} in \code{\link{outer}}}

\item{makedimnames}{ignored}

\item{...}{ignored}

\item{FUN}{a function; it may be a quoted string.}

\item{make.dimnames}{Provide dimnames that are the product of the
   dimnames of \code{X} and \code{Y}.}
}
\description{
Kronecker and Khatri-Rao products for repeated sparse matrices
}
\examples{
set.seed(1)
X <- strucMatrix(c(1, 2, 1, 2), c(1, 1, 2, 2), 1:4, rnorm(4))
Y <- strucMatrix(c(1, 2, 1), c(1, 1, 2), 1:3, rnorm(3))
(kronExample <- kron(X, Y))
(krExample <- kr(X, Y))
}
\seealso{
Other matrixCombining: \code{\link{.bind}},
  \code{\link{bind}}, \code{\link{rep.strucMatrix}};
  \code{\link{strucMatrixRowSubset}},
  \code{\link{subset.strucMatrix}}

Other matrixCombining: \code{\link{.bind}},
  \code{\link{bind}}, \code{\link{rep.strucMatrix}};
  \code{\link{strucMatrixRowSubset}},
  \code{\link{subset.strucMatrix}}
}

