% Generated by roxygen2 (4.0.1): do not edit by hand
\name{kron}
\alias{kr}
\alias{kron}
\title{Kronecker and Khatri-Rao products for repeated sparse matrices}
\usage{
kron(X, Y, trans = "*", makedimnames = FALSE, ...)

kr(X, Y, trans = "*")
}
\arguments{
\item{X,Y}{\code{repSparse} objects}

\item{trans}{see argument \code{FUN} in \code{\link{outer}}}

\item{makedimnames}{ignored}

\item{...}{ignored}
}
\description{
Kronecker and Khatri-Rao products for repeated sparse matrices
}

