% Generated by roxygen2 (4.0.1): do not edit by hand
\name{langOps}
\alias{binaryLangOp}
\alias{langOps}
\alias{listLangOp}
\alias{mkBinaryLangOp}
\alias{unaryLangOp}
\title{Tools for constructing formula terms}
\usage{
binaryLangOp(arg1, arg2, Op)

unaryLangOp(arg, Op)

listLangOp(args, Op)

mkBinaryLangOp(Op)
}
\arguments{
\item{arg1,arg2,arg}{a language object or length-one character
vector}

\item{args}{list of language objects and/or length-one character
objects}

\item{Op}{name of an operator (as language or character)}
}
\description{
Tools for constructing formula terms
}
\examples{
(oneTerm <- binaryLangOp("a", "b", "|"))
(parenTerm <- unaryLangOp(oneTerm, "("))
(manyTerms <- mapply(binaryLangOp, "x", letters[1:5], "|", USE.NAMES = FALSE))
(manyParenTerms <- lapply(manyTerms, unaryLangOp, "("))
listLangOp(manyParenTerms, "+")
}

