% Generated by roxygen2 (4.0.1): do not edit by hand
\name{kron}
\alias{kr}
\alias{kron}
\title{Kronecker and Khatri-Rao products for repeated sparse matrices}
\usage{
kron(X, Y, trans = "*", makedimnames = FALSE, ...)

kr(X, Y, trans = "*")
}
\arguments{
\item{X,Y}{repeated sparse matrices (\code{\link{repSparse-class}})}

\item{trans}{see argument \code{FUN} in \code{\link{outer}}}

\item{makedimnames}{ignored}

\item{...}{ignored}
}
\description{
Kronecker and Khatri-Rao products for repeated sparse matrices
}
\examples{
set.seed(1)
X <- repSparse(c(1, 2, 1, 2), c(1, 1, 2, 2), 1:4, rnorm(4))
Y <- repSparse(c(1, 2, 1), c(1, 1, 2), 1:3, rnorm(3))
(kronExample <- kron(X, Y))
(krExample <- kr(X, Y))
}
\seealso{
Other matrixCombining: \code{\link{.bind}},
  \code{\link{bind}}, \code{\link{rep.repSparse}};
  \code{\link{repSparseRowSubset}},
  \code{\link{subset.repSparse}}

Other matrixCombining: \code{\link{.bind}},
  \code{\link{bind}}, \code{\link{rep.repSparse}};
  \code{\link{repSparseRowSubset}},
  \code{\link{subset.repSparse}}
}

