% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{kron}
\alias{Ops.repSparse}
\alias{crossprod,repSparse,missing-method}
\alias{crossprod,repSparse,repSparse-method}
\alias{kr}
\alias{kron}
\alias{kronecker,repSparse,repSparse-method}
\alias{tcrossprod,repSparse,missing-method}
\alias{tcrossprod,repSparse,repSparse-method}
\title{Kronecker and Khatri-Rao products for repeated sparse matrices}
\usage{
kron(X, Y, trans = "*", makedimnames = FALSE, ...)

kr(X, Y, trans = "*")

\S4method{kronecker}{repSparse,repSparse}(X, Y, FUN = "*",
  make.dimnames = FALSE, ...)

\method{Ops}{repSparse}(e1, e2)

\S4method{crossprod}{repSparse,missing}(x, y = NULL, ...)

\S4method{tcrossprod}{repSparse,missing}(x, y = NULL, ...)

\S4method{crossprod}{repSparse,repSparse}(x, y = NULL, ...)

\S4method{tcrossprod}{repSparse,repSparse}(x, y = NULL, ...)
}
\arguments{
\item{X,Y}{repeated sparse matrices (\code{\link{repSparse-class}})}

\item{trans}{see argument \code{FUN} in \code{\link{outer}}}

\item{makedimnames}{ignored}

\item{...}{ignored}

\item{e1,e2}{\code{repSparse} objects}

\item{x,y}{\code{\link{repSparse}} matrix objects}

\item{FUN}{a function; it may be a quoted string.}

\item{make.dimnames}{Provide dimnames that are the product of the
   dimnames of \code{X} and \code{Y}.}
}
\description{
Kronecker and Khatri-Rao products for repeated sparse matrices
}
\note{
The \code{*} operator is matrix multiplication, not
element-wise multiplication.
}
\examples{
set.seed(1)
X <- repSparse(c(1, 2, 1, 2), c(1, 1, 2, 2), 1:4, rnorm(4))
Y <- repSparse(c(1, 2, 1), c(1, 1, 2), 1:3, rnorm(3))
(kronExample <- kron(X, Y))
(krExample <- kr(X, Y))
}
\seealso{
Other matrixCombining: \code{\link{.bind}},
  \code{\link{bind}}, \code{\link{rep.repSparse}};
  \code{\link{repSparseRowSubset}},
  \code{\link{subset.repSparse}}

Other matrixCombining: \code{\link{.bind}},
  \code{\link{bind}}, \code{\link{rep.repSparse}};
  \code{\link{repSparseRowSubset}},
  \code{\link{subset.repSparse}}
}

