% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rowWiseCombination}
\alias{kr}
\alias{kr0}
\alias{kron}
\alias{mmult}
\alias{rowWiseCombination}
\title{Row-wise combination of two repeated sparse matrices}
\usage{
rowWiseCombination(X, Y)

mmult(X, Y, trans = "*")

kron(X, Y, trans = "*", makedimnames = FALSE, saveComponents = FALSE, ...)

kr(X, Y, trans = "*", saveComponents = FALSE)

kr0(X, Y, trans = "*", saveComponents = FALSE)
}
\arguments{
\item{X,Y}{\code{repSparse} objects}

\item{trans}{see argument \code{FUN} in \code{\link{outer}}}

\item{makedimnames}{ignored}

\item{saveComponents}{should component matrices be saved?}

\item{...}{ignored}
}
\value{
a row-wise combination of repeated sparse matrices
}
\description{
Row-wise combination of two repeated sparse matrices

Standard matrix multiplication for repeated sparse matrices

Kronecker and Khatri-Rao products for repeated sparse matrices
}
\seealso{
Other repSparseTopics: \code{\link{.bind}},
  \code{\link{bind}}, \code{\link{rep.repSparse}};
  \code{\link{as}}, \code{\link{as}}, \code{\link{as}},
  \code{\link{as}}, \code{\link{as}},
  \code{\link{as.data.frame.repSparse}},
  \code{\link{as.data.table.repSparse}},
  \code{\link{as.matrix.repSparse}},
  \code{\link{as.repSparse}},
  \code{\link{as.repSparse.dsparseMatrix}},
  \code{\link{as.repSparse.factor}},
  \code{\link{as.repSparse.matrix}},
  \code{\link{as.repSparse.repSparse}};
  \code{\link{chol.repSparse}},
  \code{\link{chol.repSparseCompSymm}},
  \code{\link{chol.repSparseOneOffDiag}};
  \code{\link{dim.repSparse}},
  \code{\link{image.repSparse}},
  \code{\link{print.repSparse}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{sort.repSparse}}, \code{\link{standardSort}},
  \code{\link{t.repSparse}},
  \code{\link{update.repSparse}}; \code{\link{getInit}},
  \code{\link{getInit.default}},
  \code{\link{getInit.function}},
  \code{\link{getInit.repSparse}}, \code{\link{setInit}},
  \code{\link{setInit.default}},
  \code{\link{setInit.function}},
  \code{\link{setInit.repSparse}}; \code{\link{ind2point}},
  \code{\link{point2ind}};
  \code{\link{mkCholOneOffDiagTrans}},
  \code{\link{mkConstVarCholTrans}},
  \code{\link{mkCorMatCholTrans}},
  \code{\link{mkGenCholInPlaceTrans}},
  \code{\link{mkIdentityTrans}}, \code{\link{mkListTrans}},
  \code{\link{mkOuterTrans}}; \code{\link{rRepSparse}},
  \code{\link{repSparseCompSymm}},
  \code{\link{repSparseConstVarChol}},
  \code{\link{repSparseCorMatChol}},
  \code{\link{repSparseDiag}},
  \code{\link{repSparseIdent}},
  \code{\link{repSparseOneOffDiag}},
  \code{\link{repSparseTri}}; \code{\link{repSparse}}
}

