% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mkGblmerDevfun}
\alias{mkGblmerDevfun}
\title{Modular functions for generalized bilinear mixed models}
\usage{
mkGblmerDevfun(fr, X, linReTrms, bilinReTrms, family, nAGQ = 1,
  verbose = 0L, ...)
}
\description{
Modular functions for generalized bilinear mixed models
}

