% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mkGeneralGlmerDevfun}
\alias{mkGeneralGlmerDevfun}
\title{Make a general (Laplace) glmer deviance function}
\usage{
mkGeneralGlmerDevfun(y, X, Zt, Lambdat, weights = NULL, offset = NULL,
  etastart = NULL, mustart = NULL, devfunEnv, initPars, parInds,
  mapToCovFact = NULL, mapToModMat = NULL, mapToWeights = NULL, family,
  tolPwrss = 1e-06, verbose = 0L, pureR = FALSE, Lind = NULL)
}
\arguments{
\item{y}{response vector}

\item{X}{fixed effects model matrix}

\item{Zt}{transposed random effects model matrix}

\item{Lambdat}{relative covariance factor}

\item{weights}{prior weights}

\item{offset}{known additive offsets}

\item{initPars}{initial values for the parameter vector}

\item{parInds}{a named list of vectors for identifying each of the
three types of parameters in \code{initPars}, one vector for each
of the three types of parameters: \code{list(covar = ., fixef = .,
loads = .)} where the \code{.}'s are vectors of indices}

\item{mapToCovFact}{function taking the \code{covar} parameters
and returning the values of the non-zero elements of
\code{Lambdat} (i.e. the \code{x} slot of \code{Lambdat})}

\item{mapToModMat}{function taking the \code{loads} parameters and
returning the values of the non-zero elements of \code{Zt}
(i.e. the \code{x} slot of \code{Zt})}

\item{family}{\code{\link{family}}}

\item{tolPwrss}{tolerance for penalized weighted residual sum of
squares}

\item{verbose}{verbose}

\item{pureR}{should the PIRLS algorithm be run in
\code{lme4pureR}?}

\item{Lind}{optional indices mapping covariance parameters to the
non-zero elements of the relative covariance factor}
}
\value{
a deviance function with an environment
}
\description{
Make a general (Laplace) glmer deviance function
}

