% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mkGeneralReTrms}
\alias{mkGeneralReTrms}
\alias{mkIdentityReTrms}
\alias{mkSparseTrans}
\title{Make general random effects terms}
\usage{
mkGeneralReTrms(reStructList, parsedForm, ...)

mkSparseTrans(object)

mkIdentityReTrms(modMat, grpFac, grpName)
}
\arguments{
\item{reStructList}{list of random effects structure functions}

\item{parsedForm}{list parsed random effects terms}

\item{...}{potential extra parameters}

\item{object}{repeated sparse matrix object}

\item{modMat}{raw random effects model matrix}

\item{grpFac}{grouping factor}

\item{grpName}{grouping factor name}
}
\description{
Make general random effects terms

Make transformation function for column-compressed sparse matrix
}

