% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mkGeneralReTrms}
\alias{mkGeneralReTrms}
\title{Make general random effects terms}
\usage{
mkGeneralReTrms(reStructList, parsedForm, ...)
}
\arguments{
\item{reStructList}{list of random effects structure functions}

\item{parsedForm}{list parsed random effects terms}

\item{...}{potential extra parameters}
}
\description{
Make general random effects terms
}

