% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mkGeneralReTrms}
\alias{mkGeneralReTrms}
\alias{mkIdentityReTrms}
\alias{mkSparseTrans}
\title{Make general random effects terms}
\usage{
mkGeneralReTrms(trmList, ...)

mkSparseTrans(object)

mkIdentityReTrms(trm, ...)
}
\arguments{
\item{trmList}{list of terms}

\item{...}{potential extra parameters}

\item{object}{repeated sparse matrix object}

\item{trm}{random effects term}
}
\description{
Make general random effects terms

Make transformation function for column-compressed sparse matrix
}

