% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mkTemplateTermLambdat}
\alias{mkTemplateTermLambdat}
\title{Make covariance factor for a covariance template term}
\usage{
mkTemplateTermLambdat(covMat, nLevels, diagModel = TRUE)
}
\arguments{
\item{covMat}{template covariance matrix}

\item{nLevels}{number of levels over which the grouping factor varies}

\item{diagModel}{templates on the block diagonal (\code{TRUE}) or
densely tiled over the entire matrix (\code{FALSE})?}
}
\description{
Make covariance factor for a covariance template term
}

