% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mkIdentityTrans}
\alias{mkCholOneOffDiagTrans}
\alias{mkIdentityTrans}
\alias{mkListTrans}
\alias{mkOuterTrans}
\title{Construct functions for transforming a parameter vector to the
non-zero values of a repeated sparse matrix}
\usage{
mkIdentityTrans(init)

mkOuterTrans(Atrans, Btrans, ABtrans)

mkListTrans(transList)

mkCholOneOffDiagTrans(init)
}
\arguments{
\item{Atrans,Btrans}{functions for transforming \code{A} and
\code{B}}

\item{ABtrans}{function to pass as \code{FUN} in
\code{\link{outer}}}

\item{transList}{list of transform functions}

\item{init}{initial parameter values}
}
\description{
Construct functions for transforming a parameter vector to the
non-zero values of a repeated sparse matrix
}
\seealso{
Other repSparseTopics: \code{\link{.bind}},
  \code{\link{bind}}, \code{\link{rep.repSparse}};
  \code{\link{as}}, \code{\link{as}}, \code{\link{as}},
  \code{\link{as}}, \code{\link{as}},
  \code{\link{as.data.frame.repSparse}},
  \code{\link{as.data.table.repSparse}},
  \code{\link{as.matrix.repSparse}},
  \code{\link{as.repSparse}},
  \code{\link{as.repSparse.dsparseMatrix}},
  \code{\link{as.repSparse.matrix}},
  \code{\link{as.repSparse.repSparse}};
  \code{\link{chol.repSparse}},
  \code{\link{chol.repSparseOneOffDiag}};
  \code{\link{dim.repSparse}},
  \code{\link{image.repSparse}},
  \code{\link{print.repSparse}},
  \code{\link{repSparse-class}},
  \code{\link{sort.repSparse}}, \code{\link{t.repSparse}},
  \code{\link{update.repSparse}}; \code{\link{getInit}},
  \code{\link{getInit.default}},
  \code{\link{getInit.function}},
  \code{\link{getInit.repSparse}}, \code{\link{setInit}},
  \code{\link{setInit.default}},
  \code{\link{setInit.function}},
  \code{\link{setInit.repSparse}}; \code{\link{ind2point}},
  \code{\link{point2ind}}; \code{\link{kr}},
  \code{\link{kron}}, \code{\link{mmult}},
  \code{\link{rowWiseCombination}};
  \code{\link{rRepSparse}},
  \code{\link{repSparseCompSymm}},
  \code{\link{repSparseDiag}},
  \code{\link{repSparseIdent}},
  \code{\link{repSparseOneOffDiag}};
  \code{\link{repSparse}}
}

