% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mkIdentityTrans}
\alias{mkCholCompSymmTrans}
\alias{mkCholOneOffDiagTrans}
\alias{mkConstVarCholTrans}
\alias{mkCorMatCholTrans}
\alias{mkGenCholInPlaceTrans}
\alias{mkIdentityTrans}
\alias{mkListTrans}
\alias{mkOuterTrans}
\alias{mkVarExpTrans}
\alias{mkVarFixedTrans}
\alias{mkVarIdentTrans}
\alias{mkVarPowerTrans}
\title{Construct functions for transforming a parameter vector to the
non-zero values of a repeated sparse matrix}
\usage{
mkIdentityTrans(init)

mkCholOneOffDiagTrans(init)

mkGenCholInPlaceTrans(init)

mkConstVarCholTrans(init)

mkCorMatCholTrans(init)

mkOuterTrans(Atrans, Btrans, ABtrans)

mkListTrans(transList)

mkVarExpTrans(init, covariate, grpFac)

mkVarPowerTrans(init, covariate, grpFac)

mkVarIdentTrans(init, covariate, grpFac)

mkVarFixedTrans(init, covariate, grpFac)

mkCholCompSymmTrans(init, matSize)
}
\arguments{
\item{init}{initial parameter values}

\item{Atrans,Btrans}{functions for transforming two repeated
sparse matrices, \code{A} and \code{B} say}

\item{ABtrans}{function to pass as \code{FUN} in
\code{\link{outer}}}

\item{transList}{list of transformation functions}

\item{covariate}{covariate}

\item{grpFac}{a grouping factor (or anything coercible to
\code{numeric} really) with the number of levels equal to the
length of \code{init}}

\item{matSize}{number of rows (or columns) of the matrix}
}
\description{
These functions return a 'trans function', for transforming a
parameter vector to the repeated non-zero values of a repeated
sparse matrix.  Each trans function takes one vector argument
called \code{matPars}, which are the parameters of the matrix.
The environment of these transformation functions must contain a
vector called \code{init}, which contains the initial values (aka,
the prototype) of \code{matPars}.
}
\seealso{
Other repeated.sparse.matrix.topics: \code{\link{.bind}},
  \code{\link{bind}}, \code{\link{rep.repSparse}};
  \code{\link{as.data.frame.repSparse}},
  \code{\link{as.data.table.repSparse}},
  \code{\link{as.matrix.repSparse}},
  \code{\link{as.repSparse}},
  \code{\link{as.repSparse.dsparseMatrix}},
  \code{\link{as.repSparse.factor}},
  \code{\link{as.repSparse.matrix}},
  \code{\link{as.repSparse.repSparse}},
  \code{\link{repSparse2CsparseMatrix}},
  \code{\link{repSparse2TsparseMatrix}},
  \code{\link{repSparse2dgCMatrix}},
  \code{\link{repSparse2dgTMatrix}},
  \code{\link{repSparse2sparseMatrix}};
  \code{\link{chol.repSparse}},
  \code{\link{chol.repSparseCompSymm}},
  \code{\link{chol.repSparseOneOffDiag}};
  \code{\link{dim.repSparse}},
  \code{\link{image.repSparse}},
  \code{\link{plot.repSparse}},
  \code{\link{print.repSparse}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}}, \code{\link{t.repSparse}},
  \code{\link{update.repSparse}}; \code{\link{getInit}},
  \code{\link{getInit.default}},
  \code{\link{getInit.function}},
  \code{\link{getInit.repSparse}}, \code{\link{setInit}},
  \code{\link{setInit.default}},
  \code{\link{setInit.function}},
  \code{\link{setInit.repSparse}}; \code{\link{ind2point}},
  \code{\link{point2ind}}; \code{\link{kr}},
  \code{\link{kron}}; \code{\link{rRepSparse}},
  \code{\link{repSparseBlank}},
  \code{\link{repSparseCompSymm}},
  \code{\link{repSparseConstVarChol}},
  \code{\link{repSparseCorMatChol}},
  \code{\link{repSparseDiag}},
  \code{\link{repSparseIdent}},
  \code{\link{repSparseOneOffDiag}},
  \code{\link{repSparseSymm}}, \code{\link{repSparseTri}},
  \code{\link{repSparseVarWithCovariate}};
  \code{\link{repSparse}}
}

