% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mkIdentityTrans}
\alias{mkCholOneOffDiagTrans}
\alias{mkIdentityTrans}
\alias{mkListTrans}
\alias{mkOuterTrans}
\title{Construct functions for transforming a parameter vector to the
non-zero values of a repeated sparse matrix}
\usage{
mkIdentityTrans(init)

mkOuterTrans(A, B, Atrans, Btrans, ABtrans)

mkListTrans(valsList, transList)

mkCholOneOffDiagTrans(init)
}
\arguments{
\item{A,B}{\code{numeric} \code{vector}s}

\item{Atrans,Btrans}{functions for transforming \code{A} and
\code{B}}

\item{ABtrans}{function to pass as \code{FUN} in
\code{\link{outer}}}

\item{valsList}{list of value vectors}

\item{transList}{list of transform functions}

\item{init}{initial parameter values}
}
\description{
Construct functions for transforming a parameter vector to the
non-zero values of a repeated sparse matrix
}

