% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mkIdentityTrans}
\alias{mkCholCompSymmTrans}
\alias{mkCholOneOffDiagTrans}
\alias{mkConstVarCholTrans}
\alias{mkCorMatCholTrans}
\alias{mkExpCholTrans}
\alias{mkFlexDiagTrans}
\alias{mkGenCholInPlaceTrans}
\alias{mkIdentityTrans}
\alias{mkListTrans}
\alias{mkMultiVarExpTrans}
\alias{mkOuterTrans}
\alias{mkTrans}
\alias{mkVarExpTrans}
\alias{mkVarFixedTrans}
\alias{mkVarIdentTrans}
\alias{mkVarPowerTrans}
\title{Construct functions for transforming a parameter vector to the
non-zero values of a repeated sparse matrix}
\usage{
mkIdentityTrans(init)

mkCholOneOffDiagTrans(init)

mkGenCholInPlaceTrans(init)

mkConstVarCholTrans(init)

mkCorMatCholTrans(init)

mkExpCholTrans(init, cholObj, symmObj, vecDist)

mkFlexDiagTrans(init, defaultOutput, devfunEnv)

mkOuterTrans(Atrans, Btrans, ABtrans)

mkListTrans(transList)

mkVarExpTrans(init, covariate, grpFac)

mkMultiVarExpTrans(init, modMat, grpFac)

mkVarPowerTrans(init, covariate, grpFac)

mkVarIdentTrans(init, covariate, grpFac)

mkVarFixedTrans(init, covariate, grpFac)

mkCholCompSymmTrans(init, matSize)
}
\arguments{
\item{init}{initial parameter values}

\item{cholObj,symmObj}{corresponding triangular cholesky and
symmetric \code{\link{repSparse}} matrix objects}

\item{vecDist}{distances as a vector}

\item{defaultOutput}{default vector for diagonal elements}

\item{devfunEnv}{environment of the deviance function}

\item{Atrans,Btrans}{functions for transforming two repeated
sparse matrices, \code{A} and \code{B} say}

\item{ABtrans}{function to pass as \code{FUN} in
\code{\link{outer}}}

\item{transList}{list of transformation functions}

\item{covariate}{covariate}

\item{grpFac}{a grouping factor (or anything coercible to
\code{numeric} really) with the number of levels equal to the
length of \code{init}}

\item{modMat}{model matrix}

\item{matSize}{number of rows (or columns) of the matrix}
}
\description{
These functions return a 'trans function', for transforming a
parameter vector to the repeated non-zero values of a repeated
sparse matrix.  Each trans function takes one vector argument
called \code{matPars}, which are the parameters of the matrix.
The environment of these transformation functions must contain a
vector called \code{init}, which contains the initial values (aka,
the prototype) of \code{matPars}.
}

