% Generated by roxygen2 (4.0.1): do not edit by hand
\name{parLength}
\alias{parLength}
\alias{parLength.repSparse}
\alias{parLength.strucGlmer}
\alias{parLength.strucParseFormula}
\title{Lengths of parameter vectors}
\usage{
parLength(object, ...)

\method{parLength}{repSparse}(object, ...)

\method{parLength}{strucParseFormula}(object, ...)

\method{parLength}{strucGlmer}(object, ...)
}
\arguments{
\item{object}{parameterized object}

\item{...}{extra parameters}
}
\description{
Lengths of parameter vectors
}

