% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pars}
\alias{covar}
\alias{covar.strucGlmer}
\alias{covarPerTerm}
\alias{factors}
\alias{fixef.strucGlmer}
\alias{getStrucGlmerPar}
\alias{loads}
\alias{loads.default}
\alias{loads.strucGlmer}
\alias{loadsPerTerm}
\alias{mkParInds}
\alias{nRePerTrm}
\alias{parPerTerm}
\alias{pars}
\alias{pars.glmerMod}
\alias{pars.strucGlmer}
\alias{ranef.strucGlmer}
\alias{reIndsPerTrm}
\title{Parameter retrieval for structured generalized linear mixed models}
\usage{
pars(object, ...)

\method{pars}{strucGlmer}(object, ...)

\method{pars}{glmerMod}(object, ...)

covar(object, ...)

\method{covar}{strucGlmer}(object, ...)

loads(object, ...)

\method{loads}{default}(object, ...)

\method{loads}{strucGlmer}(object, ...)

factors(x, ...)

\method{fixef}{strucGlmer}(object, ...)

\method{ranef}{strucGlmer}(object, type = c("u", "Lu", "ZLu"), ...)

getStrucGlmerPar(object, type, ...)

parPerTerm(nParPerTrm, pars)

covarPerTerm(object)

loadsPerTerm(object)

nRePerTrm(object)

reIndsPerTrm(object)

mkParInds(parList)
}
\arguments{
\item{object}{a \code{strucGlmer} fitted model object}

\item{...}{additional arguments to pass on}

\item{x}{an object with \code{x$factors}}

\item{type}{character string giving the type of parameter
(e.g. \code{"fixef", "covar"})}

\item{nParPerTrm}{vector of the number of parameters per term}

\item{pars}{parameter vector (e.g. result of \code{covar} or
\code{loads}}

\item{parList}{named list of parameters with possible names:
(\code{covar}, \code{fixef}, \code{weigh}, \code{loads})}
}
\description{
Parameter retrieval for structured generalized linear mixed models
}

