% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fixef.glmerc}
\alias{covar}
\alias{covar.glmerc}
\alias{covarByTerms}
\alias{fixef.glmerc}
\alias{loads}
\alias{loads.glmerc}
\alias{pars}
\alias{pars.glmerc}
\title{Get parameters}
\usage{
\method{fixef}{glmerc}(object, ...)

\method{covar}{glmerc}(object, ...)

loads.glmerc(object, ...)

\method{pars}{glmerc}(object, ...)

covarByTerms(object, ...)

covar(object, ...)

loads(object, ...)

pars(object, ...)
}
\arguments{
\item{...}{not used}

\item{object}{\code{lme4ord} fitted model object}
}
\description{
Get parameters
}

