% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repSparse-class}
\alias{dim.repSparse}
\alias{image.repSparse}
\alias{plot.repSparse}
\alias{print.repSparse}
\alias{repSparse-class}
\alias{t.repSparse}
\alias{update.repSparse}
\title{A repeated sparse matrix class}
\usage{
print.repSparse(x, n = 6L, ...)

update.repSparse(object, newPars, ...)

image.repSparse(x, plain = FALSE, ...)

\method{plot}{repSparse}(x, y, plain = FALSE...)

t.repSparse(x)

\method{dim}{repSparse}(x)
}
\arguments{
\item{x}{\code{repSparse} object}

\item{n}{how many indices and repeated values to print?}

\item{...}{passed to subsequent functions}

\item{object}{\code{repSparse} object}

\item{newPars}{new parameter values}

\item{plain}{should a completely plain plot be used? (try and see)}

\item{y}{not used}
}
\description{
An \code{S3} class for repeated sparse matrices, which are lists
with a \code{Dim} attribute and with the following elements:
\describe{
  \item{\code{rowInds}}{0-based row indices}
  \item{\code{colInds}}{0-based column indices}
  \item{\code{valInds}}{1-based indices for associating each \code{rowInds} and
                        \code{colInds} pair with the elements in \code{vals}}
  \item{\code{trans}}{A function for transforming a parameter vector
                      into \code{vals}.  This function is used by
                      \code{\link{update.repSparse}}}
}
Objects in this class can be constructed with
\code{\link{repSparse}} (and, for example,
\code{\link{repSparseDiag}}), and has several methods including
the following.
}
\details{
The \code{...} arguments for the \code{update.repSparse} method
can be used to specify special parameter arguments, if
\code{object$mkNewPars} exits (which is often the case for special
matrices such as \code{\link{repSparseTri}}.  This is a more
explicit approach and can therefore be more convenient than having
to figure out what order the parameters should appear in
\code{newPars}.  For example, \code{update(., diagVals = c(1, 1),
offDiagVals = -0.2)} is more explicit than \code{update(. c(1,
-0.2, 1))}.
}
\examples{
(X <- repSparse(rowInds = 1:6,
                colInds = rep(1:2, 3),
                valInds = rep(1:2, each = 3),
                vals    = c(-pi, pi)))
(Y <- update(X, c(0.2, 0.8)))
as.matrix(X, sparse = TRUE)
as.matrix(Y, sparse = TRUE)
image(kron(t(X), Y))
image(bind(X, Y, type = "diag"))
}
\seealso{
Other repeated.sparse.matrix.topics: \code{\link{.bind}},
  \code{\link{bind}}, \code{\link{rep.repSparse}};
  \code{\link{as.data.frame.repSparse}},
  \code{\link{as.data.table.repSparse}},
  \code{\link{as.matrix.repSparse}},
  \code{\link{as.repSparse}},
  \code{\link{as.repSparse.dsparseMatrix}},
  \code{\link{as.repSparse.factor}},
  \code{\link{as.repSparse.matrix}},
  \code{\link{as.repSparse.repSparse}},
  \code{\link{repSparse2CsparseMatrix}},
  \code{\link{repSparse2TsparseMatrix}},
  \code{\link{repSparse2dgCMatrix}},
  \code{\link{repSparse2dgTMatrix}},
  \code{\link{repSparse2sparseMatrix}};
  \code{\link{chol.repSparse}},
  \code{\link{chol.repSparseCompSymm}},
  \code{\link{chol.repSparseOneOffDiag}};
  \code{\link{getInit}}, \code{\link{getInit.default}},
  \code{\link{getInit.function}},
  \code{\link{getInit.repSparse}}, \code{\link{setInit}},
  \code{\link{setInit.default}},
  \code{\link{setInit.function}},
  \code{\link{setInit.repSparse}}; \code{\link{ind2point}},
  \code{\link{point2ind}}; \code{\link{kr}},
  \code{\link{kron}}; \code{\link{mkCholCompSymmTrans}},
  \code{\link{mkCholOneOffDiagTrans}},
  \code{\link{mkConstVarCholTrans}},
  \code{\link{mkCorMatCholTrans}},
  \code{\link{mkGenCholInPlaceTrans}},
  \code{\link{mkIdentityTrans}}, \code{\link{mkListTrans}},
  \code{\link{mkOuterTrans}}, \code{\link{mkVarExpTrans}},
  \code{\link{mkVarFixedTrans}},
  \code{\link{mkVarIdentTrans}},
  \code{\link{mkVarPowerTrans}}; \code{\link{rRepSparse}},
  \code{\link{repSparseBlank}},
  \code{\link{repSparseCompSymm}},
  \code{\link{repSparseConstVarChol}},
  \code{\link{repSparseCorMatChol}},
  \code{\link{repSparseDiag}},
  \code{\link{repSparseIdent}},
  \code{\link{repSparseOneOffDiag}},
  \code{\link{repSparseSymm}}, \code{\link{repSparseTri}},
  \code{\link{repSparseVarWithCovariate}};
  \code{\link{repSparse}}
}

