% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repSparse-class}
\alias{dim.repSparse}
\alias{image.repSparse}
\alias{print.repSparse}
\alias{repSparse-class}
\alias{sort.repSparse}
\alias{standardSort}
\alias{t.repSparse}
\alias{update.repSparse}
\title{\code{repSparse} class}
\usage{
print.repSparse(x, ...)

update.repSparse(object, newPars, ...)

image.repSparse(x, ...)

t.repSparse(x)

\method{sort}{repSparse}(x, decreasing = FALSE, type = c("col", "row",
  "val"), ...)

standardSort(x)

\method{dim}{repSparse}(x)
}
\arguments{
\item{x}{\code{repSparse} object}

\item{...}{passed to subsequent functions}

\item{object}{\code{repSparse} object}

\item{newPars}{new parameter values}

\item{decreasing}{see \code{\link{sort}}}

\item{type}{sort by column, row, or value indices?}
}
\description{
\code{repSparse} class
}
\seealso{
Other repSparseTopics: \code{\link{.bind}},
  \code{\link{bind}}, \code{\link{rep.repSparse}};
  \code{\link{as}}, \code{\link{as}}, \code{\link{as}},
  \code{\link{as}}, \code{\link{as}},
  \code{\link{as.data.frame.repSparse}},
  \code{\link{as.data.table.repSparse}},
  \code{\link{as.matrix.repSparse}},
  \code{\link{as.repSparse}},
  \code{\link{as.repSparse.dsparseMatrix}},
  \code{\link{as.repSparse.factor}},
  \code{\link{as.repSparse.matrix}},
  \code{\link{as.repSparse.repSparse}};
  \code{\link{chol.repSparse}},
  \code{\link{chol.repSparseCompSymm}},
  \code{\link{chol.repSparseOneOffDiag}};
  \code{\link{getInit}}, \code{\link{getInit.default}},
  \code{\link{getInit.function}},
  \code{\link{getInit.repSparse}}, \code{\link{setInit}},
  \code{\link{setInit.default}},
  \code{\link{setInit.function}},
  \code{\link{setInit.repSparse}}; \code{\link{ind2point}},
  \code{\link{point2ind}}; \code{\link{kr}},
  \code{\link{kr0}}, \code{\link{kron}},
  \code{\link{mmult}}, \code{\link{rowWiseCombination}};
  \code{\link{mkCholOneOffDiagTrans}},
  \code{\link{mkGenCholInPlaceTrans}},
  \code{\link{mkIdentityTrans}}, \code{\link{mkListTrans}},
  \code{\link{mkOuterTrans}}; \code{\link{rRepSparse}},
  \code{\link{repSparseCompSymm}},
  \code{\link{repSparseDiag}},
  \code{\link{repSparseIdent}},
  \code{\link{repSparseOneOffDiag}},
  \code{\link{repSparseTri}}; \code{\link{repSparse}}

Other repSparseTopics: \code{\link{.bind}},
  \code{\link{bind}}, \code{\link{rep.repSparse}};
  \code{\link{as}}, \code{\link{as}}, \code{\link{as}},
  \code{\link{as}}, \code{\link{as}},
  \code{\link{as.data.frame.repSparse}},
  \code{\link{as.data.table.repSparse}},
  \code{\link{as.matrix.repSparse}},
  \code{\link{as.repSparse}},
  \code{\link{as.repSparse.dsparseMatrix}},
  \code{\link{as.repSparse.factor}},
  \code{\link{as.repSparse.matrix}},
  \code{\link{as.repSparse.repSparse}};
  \code{\link{chol.repSparse}},
  \code{\link{chol.repSparseCompSymm}},
  \code{\link{chol.repSparseOneOffDiag}};
  \code{\link{getInit}}, \code{\link{getInit.default}},
  \code{\link{getInit.function}},
  \code{\link{getInit.repSparse}}, \code{\link{setInit}},
  \code{\link{setInit.default}},
  \code{\link{setInit.function}},
  \code{\link{setInit.repSparse}}; \code{\link{ind2point}},
  \code{\link{point2ind}}; \code{\link{kr}},
  \code{\link{kr0}}, \code{\link{kron}},
  \code{\link{mmult}}, \code{\link{rowWiseCombination}};
  \code{\link{mkCholOneOffDiagTrans}},
  \code{\link{mkGenCholInPlaceTrans}},
  \code{\link{mkIdentityTrans}}, \code{\link{mkListTrans}},
  \code{\link{mkOuterTrans}}; \code{\link{rRepSparse}},
  \code{\link{repSparseCompSymm}},
  \code{\link{repSparseDiag}},
  \code{\link{repSparseIdent}},
  \code{\link{repSparseOneOffDiag}},
  \code{\link{repSparseTri}}; \code{\link{repSparse}}

Other repSparseTopics: \code{\link{.bind}},
  \code{\link{bind}}, \code{\link{rep.repSparse}};
  \code{\link{as}}, \code{\link{as}}, \code{\link{as}},
  \code{\link{as}}, \code{\link{as}},
  \code{\link{as.data.frame.repSparse}},
  \code{\link{as.data.table.repSparse}},
  \code{\link{as.matrix.repSparse}},
  \code{\link{as.repSparse}},
  \code{\link{as.repSparse.dsparseMatrix}},
  \code{\link{as.repSparse.factor}},
  \code{\link{as.repSparse.matrix}},
  \code{\link{as.repSparse.repSparse}};
  \code{\link{chol.repSparse}},
  \code{\link{chol.repSparseCompSymm}},
  \code{\link{chol.repSparseOneOffDiag}};
  \code{\link{getInit}}, \code{\link{getInit.default}},
  \code{\link{getInit.function}},
  \code{\link{getInit.repSparse}}, \code{\link{setInit}},
  \code{\link{setInit.default}},
  \code{\link{setInit.function}},
  \code{\link{setInit.repSparse}}; \code{\link{ind2point}},
  \code{\link{point2ind}}; \code{\link{kr}},
  \code{\link{kr0}}, \code{\link{kron}},
  \code{\link{mmult}}, \code{\link{rowWiseCombination}};
  \code{\link{mkCholOneOffDiagTrans}},
  \code{\link{mkGenCholInPlaceTrans}},
  \code{\link{mkIdentityTrans}}, \code{\link{mkListTrans}},
  \code{\link{mkOuterTrans}}; \code{\link{rRepSparse}},
  \code{\link{repSparseCompSymm}},
  \code{\link{repSparseDiag}},
  \code{\link{repSparseIdent}},
  \code{\link{repSparseOneOffDiag}},
  \code{\link{repSparseTri}}; \code{\link{repSparse}}

Other repSparseTopics: \code{\link{.bind}},
  \code{\link{bind}}, \code{\link{rep.repSparse}};
  \code{\link{as}}, \code{\link{as}}, \code{\link{as}},
  \code{\link{as}}, \code{\link{as}},
  \code{\link{as.data.frame.repSparse}},
  \code{\link{as.data.table.repSparse}},
  \code{\link{as.matrix.repSparse}},
  \code{\link{as.repSparse}},
  \code{\link{as.repSparse.dsparseMatrix}},
  \code{\link{as.repSparse.factor}},
  \code{\link{as.repSparse.matrix}},
  \code{\link{as.repSparse.repSparse}};
  \code{\link{chol.repSparse}},
  \code{\link{chol.repSparseCompSymm}},
  \code{\link{chol.repSparseOneOffDiag}};
  \code{\link{getInit}}, \code{\link{getInit.default}},
  \code{\link{getInit.function}},
  \code{\link{getInit.repSparse}}, \code{\link{setInit}},
  \code{\link{setInit.default}},
  \code{\link{setInit.function}},
  \code{\link{setInit.repSparse}}; \code{\link{ind2point}},
  \code{\link{point2ind}}; \code{\link{kr}},
  \code{\link{kr0}}, \code{\link{kron}},
  \code{\link{mmult}}, \code{\link{rowWiseCombination}};
  \code{\link{mkCholOneOffDiagTrans}},
  \code{\link{mkGenCholInPlaceTrans}},
  \code{\link{mkIdentityTrans}}, \code{\link{mkListTrans}},
  \code{\link{mkOuterTrans}}; \code{\link{rRepSparse}},
  \code{\link{repSparseCompSymm}},
  \code{\link{repSparseDiag}},
  \code{\link{repSparseIdent}},
  \code{\link{repSparseOneOffDiag}},
  \code{\link{repSparseTri}}; \code{\link{repSparse}}

Other repSparseTopics: \code{\link{.bind}},
  \code{\link{bind}}, \code{\link{rep.repSparse}};
  \code{\link{as}}, \code{\link{as}}, \code{\link{as}},
  \code{\link{as}}, \code{\link{as}},
  \code{\link{as.data.frame.repSparse}},
  \code{\link{as.data.table.repSparse}},
  \code{\link{as.matrix.repSparse}},
  \code{\link{as.repSparse}},
  \code{\link{as.repSparse.dsparseMatrix}},
  \code{\link{as.repSparse.factor}},
  \code{\link{as.repSparse.matrix}},
  \code{\link{as.repSparse.repSparse}};
  \code{\link{chol.repSparse}},
  \code{\link{chol.repSparseCompSymm}},
  \code{\link{chol.repSparseOneOffDiag}};
  \code{\link{getInit}}, \code{\link{getInit.default}},
  \code{\link{getInit.function}},
  \code{\link{getInit.repSparse}}, \code{\link{setInit}},
  \code{\link{setInit.default}},
  \code{\link{setInit.function}},
  \code{\link{setInit.repSparse}}; \code{\link{ind2point}},
  \code{\link{point2ind}}; \code{\link{kr}},
  \code{\link{kr0}}, \code{\link{kron}},
  \code{\link{mmult}}, \code{\link{rowWiseCombination}};
  \code{\link{mkCholOneOffDiagTrans}},
  \code{\link{mkGenCholInPlaceTrans}},
  \code{\link{mkIdentityTrans}}, \code{\link{mkListTrans}},
  \code{\link{mkOuterTrans}}; \code{\link{rRepSparse}},
  \code{\link{repSparseCompSymm}},
  \code{\link{repSparseDiag}},
  \code{\link{repSparseIdent}},
  \code{\link{repSparseOneOffDiag}},
  \code{\link{repSparseTri}}; \code{\link{repSparse}}

Other repSparseTopics: \code{\link{.bind}},
  \code{\link{bind}}, \code{\link{rep.repSparse}};
  \code{\link{as}}, \code{\link{as}}, \code{\link{as}},
  \code{\link{as}}, \code{\link{as}},
  \code{\link{as.data.frame.repSparse}},
  \code{\link{as.data.table.repSparse}},
  \code{\link{as.matrix.repSparse}},
  \code{\link{as.repSparse}},
  \code{\link{as.repSparse.dsparseMatrix}},
  \code{\link{as.repSparse.factor}},
  \code{\link{as.repSparse.matrix}},
  \code{\link{as.repSparse.repSparse}};
  \code{\link{chol.repSparse}},
  \code{\link{chol.repSparseCompSymm}},
  \code{\link{chol.repSparseOneOffDiag}};
  \code{\link{getInit}}, \code{\link{getInit.default}},
  \code{\link{getInit.function}},
  \code{\link{getInit.repSparse}}, \code{\link{setInit}},
  \code{\link{setInit.default}},
  \code{\link{setInit.function}},
  \code{\link{setInit.repSparse}}; \code{\link{ind2point}},
  \code{\link{point2ind}}; \code{\link{kr}},
  \code{\link{kr0}}, \code{\link{kron}},
  \code{\link{mmult}}, \code{\link{rowWiseCombination}};
  \code{\link{mkCholOneOffDiagTrans}},
  \code{\link{mkGenCholInPlaceTrans}},
  \code{\link{mkIdentityTrans}}, \code{\link{mkListTrans}},
  \code{\link{mkOuterTrans}}; \code{\link{rRepSparse}},
  \code{\link{repSparseCompSymm}},
  \code{\link{repSparseDiag}},
  \code{\link{repSparseIdent}},
  \code{\link{repSparseOneOffDiag}},
  \code{\link{repSparseTri}}; \code{\link{repSparse}}

Other repSparseTopics: \code{\link{.bind}},
  \code{\link{bind}}, \code{\link{rep.repSparse}};
  \code{\link{as}}, \code{\link{as}}, \code{\link{as}},
  \code{\link{as}}, \code{\link{as}},
  \code{\link{as.data.frame.repSparse}},
  \code{\link{as.data.table.repSparse}},
  \code{\link{as.matrix.repSparse}},
  \code{\link{as.repSparse}},
  \code{\link{as.repSparse.dsparseMatrix}},
  \code{\link{as.repSparse.factor}},
  \code{\link{as.repSparse.matrix}},
  \code{\link{as.repSparse.repSparse}};
  \code{\link{chol.repSparse}},
  \code{\link{chol.repSparseCompSymm}},
  \code{\link{chol.repSparseOneOffDiag}};
  \code{\link{getInit}}, \code{\link{getInit.default}},
  \code{\link{getInit.function}},
  \code{\link{getInit.repSparse}}, \code{\link{setInit}},
  \code{\link{setInit.default}},
  \code{\link{setInit.function}},
  \code{\link{setInit.repSparse}}; \code{\link{ind2point}},
  \code{\link{point2ind}}; \code{\link{kr}},
  \code{\link{kr0}}, \code{\link{kron}},
  \code{\link{mmult}}, \code{\link{rowWiseCombination}};
  \code{\link{mkCholOneOffDiagTrans}},
  \code{\link{mkGenCholInPlaceTrans}},
  \code{\link{mkIdentityTrans}}, \code{\link{mkListTrans}},
  \code{\link{mkOuterTrans}}; \code{\link{rRepSparse}},
  \code{\link{repSparseCompSymm}},
  \code{\link{repSparseDiag}},
  \code{\link{repSparseIdent}},
  \code{\link{repSparseOneOffDiag}},
  \code{\link{repSparseTri}}; \code{\link{repSparse}}

Other repSparseTopics: \code{\link{.bind}},
  \code{\link{bind}}, \code{\link{rep.repSparse}};
  \code{\link{as}}, \code{\link{as}}, \code{\link{as}},
  \code{\link{as}}, \code{\link{as}},
  \code{\link{as.data.frame.repSparse}},
  \code{\link{as.data.table.repSparse}},
  \code{\link{as.matrix.repSparse}},
  \code{\link{as.repSparse}},
  \code{\link{as.repSparse.dsparseMatrix}},
  \code{\link{as.repSparse.factor}},
  \code{\link{as.repSparse.matrix}},
  \code{\link{as.repSparse.repSparse}};
  \code{\link{chol.repSparse}},
  \code{\link{chol.repSparseCompSymm}},
  \code{\link{chol.repSparseOneOffDiag}};
  \code{\link{getInit}}, \code{\link{getInit.default}},
  \code{\link{getInit.function}},
  \code{\link{getInit.repSparse}}, \code{\link{setInit}},
  \code{\link{setInit.default}},
  \code{\link{setInit.function}},
  \code{\link{setInit.repSparse}}; \code{\link{ind2point}},
  \code{\link{point2ind}}; \code{\link{kr}},
  \code{\link{kr0}}, \code{\link{kron}},
  \code{\link{mmult}}, \code{\link{rowWiseCombination}};
  \code{\link{mkCholOneOffDiagTrans}},
  \code{\link{mkGenCholInPlaceTrans}},
  \code{\link{mkIdentityTrans}}, \code{\link{mkListTrans}},
  \code{\link{mkOuterTrans}}; \code{\link{rRepSparse}},
  \code{\link{repSparseCompSymm}},
  \code{\link{repSparseDiag}},
  \code{\link{repSparseIdent}},
  \code{\link{repSparseOneOffDiag}},
  \code{\link{repSparseTri}}; \code{\link{repSparse}}

Other repSparseTopics: \code{\link{.bind}},
  \code{\link{bind}}, \code{\link{rep.repSparse}};
  \code{\link{as}}, \code{\link{as}}, \code{\link{as}},
  \code{\link{as}}, \code{\link{as}},
  \code{\link{as.data.frame.repSparse}},
  \code{\link{as.data.table.repSparse}},
  \code{\link{as.matrix.repSparse}},
  \code{\link{as.repSparse}},
  \code{\link{as.repSparse.dsparseMatrix}},
  \code{\link{as.repSparse.factor}},
  \code{\link{as.repSparse.matrix}},
  \code{\link{as.repSparse.repSparse}};
  \code{\link{chol.repSparse}},
  \code{\link{chol.repSparseCompSymm}},
  \code{\link{chol.repSparseOneOffDiag}};
  \code{\link{getInit}}, \code{\link{getInit.default}},
  \code{\link{getInit.function}},
  \code{\link{getInit.repSparse}}, \code{\link{setInit}},
  \code{\link{setInit.default}},
  \code{\link{setInit.function}},
  \code{\link{setInit.repSparse}}; \code{\link{ind2point}},
  \code{\link{point2ind}}; \code{\link{kr}},
  \code{\link{kr0}}, \code{\link{kron}},
  \code{\link{mmult}}, \code{\link{rowWiseCombination}};
  \code{\link{mkCholOneOffDiagTrans}},
  \code{\link{mkGenCholInPlaceTrans}},
  \code{\link{mkIdentityTrans}}, \code{\link{mkListTrans}},
  \code{\link{mkOuterTrans}}; \code{\link{rRepSparse}},
  \code{\link{repSparseCompSymm}},
  \code{\link{repSparseDiag}},
  \code{\link{repSparseIdent}},
  \code{\link{repSparseOneOffDiag}},
  \code{\link{repSparseTri}}; \code{\link{repSparse}}

Other repSparseTopics: \code{\link{.bind}},
  \code{\link{bind}}, \code{\link{rep.repSparse}};
  \code{\link{as}}, \code{\link{as}}, \code{\link{as}},
  \code{\link{as}}, \code{\link{as}},
  \code{\link{as.data.frame.repSparse}},
  \code{\link{as.data.table.repSparse}},
  \code{\link{as.matrix.repSparse}},
  \code{\link{as.repSparse}},
  \code{\link{as.repSparse.dsparseMatrix}},
  \code{\link{as.repSparse.factor}},
  \code{\link{as.repSparse.matrix}},
  \code{\link{as.repSparse.repSparse}};
  \code{\link{chol.repSparse}},
  \code{\link{chol.repSparseCompSymm}},
  \code{\link{chol.repSparseOneOffDiag}};
  \code{\link{getInit}}, \code{\link{getInit.default}},
  \code{\link{getInit.function}},
  \code{\link{getInit.repSparse}}, \code{\link{setInit}},
  \code{\link{setInit.default}},
  \code{\link{setInit.function}},
  \code{\link{setInit.repSparse}}; \code{\link{ind2point}},
  \code{\link{point2ind}}; \code{\link{kr}},
  \code{\link{kr0}}, \code{\link{kron}},
  \code{\link{mmult}}, \code{\link{rowWiseCombination}};
  \code{\link{mkCholOneOffDiagTrans}},
  \code{\link{mkGenCholInPlaceTrans}},
  \code{\link{mkIdentityTrans}}, \code{\link{mkListTrans}},
  \code{\link{mkOuterTrans}}; \code{\link{rRepSparse}},
  \code{\link{repSparseCompSymm}},
  \code{\link{repSparseDiag}},
  \code{\link{repSparseIdent}},
  \code{\link{repSparseOneOffDiag}},
  \code{\link{repSparseTri}}; \code{\link{repSparse}}
}

