% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repSparse}
\alias{dim.repSparse}
\alias{image.repSparse}
\alias{print.repSparse}
\alias{repSparse}
\alias{repSparse-class}
\alias{sort.repSparse}
\alias{update.repSparse}
\title{Sparse matrix with repeated values}
\usage{
repSparse(rowInds, colInds, valInds, vals, trans, Dim)

print.repSparse(x, ...)

update.repSparse(object, newPars, ...)

\method{sort}{repSparse}(x, decreasing = FALSE, type = c("col", "row",
  "val"), ...)

\method{dim}{repSparse}(x)

image.repSparse(x, ...)
}
\arguments{
\item{rowInds}{1-based row indices (converted to 0-based in
output)}

\item{colInds}{1-based column indices (converted to 0-based in
output)}

\item{valInds}{indices for the nonzero values}

\item{vals}{nonzero values}

\item{trans}{function for transforming some parameters into
\code{vals}.  the \code{\link{environment}} of this function must
contain a named list of these parameters (called \code{matPars}),
which are to be taken as the arguments to \code{trans}.
\code{update} will update these parameters (in the future).}

\item{Dim}{matrix dimensions}

\item{newPars}{new parameter values}

\item{decreasing}{see \code{\link{sort}}}

\item{type}{sort by column, row, or value indices?}
}
\value{
object of class \code{repSparse} with list elements
\code{rowInds}, \code{colInds}, \code{valInds}, \code{vals}, and
\code{trans}, and a \code{Dim} attibute
}
\description{
Sparse matrix with repeated values

\code{repSparse} class
}

