% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repSparse}
\alias{as.data.frame.repSparse}
\alias{as.matrix.repSparse}
\alias{dim.repSparse}
\alias{image.repSparse}
\alias{print.repSparse}
\alias{repSparse}
\alias{repSparse2Sparse}
\alias{sparse2RepSparse}
\title{Sparse matrix with repeated values}
\usage{
repSparse(rowInds, colInds, valInds, vals, Dim)

print.repSparse(x, ...)

\method{as.data.frame}{repSparse}(x, ...)

as.matrix.repSparse(x, sparse = FALSE, ...)

\method{dim}{repSparse}(x)

repSparse2Sparse(object, ...)

sparse2RepSparse(object, ...)

image.repSparse(x, ...)
}
\arguments{
\item{rowInds}{1-based row indices (converted to 0-based in
output)}

\item{colInds}{1-based column indices (converted to 0-based in
output)}

\item{valInds}{indices for the nonzero values}

\item{vals}{nonzero values}

\item{Dim}{matrix dimensions}

\item{sparse}{return \code{sparseMatrix}?}

\item{object}{\code{repSparse} object}
}
\value{
object of class \code{repSparse} with list elements
\code{rowInds}, \code{colInds}, \code{valInds}, \code{vals}, and
\code{Dim} attibute

results of \code{sparseMatrix}
}
\description{
Sparse matrix with repeated values
}

