% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repSparse}
\alias{repSparse}
\title{Repeated sparse matrices}
\usage{
repSparse(rowInds, colInds, valInds, vals, trans, Dim, sortFun = standardSort)
}
\arguments{
\item{rowInds}{1-based row indices (converted to 0-based in
output)}

\item{colInds}{1-based column indices (converted to 0-based in
output)}

\item{valInds}{indices for the nonzero values}

\item{vals}{nonzero values}

\item{trans}{function for transforming some parameters into
\code{vals}.  the \code{\link{environment}} of this function
should contain a vector of these parameters (called \code{init}),
which could be taken as the arguments to \code{trans}.  these
initial values can be obtained by \code{\link{getInit}} and set by
\code{\link{setInit}}.  an \code{\link{update.repSparse}} method
will update these parameters.}

\item{Dim}{matrix dimensions}

\item{sortFun}{a function with which to sort the indices of the
resulting matrix.  please note that by default the
\code{standardSort} function is used, which provides a convenient
ordering for computing Khatri-Rao products (with
\code{\link{kr}}).  see details.}
}
\value{
A member of the \code{\link{repSparse-class}}.
}
\description{
A sparse matrix class for matrices whose elements come from a set
of a relatively small number of values or computable from a
relatively small parameter vector.
}
\details{
These parameterized matrix objects are useful for updating
matrices at each iteration of a general nonlinear optimizer.  This
\code{\link{repSparse-class}} is designed to work well with the
\code{\link{Matrix}} package, in that they can be quickly coerced
to \code{\link{sparseMatrix}} objects.  The
\code{\link{mkSparseTrans}} function can be used to automatically
generate a function for updating the values of the
\code{sparseMatrix} object from a relatively small number of
parameters or repeated values.

Such matrices are best constructed by \code{\link{bind}}ing,
\code{\link{kron}}ing, and \code{\link{kr}}ing together a series
of simpler \code{repSparse} matrices, which can be constructed
with the \code{repSparse} function or a function for constructing
special \code{repSparse} objects
(e.g. \code{\link{repSparseDiag}}).

Note that the ordering of the indices supplied to \code{rowInds},
\code{colInds}, and \code{valInds} may not appear in the same
order in the output.  This is because \code{sortFun} is used to
process the order, which by default is \code{\link{standardSort}}.
To suppress this behaviour a different \code{sortFun} may be
supplied, but this is generally not recommended because certain
functions assume certain orderings (e.g. \code{\link{kr}}).
}
\examples{
repSparse(1:4, 4:1, rep(1:2, 2), c(-pi, pi))
}
\seealso{
Other repeated.sparse.matrix.topics: \code{\link{.bind}},
  \code{\link{bind}}, \code{\link{rep.repSparse}};
  \code{\link{as.data.frame.repSparse}},
  \code{\link{as.data.table.repSparse}},
  \code{\link{as.matrix.repSparse}},
  \code{\link{as.repSparse}},
  \code{\link{as.repSparse.dCHMsimpl}},
  \code{\link{as.repSparse.dist}},
  \code{\link{as.repSparse.dsparseMatrix}},
  \code{\link{as.repSparse.factor}},
  \code{\link{as.repSparse.matrix}},
  \code{\link{as.repSparse.repSparse}},
  \code{\link{repSparse2CsparseMatrix}},
  \code{\link{repSparse2TsparseMatrix}},
  \code{\link{repSparse2dgCMatrix}},
  \code{\link{repSparse2dgTMatrix}},
  \code{\link{repSparse2sparseMatrix}};
  \code{\link{chol.repSparse}},
  \code{\link{chol.repSparseCompSymm}},
  \code{\link{chol.repSparseOneOffDiag}};
  \code{\link{diag.repSparse}},
  \code{\link{dim.repSparse}},
  \code{\link{image.repSparse}},
  \code{\link{plot.repSparse}},
  \code{\link{print.repSparse}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}}, \code{\link{t.repSparse}},
  \code{\link{update.repSparse}}; \code{\link{getInit}},
  \code{\link{getInit.default}},
  \code{\link{getInit.function}},
  \code{\link{getInit.repSparse}}, \code{\link{setInit}},
  \code{\link{setInit.default}},
  \code{\link{setInit.function}},
  \code{\link{setInit.repSparse}}; \code{\link{ind2point}},
  \code{\link{point2ind}}; \code{\link{kr}},
  \code{\link{kron}}; \code{\link{mkCholCompSymmTrans}},
  \code{\link{mkCholOneOffDiagTrans}},
  \code{\link{mkConstVarCholTrans}},
  \code{\link{mkCorMatCholTrans}},
  \code{\link{mkExpCholTrans}},
  \code{\link{mkFlexObsLevelTrans}},
  \code{\link{mkGenCholInPlaceTrans}},
  \code{\link{mkIdentityTrans}}, \code{\link{mkListTrans}},
  \code{\link{mkOuterTrans}}, \code{\link{mkTrans}},
  \code{\link{mkVarExpTrans}},
  \code{\link{mkVarFixedTrans}},
  \code{\link{mkVarIdentTrans}},
  \code{\link{mkVarPowerTrans}};
  \code{\link{repSparseBlank}}, \code{\link{repSparseCol}},
  \code{\link{repSparseCompSymm}},
  \code{\link{repSparseConstVarChol}},
  \code{\link{repSparseCorMatChol}},
  \code{\link{repSparseDiag}},
  \code{\link{repSparseExpChol}},
  \code{\link{repSparseIdent}}, \code{\link{repSparseInd}},
  \code{\link{repSparseOneOffDiag}},
  \code{\link{repSparseOnes}},
  \code{\link{repSparseSpecial}},
  \code{\link{repSparseSymm}}, \code{\link{repSparseTri}},
  \code{\link{repSparseVarWithCovariate}}
}

