% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repSparse}
\alias{repSparse}
\title{Sparse matrix with repeated values}
\usage{
repSparse(rowInds, colInds, valInds, vals, trans, Dim)
}
\arguments{
\item{rowInds}{1-based row indices (converted to 0-based in
output)}

\item{colInds}{1-based column indices (converted to 0-based in
output)}

\item{valInds}{indices for the nonzero values}

\item{vals}{nonzero values}

\item{trans}{function for transforming some parameters into
\code{vals}.  the \code{\link{environment}} of this function
should contain a vector of these parameters (called \code{init}),
which could be taken as the arguments to \code{trans}.  these
initial values can be obtained by \code{\link{getInit}} and set by
\code{\link{setInit}}.  an \code{\link{update.repSparse}} method
will update these parameters.}

\item{Dim}{matrix dimensions}
}
\value{
object of class \code{repSparse} with list elements
\code{rowInds}, \code{colInds}, \code{valInds}, \code{vals}, and
\code{trans}, and a \code{Dim} attibute
}
\description{
Sparse matrix with repeated values
}
\seealso{
Other repSparseTopics: \code{\link{.bind}},
  \code{\link{bind}}, \code{\link{rep.repSparse}};
  \code{\link{as}}, \code{\link{as}}, \code{\link{as}},
  \code{\link{as}}, \code{\link{as}},
  \code{\link{as.data.frame.repSparse}},
  \code{\link{as.data.table.repSparse}},
  \code{\link{as.matrix.repSparse}},
  \code{\link{as.repSparse}},
  \code{\link{as.repSparse.dsparseMatrix}},
  \code{\link{as.repSparse.factor}},
  \code{\link{as.repSparse.matrix}},
  \code{\link{as.repSparse.repSparse}};
  \code{\link{chol.repSparse}},
  \code{\link{chol.repSparseCompSymm}},
  \code{\link{chol.repSparseOneOffDiag}};
  \code{\link{dim.repSparse}},
  \code{\link{image.repSparse}},
  \code{\link{print.repSparse}},
  \code{\link{repSparse-class}},
  \code{\link{sort.repSparse}}, \code{\link{t.repSparse}},
  \code{\link{update.repSparse}}; \code{\link{getInit}},
  \code{\link{getInit.default}},
  \code{\link{getInit.function}},
  \code{\link{getInit.repSparse}}, \code{\link{setInit}},
  \code{\link{setInit.default}},
  \code{\link{setInit.function}},
  \code{\link{setInit.repSparse}}; \code{\link{ind2point}},
  \code{\link{point2ind}}; \code{\link{kr}},
  \code{\link{kron}}, \code{\link{mmult}},
  \code{\link{rowWiseCombination}};
  \code{\link{mkCholOneOffDiagTrans}},
  \code{\link{mkIdentityTrans}}, \code{\link{mkListTrans}},
  \code{\link{mkOuterTrans}}; \code{\link{rRepSparse}},
  \code{\link{repSparseCompSymm}},
  \code{\link{repSparseDiag}},
  \code{\link{repSparseIdent}},
  \code{\link{repSparseOneOffDiag}}
}

