% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repSparseCompSymm}
\alias{repSparseCompSymm}
\title{Repeated sparse matrix with compound symmetry}
\usage{
repSparseCompSymm(diagVal, offDiagVal, matSize)
}
\arguments{
\item{diagVal}{value for the diagonal}

\item{offDiagVal}{value for the off-diagonal}

\item{matSize}{size of the resulting matrix}
}
\description{
Repeated sparse matrix with compound symmetry
}
\examples{
(xCompSymm <- repSparseCompSymm(1, -0.2, 5))
}
\seealso{
Other repSparseSpecial: \code{\link{repSparseBlank}};
  \code{\link{repSparseCol}};
  \code{\link{repSparseConstVarChol}};
  \code{\link{repSparseCorFactor}};
  \code{\link{repSparseCorMatChol}};
  \code{\link{repSparseDiag}};
  \code{\link{repSparseExpChol}};
  \code{\link{repSparseFull}};
  \code{\link{repSparseGenFullTri}};
  \code{\link{repSparseIdent}}; \code{\link{repSparseInd}};
  \code{\link{repSparseOneOffDiag}};
  \code{\link{repSparseOnes}}; \code{\link{repSparseSymm}};
  \code{\link{repSparseTri}};
  \code{\link{repSparseVarWithCovariate}}
}

