% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repSparseDiag}
\alias{repSparseDiag}
\title{Diagonal repeated sparse matrix}
\usage{
repSparseDiag(vals, valInds)
}
\arguments{
\item{vals}{vector of values}

\item{valInds}{vector of value indices}
}
\description{
Diagonal repeated sparse matrix
}
\examples{
set.seed(1)
(xDiag <- repSparseDiag(rnorm(5)))
}
\seealso{
Other repSparseSpecial: \code{\link{repSparseBlank}};
  \code{\link{repSparseCol}};
  \code{\link{repSparseCompSymm}};
  \code{\link{repSparseConstVarChol}};
  \code{\link{repSparseCorMatChol}};
  \code{\link{repSparseExpChol}};
  \code{\link{repSparseIdent}}; \code{\link{repSparseInd}};
  \code{\link{repSparseOneOffDiag}};
  \code{\link{repSparseOnes}}; \code{\link{repSparseSymm}};
  \code{\link{repSparseTri}};
  \code{\link{repSparseVarWithCovariate}}
}

