% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repSparseOneOffDiag}
\alias{repSparseOneOffDiag}
\title{Repeated sparse matrix with only one non-zero value off the
diagonal}
\usage{
repSparseOneOffDiag(diagVal, offDiagVal, offDiagInds, matSize)
}
\arguments{
\item{diagVal}{unique value for the diagonal}

\item{offDiagVal}{unique value for the off-diagonal}

\item{offDiagInds}{indices for the two correlated objects}

\item{matSize}{size of the matrix}
}
\description{
Repeated sparse matrix with only one non-zero value off the
diagonal
}
\examples{
(xOneOffDiag <- repSparseOneOffDiag(1, -0.2, c(5, 2), 5))
}
\seealso{
Other repSparseSpecial: \code{\link{repSparseBlank}};
  \code{\link{repSparseCol}};
  \code{\link{repSparseCompSymm}};
  \code{\link{repSparseConstVarChol}};
  \code{\link{repSparseCorFactor}};
  \code{\link{repSparseCorMatChol}};
  \code{\link{repSparseDiag}};
  \code{\link{repSparseExpChol}};
  \code{\link{repSparseFull}};
  \code{\link{repSparseIdent}}; \code{\link{repSparseInd}};
  \code{\link{repSparseOnes}}; \code{\link{repSparseSymm}};
  \code{\link{repSparseTri}};
  \code{\link{repSparseVarWithCovariate}}
}

