% Generated by roxygen2 (4.0.1): do not edit by hand
\name{resetTransConst}
\alias{resetTransConst}
\title{Constant repeated sparse matrix}
\usage{
resetTransConst(object)
}
\arguments{
\item{object}{repeated sparse matrix}
}
\description{
Convert a parameterized repeated sparse matrix into a constant
repeated sparse matrix (i.e. \code{getInit(object)} has length
zero).
}
\seealso{
Other modifications: \code{\link{scalarMult}};
  \code{\link{simplifyRepSparse}}
}

