% Generated by roxygen2 (4.0.1): do not edit by hand
\name{scalarMult}
\alias{scalarMult}
\title{Add scalar multiple to a repeated sparse matrix}
\usage{
scalarMult(object, mult)
}
\arguments{
\item{object}{repeated sparse matrix}

\item{mult}{initial value for the multiplier parameter}
}
\description{
Adjust the \code{trans} function of \code{object} such that a
scalar multiplier parameter is concatenated at the beginning of
the parameter vector.
}
\seealso{
Other modifications: \code{\link{resetTransConst}};
  \code{\link{simplifyRepSparse}}
}

