% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setReTrm.expDecay}
\alias{expDecay}
\alias{setReTrm.expDecay}
\title{Random effects term with exponential decay in distance-based covariance}
\usage{
\method{setReTrm}{expDecay}(object, addArgsList, devfunEnv = NULL)
}
\arguments{
\item{object}{a \code{reTrmStruct} object}

\item{addArgsList}{a list of named quantities within which
\code{addArgsExpr} is evaluated}

\item{devfunEnv}{optional environment of the deviance function}
}
\description{
Random effects term with exponential decay in distance-based covariance
}
\section{Mixed model formula usage}{

\code{expDecay(1 | grpFac, distCutoff, minCov, distmat)}
\describe{
\item{grpFac}{grouping factor (e.g. with levels given by geographical sites)} \item{distCutoff}{maximum distance with covariance greater than minCov} \item{minCov}{minimum covariance} \item{distmat}{distance matrix object over the levels of grpFac}
}
}
\seealso{
Other setReTrm: \code{\link{cooccur}},
  \code{\link{setReTrm.cooccur}}; \code{\link{edge}},
  \code{\link{setReTrm.edge}}; \code{\link{factAnal}},
  \code{\link{setReTrm.factAnal}};
  \code{\link{findReTrmClasses}}, \code{\link{getAddArgs}},
  \code{\link{packReTrm}}, \code{\link{setReTrm}},
  \code{\link{setReTrm.default}}; \code{\link{flexvar}},
  \code{\link{setReTrm.flexvar}}; \code{\link{identity}},
  \code{\link{setReTrm.identity}};
  \code{\link{nlmeCorStruct}},
  \code{\link{setReTrm.nlmeCorStruct}};
  \code{\link{obslev}}, \code{\link{setReTrm.obslev}};
  \code{\link{setReTrm.varExp}}, \code{\link{varExp}};
  \code{\link{setReTrm.varIdent}}, \code{\link{varIdent}}
}

