% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setReTrm.factAnal}
\alias{factAnal}
\alias{setReTrm.factAnal}
\title{Random effects term for factor analysis}
\usage{
\method{setReTrm}{factAnal}(object, addArgsList, devfunEnv = NULL)
}
\arguments{
\item{object}{a \code{reTrmStruct} object}

\item{addArgsList}{a list of named quantities within which
\code{addArgsExpr} is evaluated}

\item{devfunEnv}{optional environment of the deviance function}
}
\description{
Random effects term for factor analysis
}
\section{Mixed model formula usage}{

\code{factAnal(1 | grpFac, nAxes, seed)}
\describe{
\item{grpFac}{grouping factor for loadings} \item{nAxes}{number of axes} \item{seed}{random seed for initial axis values}
}
}
\seealso{
Other setReTrm: \code{\link{cooccur}},
  \code{\link{setReTrm.cooccur}}; \code{\link{edge}},
  \code{\link{setReTrm.edge}}; \code{\link{expDecay}},
  \code{\link{setReTrm.expDecay}};
  \code{\link{findReTrmClasses}}, \code{\link{getAddArgs}},
  \code{\link{packReTrm}}, \code{\link{setReTrm}},
  \code{\link{setReTrm.default}}; \code{\link{flexvar}},
  \code{\link{setReTrm.flexvar}}; \code{\link{identity}},
  \code{\link{setReTrm.identity}};
  \code{\link{nlmeCorStruct}},
  \code{\link{setReTrm.nlmeCorStruct}};
  \code{\link{obslev}}, \code{\link{setReTrm.obslev}};
  \code{\link{setReTrm.varExp}}, \code{\link{varExp}};
  \code{\link{setReTrm.varIdent}}, \code{\link{varIdent}}
}

