% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setReTrm.identity}
\alias{identity}
\alias{setReTrm.identity}
\title{Random effects term for covariance proportional to identity matrix}
\usage{
\method{setReTrm}{identity}(object, addArgsList, devfunEnv = NULL)
}
\arguments{
\item{object}{a \code{reTrmStruct} object}

\item{addArgsList}{a list of named quantities within which
\code{addArgsExpr} is evaluated}

\item{devfunEnv}{optional environment of the deviance function}
}
\description{
Random effects term for covariance proportional to identity matrix
}
\section{Mixed model formula usage}{

\code{identity(linForm | grpFac)}
\describe{
\item{linForm}{linear model formula decribing effects} \item{grpFac}{grouping factor}
}
}
\seealso{
Other setReTrm: \code{\link{cooccur}},
  \code{\link{setReTrm.cooccur}}; \code{\link{edge}},
  \code{\link{setReTrm.edge}}; \code{\link{expDecay}},
  \code{\link{setReTrm.expDecay}}; \code{\link{factAnal}},
  \code{\link{setReTrm.factAnal}};
  \code{\link{findReTrmClasses}}, \code{\link{getAddArgs}},
  \code{\link{packReTrm}}, \code{\link{setReTrm}},
  \code{\link{setReTrm.default}}; \code{\link{flexvar}},
  \code{\link{setReTrm.flexvar}};
  \code{\link{multiVarExp}},
  \code{\link{setReTrm.multiVarExp}}; \code{\link{obslev}},
  \code{\link{setReTrm.obslev}};
  \code{\link{setReTrm.varIdent}}, \code{\link{varIdent}}
}

