% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setReTrm.nlmeCorStruct}
\alias{obslev}
\alias{setReTrm.nlmeCorStruct}
\title{Random effects term from an \code{nlme}-style \code{corStruct}
object}
\usage{
\method{setReTrm}{nlmeCorStruct}(object, addArgsList, devfunEnv = NULL)
}
\arguments{
\item{object}{a \code{reTrmStruct} object}

\item{addArgsList}{a list of named quantities within which
\code{addArgsExpr} is evaluated}

\item{devfunEnv}{optional environment of the deviance function}
}
\description{
Random effects term from an \code{nlme}-style \code{corStruct}
object
}
\section{Mixed model formula usage}{

\code{nlmeCorStruct(1 | grpFac)}
\describe{
\item{corObj}{corStruct object}
}
}
\seealso{
Other setReTrm: \code{\link{cooccur}},
  \code{\link{setReTrm.cooccur}}; \code{\link{edge}},
  \code{\link{setReTrm.edge}}; \code{\link{expDecay}},
  \code{\link{setReTrm.expDecay}}; \code{\link{factAnal}},
  \code{\link{setReTrm.factAnal}};
  \code{\link{findReTrmClasses}}, \code{\link{getAddArgs}},
  \code{\link{packReTrm}}, \code{\link{setReTrm}},
  \code{\link{setReTrm.default}}; \code{\link{flexvar}},
  \code{\link{setReTrm.flexvar}}; \code{\link{identity}},
  \code{\link{setReTrm.identity}}; \code{\link{obslev}},
  \code{\link{setReTrm.obslev}};
  \code{\link{setReTrm.varExp}}, \code{\link{varExp}};
  \code{\link{setReTrm.varIdent}}, \code{\link{varIdent}}
}

