% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setReTrm}
\alias{setReTrm}
\alias{setReTrm.cooccur}
\alias{setReTrm.default}
\alias{setReTrm.edge}
\alias{setReTrm.identity}
\title{Set model matrix slice and relative covariance factor block for a
random effects term}
\usage{
setReTrm(object, addArgs)

\method{setReTrm}{default}(object, addArgs)

\method{setReTrm}{identity}(object, addArgs)

\method{setReTrm}{edge}(object, addArgs)

\method{setReTrm}{cooccur}(object, addArgs)
}
\arguments{
\item{object}{a \code{reTrmStruct} object}

\item{addArgs}{named list of additional arguments}
}
\description{
Set model matrix slice and relative covariance factor block for a
random effects term
}

