% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setReTrm}
\alias{findReTrmClasses}
\alias{setReTrm}
\alias{setReTrm.cooccur}
\alias{setReTrm.default}
\alias{setReTrm.edge}
\alias{setReTrm.flexvar}
\alias{setReTrm.identity}
\title{Set model matrix slice and relative covariance factor block for a
random effects term}
\usage{
setReTrm(object, addArgs, devfunEnv = NULL)

\method{setReTrm}{default}(object, addArgs, devfunEnv = NULL)

\method{setReTrm}{identity}(object, addArgs, devfunEnv = NULL)

\method{setReTrm}{flexvar}(object, addArgs, devfunEnv = NULL)

\method{setReTrm}{edge}(object, addArgs, devfunEnv = NULL)

\method{setReTrm}{cooccur}(object, addArgs, devfunEnv = NULL)

findReTrmClasses(formula = NULL)
}
\arguments{
\item{object}{a \code{reTrmStruct} object}

\item{addArgs}{named list of additional arguments}

\item{devfunEnv}{optional environment of the deviance function}

\item{formula}{generalized mixed model formula.  if \code{NULL}
(the default) \code{findReTrmClasses} returns classes available
(on the search path).}
}
\value{
\code{object} with the following additional elements:

\item{Zt}{A \code{\link{repSparse}} object describing the slice of
the model matrix associated with the random effects term.}

\item{Lambdat}{A \code{\link{repSparse}} object describing the
block of the relative covariance factor associated with the random
effects term.}
}
\description{
Set model matrix slice and relative covariance factor block for a
random effects term
}
\seealso{
\code{\link{mkReTrmStructs}} for construction of these objects
}

