% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setReTrm}
\alias{findReTrmClasses}
\alias{getAddArgs}
\alias{packReTrm}
\alias{setReTrm}
\alias{setReTrm.cooccur}
\alias{setReTrm.default}
\alias{setReTrm.edge}
\alias{setReTrm.expcutdist}
\alias{setReTrm.factAnal}
\alias{setReTrm.flexvar}
\alias{setReTrm.identity}
\alias{setReTrm.multiVarExp}
\alias{setReTrm.obslev}
\alias{setReTrm.varIdent}
\title{Set model matrix slice and relative covariance factor block for a
random effects term}
\usage{
setReTrm(object, addArgsList, devfunEnv = NULL)

\method{setReTrm}{default}(object, addArgsList, devfunEnv = NULL)

\method{setReTrm}{factAnal}(object, addArgsList, devfunEnv = NULL)

\method{setReTrm}{identity}(object, addArgsList, devfunEnv = NULL)

\method{setReTrm}{flexvar}(object, addArgsList, devfunEnv = NULL)

\method{setReTrm}{expcutdist}(object, addArgsList, devfunEnv = NULL)

\method{setReTrm}{edge}(object, addArgsList, devfunEnv = NULL)

\method{setReTrm}{cooccur}(object, addArgsList, devfunEnv = NULL)

\method{setReTrm}{varIdent}(object, addArgsList, devfunEnv = NULL)

\method{setReTrm}{multiVarExp}(object, addArgsList, devfunEnv = NULL)

\method{setReTrm}{obslev}(object, addArgsList, devfunEnv = NULL)

packReTrm(object, Zt, Lambdat, lowerLoads, upperLoads, lowerCovar, upperCovar,
  ...)

findReTrmClasses(formula = NULL)

getAddArgs(addArgsExpr, addArgsList)
}
\arguments{
\item{object}{a \code{reTrmStruct} object}

\item{addArgsList}{named list of additional arguments}

\item{devfunEnv}{optional environment of the deviance function}

\item{Zt}{transposed model matrix}

\item{Lambdat}{relative covariance factor}

\item{lowerLoads,upperLoads,lowerCovar,upperCovar}{lower and upper
bounds on possible loadings parameters (for \code{Zt}) and
possible covariance parameters (for \code{Lambdat})}

\item{formula}{generalized mixed model formula.  if \code{NULL}
(the default) \code{findReTrmClasses} returns classes available
(on the search path).}

\item{addArgsExpr}{a list of expressions for evaluating within
\code{addArgsList}}

\item{addArgsList}{a list of named quantities within which
\code{addArgsExpr} is evaluated}
}
\value{
\code{object} with the following additional elements:

\item{Zt}{A \code{\link{repSparse}} object describing the slice of
the model matrix associated with the random effects term.}

\item{Lambdat}{A \code{\link{repSparse}} object describing the
block of the relative covariance factor associated with the random
effects term.}

\item{lowerLoads,upperLoads,lowerCovar,upperCovar}{Lower/upper
bounds on loading parameters (possible parameters for \code{Zt})
and/or covariance parameters (possible parameters for \code{Lambdat})}
}
\description{
Set model matrix slice and relative covariance factor block for a
random effects term
}
\seealso{
\code{\link{mkReTrmStructs}} for construction of these objects
}

