% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setReTrm}
\alias{findReTrmClasses}
\alias{getAddArgs}
\alias{packReTrm}
\alias{setReTrm}
\alias{setReTrm.default}
\title{Set model matrix slice and relative covariance factor block for a
random effects term}
\usage{
setReTrm(object, addArgsList, devfunEnv = NULL)

\method{setReTrm}{default}(object, addArgsList, devfunEnv = NULL)

packReTrm(object, Zt, Lambdat, lowerLoads, upperLoads, lowerCovar, upperCovar)

findReTrmClasses(formula = NULL)

getAddArgs(addArgsExpr, addArgsList)
}
\arguments{
\item{object}{a \code{reTrmStruct} object}

\item{addArgsList}{a list of named quantities within which
\code{addArgsExpr} is evaluated}

\item{devfunEnv}{optional environment of the deviance function}

\item{Zt}{transposed model matrix}

\item{Lambdat}{relative covariance factor}

\item{lowerLoads,upperLoads,lowerCovar,upperCovar}{lower and upper
bounds on possible loadings parameters (for \code{Zt}) and
possible covariance parameters (for \code{Lambdat})}

\item{formula}{generalized mixed model formula.  if \code{NULL}
(the default) \code{findReTrmClasses} returns classes available
(on the search path).}

\item{addArgsExpr}{a list of expressions for evaluating within
\code{addArgsList}}
}
\value{
\code{object} with the following additional elements:

\item{Zt}{A \code{\link{repSparse}} object describing the slice of
the model matrix associated with the random effects term.}

\item{Lambdat}{A \code{\link{repSparse}} object describing the
block of the relative covariance factor associated with the random
effects term.}

\item{lowerLoads,upperLoads,lowerCovar,upperCovar}{Lower/upper
bounds on loading parameters (possible parameters for \code{Zt})
and/or covariance parameters (possible parameters for \code{Lambdat})}
}
\description{
Set model matrix slice and relative covariance factor block for a
random effects term
}
\section{Mixed model formula usage}{

The default method is identical to a standard \code{lme4} random effects term.
\code{(linForm | grpFac)}
\describe{
\item{linForm}{linear model formula}
\item{grpFac}{grouping factor}
}
}
\seealso{
\code{\link{mkReTrmStructs}} for construction of these objects

Other setReTrm: \code{\link{cooccur}},
  \code{\link{setReTrm.cooccur}}; \code{\link{edge}},
  \code{\link{setReTrm.edge}}; \code{\link{expDecay}},
  \code{\link{setReTrm.expDecay}}; \code{\link{factAnal}},
  \code{\link{setReTrm.factAnal}}; \code{\link{flexvar}},
  \code{\link{setReTrm.flexvar}}; \code{\link{identity}},
  \code{\link{setReTrm.identity}}; \code{\link{lme4}},
  \code{\link{setReTrm.lme4}}; \code{\link{nlmeCorStruct}},
  \code{\link{setReTrm.nlmeCorStruct}};
  \code{\link{obslev}}, \code{\link{setReTrm.obslev}};
  \code{\link{setReTrm.varExp}}, \code{\link{varExp}};
  \code{\link{setReTrm.varIdent}}, \code{\link{varIdent}}
}

