% Generated by roxygen2 (4.0.1): do not edit by hand
\name{showSkeleton}
\alias{showSkeleton}
\title{Show skeleton for some type of function}
\usage{
showSkeleton(whichSkel, package = "lme4ord")
}
\arguments{
\item{whichSkel}{length-one character giving the name of the
skeleton}

\item{package}{the default is probably what you want}
}
\description{
Show skeleton for some type of function
}
\examples{
showSkeleton("setReTrm")
}

