% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simAddArgs}
\alias{simAddArgs}
\alias{simAddArgs.default}
\alias{simAddArgs.phyloEdge}
\alias{simAddArgsList}
\title{Simulate additional arguments}
\usage{
simAddArgs(object, ...)

simAddArgsList(object, ...)

\method{simAddArgs}{default}(object, ...)

\method{simAddArgs}{phyloEdge}(object, rtreeArgs = list(),
  compute.brlenArgs = list(), ...)
}
\arguments{
\item{object}{a \code{reTrmStruct} object}

\item{...}{dots}

\item{rtreeArgs}{arguments for \code{\link{rtree}}}

\item{compute.brlenArgs}{arguments for \code{\link{compute.brlen}}}
}
\description{
Simulate additional arguments
}

