% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simplifyFacList}
\alias{simplifyFacList}
\title{Simplify factor list over random effects terms}
\usage{
simplifyFacList(facList)
}
\arguments{
\item{facList}{list of grouping factors over random effects terms}
}
\value{
collapse repeated factors and add an \code{assign}
attribute for indicating which factors relate to which terms
}
\description{
Simplify factor list over random effects terms
}

