% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simplifyRepSparse}
\alias{simplifyRepSparse}
\title{Simplify repeated sparse matrices}
\usage{
simplifyRepSparse(object, ...)
}
\arguments{
\item{object}{repeated sparse matrix}

\item{...}{not yet used}
}
\description{
Simplify repeated sparse matrices by (1) forcing non structural
zeros to be structural zeros, (2) eliminate duplicates in the
repeated values, and (3) resetting the transformation function to
be the identity.
}
\seealso{
Other modifications: \code{\link{resetTransConst}};
  \code{\link{scalarMult}}
}

