% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simplifyRepSparse}
\alias{simplifyRepSparse}
\title{Simplify structured sparse matrices}
\usage{
simplifyRepSparse(object, force = TRUE, eliminate = TRUE, reset = TRUE,
  ...)
}
\arguments{
\item{object}{structured sparse matrix}

\item{force}{force non structural zeros to be structural zeros?}

\item{eliminate}{eliminate duplicates in the repeated values?}

\item{reset}{reset the transformation function to be the identity?}

\item{...}{not yet used}
}
\description{
Simplify structured sparse matrices by (1) forcing non structural
zeros to be structural zeros, (2) eliminate duplicates in the
repeated values, and (3) resetting the transformation function to
be the identity.
}
\seealso{
Other modifications: \code{\link{resetTransConst}};
  \code{\link{scalarMult}}
}

