% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simplifyRepSparse}
\alias{simplifyRepSparse}
\title{Simplify a repeated sparse matrix}
\usage{
simplifyRepSparse(object)
}
\arguments{
\item{object}{a \code{repSparse} object}
}
\description{
Remove unused values and renumber value indices of a
\code{repSparse} object that has values that are not used in any
matrix element.  (FIXME: not currently used anywhere)
}

