% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simStrucParsedForm}
\alias{simStrucParsedForm}
\title{Simulate from a parsed formula}
\usage{
simStrucParsedForm(parsedForm, family = binomial, weights, nsim)
}
\arguments{
\item{nsim}{number of simulations (FIXME: inconsistent between
binomial and poisson)}

\item{parsedForm}{result of \code{strucParseFormula}}

\item{family}{family object}

\item{weights}{optional weights}
}
\description{
Simulate from a parsed formula
}

