% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sortedBind}
\alias{sortedBind}
\title{Bind repeated sparse matrices and sort their indices to be
compatible with column-compressed sparse matrices}
\usage{
sortedBind(mats, type = c("row", "col", "diag"))
}
\arguments{
\item{mats}{list of \code{strucMatrix} matrix objects}

\item{type}{type of bind}
}
\description{
Bind repeated sparse matrices and sort their indices to be
compatible with column-compressed sparse matrices
}

