% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repSparseCompSymm}
\alias{rRepSparse}
\alias{repSparseCompSymm}
\alias{repSparseDiag}
\alias{repSparseIdent}
\alias{repSparseOneOffDiag}
\title{Repeated sparse matrix with compound symmetry}
\usage{
repSparseCompSymm(diagVal, offDiagVal, matSize)

repSparseOneOffDiag(diagVal, offDiagVal, offDiagInds, matSize)

repSparseDiag(vals, valInds)

repSparseIdent(matSize)

rRepSparse(nrows, ncols, nvals, nnonzeros, rfunc = rnorm, ...)
}
\arguments{
\item{diagVal}{value for the diagonal}

\item{offDiagVal}{value for the off-diagonal}

\item{matSize}{size of the resulting matrix}

\item{diagVal}{value for the diagonal}

\item{offDiagVal}{value for the off-diagonal}

\item{offDiagInds}{indices for the two correlated objects}

\item{matSize}{size of the resulting matrix}

\item{vals}{vector of values}

\item{valInds}{vector of value indices}

\item{matSize}{size of matrix}

\item{nrows,ncols}{numbers of rows and columns}

\item{nvals}{number of values}

\item{nnonzeros}{number of nonzero elements}

\item{rfunc}{random number function}

\item{...}{dots}
}
\description{
Repeated sparse matrix with compound symmetry

Repeated sparse covariance matrix with equal variances and one
off-diagonal covariance

Diagonal repeated sparse matrix

Identity repeated sparse matrix

Random repeated sparse matrix
}

