% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repSparseBlank}
\alias{rRepSparse}
\alias{repSparseBlank}
\alias{repSparseCompSymm}
\alias{repSparseConstVarChol}
\alias{repSparseCorMatChol}
\alias{repSparseDiag}
\alias{repSparseIdent}
\alias{repSparseOneOffDiag}
\alias{repSparseSymm}
\alias{repSparseTri}
\alias{repSparseVarWithCovariate}
\title{Special sparse repeated matrices}
\usage{
repSparseBlank(nrow, ncol)

repSparseIdent(matSize)

repSparseDiag(vals, valInds)

repSparseTri(diagVals, offDiagVals, low = TRUE)

repSparseSymm(diagVals, offDiagVals)

repSparseCompSymm(diagVal, offDiagVal, matSize)

repSparseOneOffDiag(diagVal, offDiagVal, offDiagInds, matSize)

repSparseConstVarChol(sdVal, offDiagVals)

repSparseCorMatChol(offDiagPars)

repSparseVarWithCovariate(varPars, covariate, grpFac,
  mkTransFunc = mkVarExpTrans)

rRepSparse(nrows, ncols, nvals, nnonzeros, rfunc = rnorm, ...)
}
\arguments{
\item{nrow,ncol}{number of rows and columns}

\item{vals}{vector of values}

\item{valInds}{vector of value indices}

\item{diagVals}{values for the diagonal}

\item{offDiagVals}{values for the off-diagonal}

\item{low}{lower triangular?}

\item{diagVal}{value for the diagonal}

\item{offDiagVal}{value for the off-diagonal}

\item{matSize}{size of the resulting matrix}

\item{offDiagInds}{indices for the two correlated objects}

\item{sdVal}{standard deviation of crossproduct of the result}

\item{offDiagVals}{values for the off-diagonal of the Cholesky
factor}

\item{offDiagPars}{parameters determining the off-diagonal of the
Cholesky factor}

\item{varPars}{vector of variance parameters (one per level of
\code{grpFac})}

\item{covariate}{covariate}

\item{grpFac}{a grouping factor (or anything coercible to
\code{numeric} really) with the number of levels equal to the
length of \code{varPars}}

\item{mkTransFunc}{a function taking arguments \code{varPars},
\code{covariate}, and \code{grpFac} for constructing a
\code{trans} function (see \code{\link{mkVarExpTrans}} for an
example).}

\item{nrows,ncols}{numbers of rows and columns}

\item{nvals}{number of values}

\item{nnonzeros}{number of nonzero elements}

\item{rfunc}{random number function}

\item{...}{dots}
}
\description{
Special sparse repeated matrices
}
\seealso{
Other repeated.sparse.matrix.topics: \code{\link{.bind}},
  \code{\link{bind}}, \code{\link{rep.repSparse}};
  \code{\link{as.data.frame.repSparse}},
  \code{\link{as.data.table.repSparse}},
  \code{\link{as.matrix.repSparse}},
  \code{\link{as.repSparse}},
  \code{\link{as.repSparse.dist}},
  \code{\link{as.repSparse.dsparseMatrix}},
  \code{\link{as.repSparse.factor}},
  \code{\link{as.repSparse.matrix}},
  \code{\link{as.repSparse.repSparse}},
  \code{\link{repSparse2CsparseMatrix}},
  \code{\link{repSparse2TsparseMatrix}},
  \code{\link{repSparse2dgCMatrix}},
  \code{\link{repSparse2dgTMatrix}},
  \code{\link{repSparse2sparseMatrix}};
  \code{\link{chol.repSparse}},
  \code{\link{chol.repSparseCompSymm}},
  \code{\link{chol.repSparseOneOffDiag}};
  \code{\link{dim.repSparse}},
  \code{\link{image.repSparse}},
  \code{\link{plot.repSparse}},
  \code{\link{print.repSparse}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}}, \code{\link{t.repSparse}},
  \code{\link{update.repSparse}}; \code{\link{getInit}},
  \code{\link{getInit.default}},
  \code{\link{getInit.function}},
  \code{\link{getInit.repSparse}}, \code{\link{setInit}},
  \code{\link{setInit.default}},
  \code{\link{setInit.function}},
  \code{\link{setInit.repSparse}}; \code{\link{ind2point}},
  \code{\link{point2ind}}; \code{\link{kr}},
  \code{\link{kron}}; \code{\link{mkCholCompSymmTrans}},
  \code{\link{mkCholOneOffDiagTrans}},
  \code{\link{mkConstVarCholTrans}},
  \code{\link{mkCorMatCholTrans}},
  \code{\link{mkGenCholInPlaceTrans}},
  \code{\link{mkIdentityTrans}}, \code{\link{mkListTrans}},
  \code{\link{mkOuterTrans}}, \code{\link{mkVarExpTrans}},
  \code{\link{mkVarFixedTrans}},
  \code{\link{mkVarIdentTrans}},
  \code{\link{mkVarPowerTrans}}; \code{\link{repSparse}}
}

