% Generated by roxygen2 (4.0.1): do not edit by hand
\name{repSparseCompSymm}
\alias{rRepSparse}
\alias{repSparseCompSymm}
\alias{repSparseDiag}
\alias{repSparseIdent}
\alias{repSparseOneOffDiag}
\alias{repSparseTri}
\title{Repeated sparse matrix with compound symmetry}
\usage{
repSparseCompSymm(diagVal, offDiagVal, matSize)

repSparseOneOffDiag(diagVal, offDiagVal, offDiagInds, matSize)

repSparseDiag(vals, valInds)

repSparseIdent(matSize)

repSparseTri(diagVals, offDiagVals, low = TRUE)

rRepSparse(nrows, ncols, nvals, nnonzeros, rfunc = rnorm, ...)
}
\arguments{
\item{diagVal}{value for the diagonal}

\item{offDiagVal}{value for the off-diagonal}

\item{matSize}{size of the resulting matrix}

\item{offDiagInds}{indices for the two correlated objects}

\item{vals}{vector of values}

\item{valInds}{vector of value indices}

\item{diagVals}{values for the diagonal}

\item{offDiagVals}{values for the off-diagonal}

\item{low}{lower triangular?}

\item{nrows,ncols}{numbers of rows and columns}

\item{nvals}{number of values}

\item{nnonzeros}{number of nonzero elements}

\item{rfunc}{random number function}

\item{...}{dots}
}
\description{
Repeated sparse matrix with compound symmetry

Diagonal repeated sparse matrix

Identity repeated sparse matrix

Triangular repeated sparse matrix

Random repeated sparse matrix
}
\seealso{
Other repSparseTopics: \code{\link{.bind}},
  \code{\link{bind}}, \code{\link{rep.repSparse}};
  \code{\link{as}}, \code{\link{as}}, \code{\link{as}},
  \code{\link{as}}, \code{\link{as}},
  \code{\link{as.data.frame.repSparse}},
  \code{\link{as.data.table.repSparse}},
  \code{\link{as.matrix.repSparse}},
  \code{\link{as.repSparse}},
  \code{\link{as.repSparse.dsparseMatrix}},
  \code{\link{as.repSparse.factor}},
  \code{\link{as.repSparse.matrix}},
  \code{\link{as.repSparse.repSparse}};
  \code{\link{chol.repSparse}},
  \code{\link{chol.repSparseCompSymm}},
  \code{\link{chol.repSparseOneOffDiag}};
  \code{\link{dim.repSparse}},
  \code{\link{image.repSparse}},
  \code{\link{print.repSparse}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{repSparse-class}},
  \code{\link{sort.repSparse}}, \code{\link{standardSort}},
  \code{\link{t.repSparse}},
  \code{\link{update.repSparse}}; \code{\link{getInit}},
  \code{\link{getInit.default}},
  \code{\link{getInit.function}},
  \code{\link{getInit.repSparse}}, \code{\link{setInit}},
  \code{\link{setInit.default}},
  \code{\link{setInit.function}},
  \code{\link{setInit.repSparse}}; \code{\link{ind2point}},
  \code{\link{point2ind}}; \code{\link{kr}},
  \code{\link{kr0}}, \code{\link{kron}},
  \code{\link{mmult}}, \code{\link{rowWiseCombination}};
  \code{\link{mkCholOneOffDiagTrans}},
  \code{\link{mkGenCholInPlaceTrans}},
  \code{\link{mkIdentityTrans}}, \code{\link{mkListTrans}},
  \code{\link{mkOuterTrans}}; \code{\link{repSparse}}
}

