% Generated by roxygen2 (4.0.1): do not edit by hand
\name{splitForm}
\alias{noSpecials}
\alias{reForm}
\alias{splitForm}
\title{Split a formula}
\usage{
splitForm(formula)

reForm(splitFormula)

noSpecials(term)
}
\arguments{
\item{formula}{Generalized mixed model formula}

\item{splitFormula}{results of \code{splitForm}}
}
\description{
Split a formula
}

