% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strucGlmer-class}
\alias{deviance.strucGlmer}
\alias{df.residual.strucGlmer}
\alias{formula.strucGlmer}
\alias{isREML.strucGlmer}
\alias{logLik.strucGlmer}
\alias{nobs}
\alias{print.strucGlmer}
\alias{print.summary.strucGlmer}
\alias{ranef.strucGlmer}
\alias{strucGlmer-class}
\alias{summary.strucGlmer}
\title{Structured GLMM class}
\usage{
print.strucGlmer(x, digits = max(3, getOption("digits") - 3), ...)

summary.strucGlmer(object, use.hessian = TRUE, ...)

print.summary.strucGlmer(x, digits = max(3, getOption("digits") - 3),
  correlation = NULL, signif.stars = getOption("show.signif.stars"),
  show.resids = TRUE, ...)

\method{isREML}{strucGlmer}(x, ...)

\method{df.residual}{strucGlmer}(object, ...)

nobs(object, ...)

\method{deviance}{strucGlmer}(object, ...)

\method{logLik}{strucGlmer}(object, ...)

\method{formula}{strucGlmer}(x, ...)

\method{ranef}{strucGlmer}(object, type = c("u", "Lu", "ZLu"), ...)
}
\arguments{
\item{x,object}{\code{strucGlmer} objects}

\item{...}{additional arguments to methods}
}
\description{
An \code{S3} class for generalized linear mixed models with
structured (co)variance terms, which are lists with the following
elements:
\describe{
  \item{opt}{Result of the optimizer (currently \code{\link{minqa}}).}
  \item{parsedForm}{Results of \code{\link{strucParseFormula}}.}
  \item{dfun}{A function for computing the model deviance.
              The environment of \code{dfun} contains objects for
              representing the model.}
  \item{mc}{Matched call}}
}

