% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strucGlmer-class}
\alias{VarCorr.strucGlmer}
\alias{deviance.strucGlmer}
\alias{df.residual.strucGlmer}
\alias{fitted.strucGlmer}
\alias{formula.strucGlmer}
\alias{isREML.strucGlmer}
\alias{logLik.strucGlmer}
\alias{mkStrucGlmer}
\alias{nobs}
\alias{print.strucGlmer}
\alias{print.summary.strucGlmer}
\alias{residuals.strucGlmer}
\alias{strucGlmer-class}
\alias{summary.strucGlmer}
\alias{vcov.strucGlmer}
\title{Structured GLMM class}
\usage{
mkStrucGlmer(opt, parsedForm, dfun, mc)

print.strucGlmer(x, digits = max(3, getOption("digits") - 3), ...)

summary.strucGlmer(object, use.hessian = TRUE, ...)

print.summary.strucGlmer(x, digits = max(3, getOption("digits") - 3),
  correlation = NULL, signif.stars = getOption("show.signif.stars"),
  show.resids = TRUE, ...)

\method{residuals}{strucGlmer}(object, ...)

\method{fitted}{strucGlmer}(object, ranefTrms, fixef = TRUE, ...)

\method{vcov}{strucGlmer}(object, correlation = TRUE, use.hessian = TRUE,
  justFixef = TRUE, ...)

\method{VarCorr}{strucGlmer}(x, sigma = 1, rdig = 3)

\method{isREML}{strucGlmer}(x, ...)

\method{df.residual}{strucGlmer}(object, ...)

nobs(object, ...)

\method{deviance}{strucGlmer}(object, ...)

\method{logLik}{strucGlmer}(object, ...)

\method{formula}{strucGlmer}(x, ...)
}
\arguments{
\item{opt,parsedForm,dfun,mc}{See desciption of \code{strucGlmer}
objects above}

\item{x,object}{\code{strucGlmer} objects}

\item{digits}{number of significant digits}

\item{...}{additional arguments to methods}

\item{use.hessian}{use numerical hessian in covariance
calculations if available}

\item{correlation}{display correlation among fixed effects?}

\item{signif.stars}{display significance stars?}

\item{show.resids}{show residuals?}

\item{ranefTrms}{character vector naming random effects terms to
be used}

\item{fixef}{should fixed effects be used?}

\item{justFixef}{should matrices be over fixed effects only?}

\item{sigma,rdig}{for consistency}
}
\description{
An \code{S3} class for generalized linear mixed models with
structured (co)variance terms, which are lists with the following
elements:
\describe{
  \item{opt}{Result of the optimizer (currently \code{\link[minqa]{bobyqa}}).}
  \item{parsedForm}{Results of \code{\link{strucParseFormula}}.}
  \item{dfun}{A function for computing the model deviance.
              The environment of \code{dfun} contains objects for
              representing the model.}
  \item{mc}{Matched call}}
}

