% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strucGlmer}
\alias{strucGlmer}
\title{Generalized linear mixed model with structured (co)variance terms}
\usage{
strucGlmer(formula, data, family, addArgs = list(), optVerb = 0L,
  weights = NULL, offset = NULL, etastart = NULL, devfunOnly = FALSE,
  ...)
}
\arguments{
\item{formula}{extended mixed model formula}

\item{data,family,weights,offset,etastart,devfunOnly}{see \code{\link{glmer}}}

\item{addArgs}{list of additional arguments to pass to}

\item{optVerb}{verbose}

\item{...}{further arguments to \code{\link{mkGeneralGlmerDevfun}}}
}
\description{
Fits a generalized linear mixed model with structured (co)variance
terms by the following fully modularized steps:
\describe{
\item{\code{\link{strucParseFormula}}}{parses the mixed model
formula with structured terms}
\item{\code{\link{mkGeneralGlmerDevfun}}}{constructs a generalized
linear mixed model deviance function}
\item{\code{bobyqa}}{optimizer the deviance function}
\item{\code{\link{mkStrucGlmer}}}{constructs an object of
\code{\link{strucGlmer-class}}}}
}

