% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strucGlmer}
\alias{covar.strucGlmer}
\alias{fixef.strucGlmer}
\alias{getStrucGlmerPar}
\alias{loads.strucGlmer}
\alias{print.strucGlmer}
\alias{strucGlmer}
\title{Structured glmer}
\usage{
strucGlmer(formula, data, addArgs = list(), optVerb = 0L, weights, offset,
  etastart, ...)

print.strucGlmer(x, ...)

getStrucGlmerPar(object, type, ...)

\method{fixef}{strucGlmer}(object, ...)

\method{covar}{strucGlmer}(object, ...)

loads.strucGlmer(object, ...)
}
\arguments{
\item{formula}{extended mixed model formula}

\item{data}{data frame}

\item{addArgs}{list of additional arguments to pass to}

\item{optVerb}{verbose
\code{\link{setReTrm}} methods}

\item{...}{further arguments to \code{\link{mkGeneralGlmerDevfun}}}

\item{type}{character string giving the type of parameter
(e.g. \code{"fixef", "covar"})}
}
\description{
Structured glmer
}

