% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strucGlmer}
\alias{VarCorr.strucGlmer}
\alias{covar.strucGlmer}
\alias{covarPerTerm}
\alias{fixef.strucGlmer}
\alias{getStrucGlmerPar}
\alias{loads.strucGlmer}
\alias{loadsPerTerm}
\alias{pars.strucGlmer}
\alias{residuals.strucGlmer}
\alias{strucGlmer}
\alias{vcov.strucGlmer}
\title{Generalized linear mixed model with structured (co)variance terms}
\usage{
strucGlmer(formula, data, family, addArgs = list(), optVerb = 0L,
  weights = NULL, offset = NULL, etastart = NULL, ...)

\method{residuals}{strucGlmer}(object, ...)

\method{vcov}{strucGlmer}(object, correlation = TRUE, use.hessian = TRUE,
  justFixef = TRUE, ...)

\method{VarCorr}{strucGlmer}(x, sigma = 1, rdig = 3)

getStrucGlmerPar(object, type, ...)

\method{fixef}{strucGlmer}(object, ...)

\method{covar}{strucGlmer}(object, ...)

loads.strucGlmer(object, ...)

\method{pars}{strucGlmer}(object, ...)

covarPerTerm(object)

loadsPerTerm(object)
}
\arguments{
\item{formula}{extended mixed model formula}

\item{data}{data frame}

\item{addArgs}{list of additional arguments to pass to}

\item{optVerb}{verbose
\code{\link{setReTrm}} methods}

\item{...}{further arguments to \code{\link{mkGeneralGlmerDevfun}}}

\item{type}{character string giving the type of parameter
(e.g. \code{"fixef", "covar"})}
}
\description{
Fits a generalized linear mixed model with structured (co)variance
terms by the following fully modularized steps:
\describe{
\item{\code{\link{strucParseFormula}}}{parses the mixed model
formula with structured terms}
\item{\code{\link{mkGeneralGlmerDevfun}}}{constructs a generalized
linear mixed model deviance function}
\item{\code{\link{bobyqa}}}{optimizer the deviance function}
\item{\code{\link{mkStrucGlmer}}}{constructs an object of
\code{\link{strucGlmer-class}}}}
}

