% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strucGlmer}
\alias{strucGlmer}
\alias{strucGlmer.default}
\alias{strucGlmer.formula}
\alias{strucGlmer.function}
\alias{strucGlmer.strucParseFormula}
\alias{update.strucGlmer}
\title{Generalized linear mixed model with structured (co)variance terms}
\usage{
strucGlmer(formula, ...)

\method{strucGlmer}{default}(formula, ...)

\method{strucGlmer}{formula}(formula, data, family, addArgs = list(),
  modularVerb = FALSE, optVerb = 0L, optMaxit = 10000, weights = NULL,
  offset = NULL, etastart = NULL, devfunOnly = FALSE,
  numModularSteps = 4L, parList = NULL, ...)

\method{strucGlmer}{strucParseFormula}(formula, family, optVerb = 0L,
  optMaxit = 10000, weights = NULL, offset = NULL, etastart = NULL, ...)

\method{strucGlmer}{function}(formula, optVerb = 0L, optMaxit = 10000,
  parsedForm, initPars, lower, upper, ...)

\method{update}{strucGlmer}(object, formula, parList, data = NULL,
  addArgs = list(), optVerb = 0L, optMaxit = 10000, weights = NULL,
  offset = NULL, etastart = NULL, ...)
}
\arguments{
\item{formula}{extended mixed model formula or
\code{\link{strucParseFormula}} object}

\item{data,family,weights,offset,etastart,devfunOnly}{see \code{\link{glmer}}}

\item{addArgs}{list of additional arguments to pass to}

\item{modularVerb}{should the modular steps be announced during the fitting process?}

\item{optVerb}{verbose}

\item{optMaxit}{maximum number of iterations for the nonlinear
optimizers}

\item{numModularSteps}{number of modular steps to perform
(defaults to \code{4}, which is all of them). \code{1} returns the
\code{\link{strucParseFormula}} object, \code{2} the deviance
function, \code{3} the optimizer output, and \code{4} a
\code{strucGlmer} object.}

\item{parList}{potential named list of initial parameters}

\item{parsedForm}{\code{\link{strucParseFormula}} object}

\item{initPars}{initial parameter values (required for
\code{strucGlmer.function} method if missing \code{parsedForm})}

\item{lower,upper}{lower and upper bounds of parameter values
(only useful for \code{strucGlmer.function} method)}

\item{...}{further arguments to \code{\link{mkGeneralGlmerDevfun}}}

\item{object}{\code{strucGlmer} object}
}
\description{
Fits a generalized linear mixed model with structured (co)variance
terms by the following fully modularized steps:
\describe{
\item{\code{\link{strucParseFormula}}}{parses the mixed model
formula with structured terms}
\item{\code{\link{mkGeneralGlmerDevfun}}}{constructs a generalized
linear mixed model deviance function}
\item{\code{bobyqa}}{optimizer the deviance function}
\item{\code{\link{mkStrucGlmer}}}{constructs an object of
\code{\link{strucGlmer-class}}}}
}
\examples{
cbpp$incidenceBySize <- with(cbpp, incidence/size)
gm <- strucGlmer(incidenceBySize ~ factAnal(0 + herd | period, nAxes = 1),
                 data = cbpp, family = binomial, weights = cbpp$size,
                 penLoads = mkPenLpNorm())
}

