% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strucMatrixBlank}
\alias{strucMatrixBlank}
\title{Blank repeated sparse matrix}
\usage{
strucMatrixBlank(nrow, ncol)
}
\arguments{
\item{nrow,ncol}{number of rows and columns}
}
\description{
Blank repeated sparse matrix
}
\examples{
(xBlank <- strucMatrixBlank(5, 5))
}
\seealso{
Other strucMatrixSpecial: \code{\link{strucMatrixCol}};
  \code{\link{strucMatrixCompSymm}};
  \code{\link{strucMatrixConstVarChol}};
  \code{\link{strucMatrixCorFactor}};
  \code{\link{strucMatrixCorMatChol}};
  \code{\link{strucMatrixDiag}};
  \code{\link{strucMatrixExpChol}};
  \code{\link{strucMatrixFull}};
  \code{\link{strucMatrixGenFullTri}};
  \code{\link{strucMatrixIdent}};
  \code{\link{strucMatrixInd}};
  \code{\link{strucMatrixOneOffDiag}};
  \code{\link{strucMatrixOnes}};
  \code{\link{strucMatrixSymm}};
  \code{\link{strucMatrixTri}};
  \code{\link{strucMatrixVarWithCovariate}}
}

