% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strucMatrixIdent}
\alias{strucMatrixIdent}
\title{Identity structured sparse matrix}
\usage{
strucMatrixIdent(matSize, val)
}
\arguments{
\item{matSize}{matrix size}

\item{val}{value to be repeated on the diagonal (defaults to
\code{1})}
}
\description{
Identity structured sparse matrix
}
\examples{
(xIdent <- strucMatrixIdent(5))
}
\seealso{
Other strucMatrixSpecial: \code{\link{strucMatrixBlank}};
  \code{\link{strucMatrixCol}};
  \code{\link{strucMatrixCompSymm}};
  \code{\link{strucMatrixConstVarChol}};
  \code{\link{strucMatrixCorFactor}};
  \code{\link{strucMatrixCorMatChol}};
  \code{\link{strucMatrixDiag}};
  \code{\link{strucMatrixExpChol}};
  \code{\link{strucMatrixFull}};
  \code{\link{strucMatrixGenFullTri}};
  \code{\link{strucMatrixInd}};
  \code{\link{strucMatrixOneOffDiag}};
  \code{\link{strucMatrixOnes}};
  \code{\link{strucMatrixSymm}};
  \code{\link{strucMatrixTri}};
  \code{\link{strucMatrixVarWithCovariate}}
}

