% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strucMatrixSymm}
\alias{strucMatrixSymm}
\title{Symmetric structured sparse matrix}
\usage{
strucMatrixSymm(diagVals, offDiagVals)
}
\arguments{
\item{diagVals}{values for the diagonal}

\item{offDiagVals}{values for the off-diagonal}
}
\description{
Symmetric structured sparse matrix
}
\examples{
set.seed(1)
(xSymm <- strucMatrixSymm(rnorm(5), rnorm(choose(5, 2))))
}
\seealso{
Other strucMatrixSpecial: \code{\link{strucMatrixBlank}};
  \code{\link{strucMatrixCol}};
  \code{\link{strucMatrixCompSymm}};
  \code{\link{strucMatrixConstVarChol}};
  \code{\link{strucMatrixCorFactor}};
  \code{\link{strucMatrixCorMatChol}};
  \code{\link{strucMatrixDiag}};
  \code{\link{strucMatrixExpChol}};
  \code{\link{strucMatrixFull}};
  \code{\link{strucMatrixGenFullTri}};
  \code{\link{strucMatrixIdent}};
  \code{\link{strucMatrixInd}};
  \code{\link{strucMatrixOneOffDiag}};
  \code{\link{strucMatrixOnes}};
  \code{\link{strucMatrixTri}};
  \code{\link{strucMatrixVarWithCovariate}}
}

