% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strucParseFormula}
\alias{simStrucParsedForm}
\alias{strucParseFormula}
\title{Parse a mixed model formula}
\usage{
strucParseFormula(formula, data, addArgs = list(), reTrmsList = NULL, ...)

simStrucParsedForm(parsedForm, family = binomial, weights, nsim)
}
\arguments{
\item{formula}{mixed model formula}

\item{data}{an object coercible to data frame}

\item{addArgs}{list of additional arguments to
\code{\link{setReTrm}} methods}

\item{reTrmsList}{if \code{NULL} \code{\link{mkReTrmStructs}} is used}

\item{...}{additional parameters to \code{\link{as.data.frame}}}

\item{parsedForm}{result of \code{strucParseFormula}}

\item{family}{family object}

\item{weights}{optional weights}
}
\value{
A list with components:
\item{response}{The response vector}
\item{fixed}{The fixed effects model matrix}
\item{random}{List of random effects terms returned by
\code{\link{mkReTrmStructs}} and updated with
\code{\link{setReTrm}}.}
\item{Zt}{The sparse transposed random effects model matrix, of
class \code{dgCMatrix}.}
\item{Lambdat}{The sparse transposed relative covariance factor,
of class \code{dgCMatrix}}
\item{initPars}{The initial parameter vector}
\item{parInds}{List of indices to the different types of
parameters.}
\item{lower, upper}{Vectors of lower and upper bounds on
\code{initPars}.}
\item{devfunEnv}{Environment of the deviance function.}
\item{formula}{Model formula.}
}
\description{
Parse a mixed model formula
}

