% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strucParseFormula}
\alias{simStrucParsedForm}
\alias{strucParseFormula}
\title{Parse a mixed model formula}
\usage{
strucParseFormula(formula, data, addArgs = list(), reTrmsList = NULL, ...)

simStrucParsedForm(parsedForm, family = binomial, weights, ...)
}
\arguments{
\item{formula}{mixed model formula}

\item{data}{an object coercible to data frame}

\item{addArgs}{list of additional arguments to
\code{\link{setReTrm}} methods}

\item{reTrmsList}{if \code{NULL} \code{\link{mkReStructs}} is used}

\item{...}{additional parameters to \code{\link{as.data.frame}}}

\item{parsedForm}{result of \code{strucParseFormula}}

\item{family}{family object}

\item{weights}{optional weights}
}
\description{
Parse a mixed model formula
}

