% Generated by roxygen2 (4.0.1): do not edit by hand
\name{subset.repSparse}
\alias{repSparseRowSubset}
\alias{subset.repSparse}
\title{Subsetting repeated sparse matices}
\usage{
\method{subset}{repSparse}(x, rowInds = NULL, colInds = NULL, ...)

repSparseRowSubset(x, rowInds)
}
\arguments{
\item{x}{a \code{\link{repSparse-class}} object}

\item{rowInds,colInds}{1-based integer indices for rows and
columns}

\item{...}{unused}
}
\description{
Subsetting repeated sparse matices
}
\examples{
set.seed(1)
n <- 8; m <- 5
X <- repSparse(1:m,
               rep(1:2, c(2, m - 2)),
               1:m, rep(1, m))
fac <- factor(letters[rep(sample(m), n)])
levels(fac) <- levels(fac)[sample(m)]
Z <- subset(X, as.numeric(fac))
image(update(Z, rnorm(m)))
image(update(Z, rnorm(m)))
}

