% Generated by roxygen2 (4.0.1): do not edit by hand
\name{subset.strucMatrix}
\alias{strucMatrixRowSubset}
\alias{subset.strucMatrix}
\title{Subsetting repeated sparse matices}
\usage{
\method{subset}{strucMatrix}(x, rowInds = NULL, colInds = NULL, ...)

strucMatrixRowSubset(x, rowInds)
}
\arguments{
\item{x}{a \code{\link{strucMatrix-class}} object}

\item{rowInds,colInds}{1-based integer indices for rows and
columns}

\item{...}{unused}
}
\description{
Subsetting repeated sparse matices
}
\examples{
set.seed(1)
X <- strucMatrix(c(1, 2, 1, 2), c(1, 1, 2, 2), 1:4, rnorm(4))
(subsetExample <- subset(X, c(1, 2, 2, 2, 1, 1)))
}
\seealso{
Other matrixCombining: \code{\link{.bind}},
  \code{\link{bind}}, \code{\link{rep.strucMatrix}};
  \code{\link{kr}}, \code{\link{kron}},
  \code{\link{kronecker,strucMatrix,strucMatrix-method}}
}

