% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tipNodeRelationship}
\alias{tipNodeRelationship}
\title{Tip-node bigraph adjacency matrix}
\usage{
tipNodeRelationship(phy)
}
\arguments{
\item{phy}{phylo object}
}
\value{
adjacency matrix
}
\description{
Tip-node bigraph adjacency matrix
}

