% Generated by roxygen2 (4.0.1): do not edit by hand
\name{update.reTrmStruct}
\alias{update.flexvar}
\alias{update.reTrmStruct}
\title{Update a random effects term structure with new parameters}
\usage{
\method{update}{reTrmStruct}(object, newCovar, newLoads, ...)

\method{update}{flexvar}(object, newCovar, newLoads, ...)
}
\arguments{
\item{object}{a \code{\link{reTrmStruct}} object}

\item{newCovar}{new covariance parameters}

\item{newLoads}{new loadings parameters}

\item{...}{potential additional arguments (ignored by the default
method)}
}
\description{
Update a random effects term structure with new parameters
}

